feat: URL taxonomy refactor — semantic /ship/{initiative} routes
Replace the flat /build?initiative= query-param pattern with semantic
path-based URLs throughout the stack.

Browser / HTMX:
  /build?initiative={x}              → /ship/{initiative}
  /build/board?initiative={x}        → /ship/{initiative}/board
  /build/agent/{id}/stream           → /ship/runs/{id}/stream
  Old /ship PR board                 → REMOVED

API — user-facing:
  POST /api/build/agent/{id}/message → POST /api/runs/{id}/message
  POST /api/build/agent/{id}/stop    → POST /api/runs/{id}/stop
  POST /api/build/advance-phase      → POST /api/ship/{initiative}/advance
  POST /api/build/dispatch           → POST /api/dispatch/issue
  POST /api/build/dispatch-label     → POST /api/dispatch/label
  GET  /api/build/label-context      → GET  /api/dispatch/context
  GET  /api/build/dispatcher-prompt  → GET  /api/dispatch/prompt

API — agent-facing:
  GET  /api/build/pending-launches   → GET  /api/runs/pending
  POST /api/build/acknowledge/{id}   → POST /api/runs/{id}/acknowledge
  POST /api/build/spawn-child        → POST /api/runs/{id}/children
  POST /api/build/report/step        → POST /api/runs/{id}/step
  POST /api/build/report/blocker     → POST /api/runs/{id}/blocker
  POST /api/build/report/decision    → POST /api/runs/{id}/decision
  POST /api/build/report/done        → POST /api/runs/{id}/done

run_id moves from JSON body to URL path on all report endpoints.

Split routes/api/build.py → dispatch.py + runs.py + ship_api.py.
Delete routes/ui/ship.py (old PR board). Nav trimmed to Plan + Ship.
All tests, mypy, and typing audit pass (0 Any, 0 errors, 52/52 tests).
