+++ agentception/routes/api/__init__.py
@@ -12,7 +12,9 @@ import path continues to work unchanged.
 
 from fastapi import APIRouter
 
-from .build import router as _build
+from .dispatch import router as _dispatch
+from .runs import router as _runs
+from .ship_api import router as _ship_api
 from .config import router as _config
 from .control import router as _control
 from .health import router as _health
@@ -25,7 +27,9 @@ from .wizard import router as _wizard
 from .worktrees import router as _worktrees
 
 router = APIRouter(prefix="/api", tags=["api"])
-router.include_router(_build)
+router.include_router(_dispatch)
+router.include_router(_runs)
+router.include_router(_ship_api)
 router.include_router(_pipeline)
 router.include_router(_control)
 router.include_router(_config)
