+++ agentception/static/app.js
@@ -13,8 +13,8 @@
 `))};if((x.mac||x.android)&&f.from==o-1&&/^\. ?$/.test(i.text)&&t.contentDOM.getAttribute("autocorrect")=="off"&&(f={from:l,to:a,insert:R.of([i.text.replace("."," ")])}),this.pendingContextChange=f,!t.state.readOnly){let u=this.to-this.from+(f.to-f.from+f.insert.length);kr(t,f,w.single(this.toEditorPos(i.selectionStart,u),this.toEditorPos(i.selectionEnd,u)))}this.pendingContextChange&&(this.revertPending(t.state),this.setSelection(t.state)),f.from<f.to&&!f.insert.length&&t.inputState.composing>=0&&!/[\\p{Alphabetic}\\p{Number}_]/.test(e.text.slice(Math.max(0,i.updateRangeStart-1),Math.min(e.text.length,i.updateRangeStart+1)))&&this.handlers.compositionend(i)},this.handlers.characterboundsupdate=i=>{let n=[],r=null;for(let o=this.toEditorPos(i.rangeStart),l=this.toEditorPos(i.rangeEnd);o<l;o++){let a=t.coordsForChar(o);r=a&&new DOMRect(a.left,a.top,a.right-a.left,a.bottom-a.top)||r||new DOMRect,n.push(r)}e.updateCharacterBounds(i.rangeStart,n)},this.handlers.textformatupdate=i=>{let n=[];for(let r of i.getTextFormats()){let o=r.underlineStyle,l=r.underlineThickness;if(!/none/i.test(o)&&!/none/i.test(l)){let a=this.toEditorPos(r.rangeStart),h=this.toEditorPos(r.rangeEnd);if(a<h){let c=`text-decoration: underline ${/^[a-z]/.test(o)?o+" ":o=="Dashed"?"dashed ":o=="Squiggle"?"wavy ":""}${/thin/i.test(l)?1:2}px`;n.push($.mark({attributes:{style:c}}).range(a,h))}}}t.dispatch({effects:wa.of($.set(n))})},this.handlers.compositionstart=()=>{t.inputState.composing<0&&(t.inputState.composing=0,t.inputState.compositionFirstChange=!0)},this.handlers.compositionend=()=>{if(t.inputState.composing=-1,t.inputState.compositionFirstChange=null,this.composing){let{drifted:i}=this.composing;this.composing=null,i&&this.reset(t.state)}};for(let i in this.handlers)e.addEventListener(i,this.handlers[i]);this.measureReq={read:i=>{this.editContext.updateControlBounds(i.contentDOM.getBoundingClientRect());let n=Oi(i.root);n&&n.rangeCount&&this.editContext.updateSelectionBounds(n.getRangeAt(0).getBoundingClientRect())}}}applyEdits(t){let e=0,i=!1,n=this.pendingContextChange;return t.changes.iterChanges((r,o,l,a,h)=>{if(i)return;let c=h.length-(o-r);if(n&&o>=n.to)if(n.from==r&&n.to==o&&n.insert.eq(h)){n=this.pendingContextChange=null,e+=c,this.to+=c;return}else n=null,this.revertPending(t.state);if(r+=e,o+=e,o<=this.from)this.from+=c,this.to+=c;else if(r<this.to){if(r<this.from||o>this.to||this.to-this.from+h.length>3e4){i=!0;return}this.editContext.updateText(this.toContextPos(r),this.toContextPos(o),h.toString()),this.to+=c}e+=c}),n&&!i&&this.revertPending(t.state),!i}update(t){let e=this.pendingContextChange,i=t.startState.selection.main;this.composing&&(this.composing.drifted||!t.changes.touchesRange(i.from,i.to)&&t.transactions.some(n=>!n.isUserEvent("input.type")&&n.changes.touchesRange(this.from,this.to)))?(this.composing.drifted=!0,this.composing.editorBase=t.changes.mapPos(this.composing.editorBase)):!this.applyEdits(t)||!this.rangeIsValid(t.state)?(this.pendingContextChange=null,this.reset(t.state)):(t.docChanged||t.selectionSet||e)&&this.setSelection(t.state),(t.geometryChanged||t.docChanged||t.selectionSet)&&t.view.requestMeasure(this.measureReq)}resetRange(t){let{head:e}=t.selection.main;this.from=Math.max(0,e-1e4),this.to=Math.min(t.doc.length,e+1e4)}reset(t){this.resetRange(t),this.editContext.updateText(0,this.editContext.text.length,t.doc.sliceString(this.from,this.to)),this.setSelection(t)}revertPending(t){let e=this.pendingContextChange;this.pendingContextChange=null,this.editContext.updateText(this.toContextPos(e.from),this.toContextPos(e.from+e.insert.length),t.doc.sliceString(e.from,e.to))}setSelection(t){let{main:e}=t.selection,i=this.toContextPos(Math.max(this.from,Math.min(this.to,e.anchor))),n=this.toContextPos(e.head);(this.editContext.selectionStart!=i||this.editContext.selectionEnd!=n)&&this.editContext.updateSelection(i,n)}rangeIsValid(t){let{head:e}=t.selection.main;return!(this.from>0&&e-this.from<500||this.to<t.doc.length&&this.to-e<500||this.to-this.from>1e4*3)}toEditorPos(t,e=this.to-this.from){t=Math.min(t,e);let i=this.composing;return i&&i.drifted?i.editorBase+(t-i.contextBase):t+this.from}toContextPos(t){let e=this.composing;return e&&e.drifted?e.contextBase+(t-e.editorBase):t-this.from}destroy(){for(let t in this.handlers)this.editContext.removeEventListener(t,this.handlers[t])}},A=class s{get state(){return this.viewState.state}get viewport(){return this.viewState.viewport}get visibleRanges(){return this.viewState.visibleRanges}get inView(){return this.viewState.inView}get composing(){return!!this.inputState&&this.inputState.composing>0}get compositionStarted(){return!!this.inputState&&this.inputState.composing>=0}get root(){return this._root}get win(){return this.dom.ownerDocument.defaultView||window}constructor(t={}){var e;this.plugins=[],this.pluginMap=new Map,this.editorAttrs={},this.contentAttrs={},this.bidiCache=[],this.destroyed=!1,this.updateState=2,this.measureScheduled=-1,this.measureRequests=[],this.contentDOM=document.createElement("div"),this.scrollDOM=document.createElement("div"),this.scrollDOM.tabIndex=-1,this.scrollDOM.className="cm-scroller",this.scrollDOM.appendChild(this.contentDOM),this.announceDOM=document.createElement("div"),this.announceDOM.className="cm-announced",this.announceDOM.setAttribute("aria-live","polite"),this.dom=document.createElement("div"),this.dom.appendChild(this.announceDOM),this.dom.appendChild(this.scrollDOM),t.parent&&t.parent.appendChild(this.dom);let{dispatch:i}=t;this.dispatchTransactions=t.dispatchTransactions||i&&(n=>n.forEach(r=>i(r,this)))||(n=>this.update(n)),this.dispatch=this.dispatch.bind(this),this._root=t.root||ef(t.parent)||document,this.viewState=new Os(this,t.state||W.create(t)),t.scrollTo&&t.scrollTo.is(ss)&&(this.viewState.scrollTarget=t.scrollTo.value.clip(this.viewState.state)),this.plugins=this.state.facet(Ve).map(n=>new yi(n));for(let n of this.plugins)n.update(this);this.observer=new lr(this),this.inputState=new Un(this),this.inputState.ensureHandlers(this.plugins),this.docView=new ws(this),this.mountStyles(),this.updateAttrs(),this.updateState=0,this.requestMeasure(),!((e=document.fonts)===null||e===void 0)&&e.ready&&document.fonts.ready.then(()=>{this.viewState.mustMeasureContent="refresh",this.requestMeasure()})}dispatch(...t){let e=t.length==1&&t[0]instanceof K?t:t.length==1&&Array.isArray(t[0])?t[0]:[this.state.update(...t)];this.dispatchTransactions(e,this)}update(t){if(this.updateState!=0)throw new Error("Calls to EditorView.update are not allowed while an update is in progress");let e=!1,i=!1,n,r=this.state;for(let u of t){if(u.startState!=r)throw new RangeError("Trying to update state with a transaction that doesn't start from the previous state.");r=u.state}if(this.destroyed){this.viewState.state=r;return}let o=this.hasFocus,l=0,a=null;t.some(u=>u.annotation(Na))?(this.inputState.notifiedFocused=o,l=1):o!=this.inputState.notifiedFocused&&(this.inputState.notifiedFocused=o,a=_a(r,o),a||(l=1));let h=this.observer.delayedAndroidKey,c=null;if(h?(this.observer.clearDelayedAndroidKey(),c=this.observer.readChange(),(c&&!this.state.doc.eq(r.doc)||!this.state.selection.eq(r.selection))&&(c=null)):this.observer.clear(),r.facet(W.phrases)!=this.state.facet(W.phrases))return this.setState(r);n=ys.create(this,r,t),n.flags|=l;let f=this.viewState.scrollTarget;try{this.updateState=2;for(let u of t){if(f&&(f=f.map(u.changes)),u.scrollIntoView){let{main:d}=u.state.selection;f=new mi(d.empty?d:w.cursor(d.head,d.head>d.anchor?-1:1))}for(let d of u.effects)d.is(ss)&&(f=d.value.clip(this.state))}this.viewState.update(n,f),this.bidiCache=Cs.update(this.bidiCache,n.changes),n.empty||(this.updatePlugins(n),this.inputState.update(n)),e=this.docView.update(n),this.state.facet(ui)!=this.styleModules&&this.mountStyles(),i=this.updateAttrs(),this.showAnnouncements(t),this.docView.updateSelection(e,t.some(u=>u.isUserEvent("select.pointer")))}finally{this.updateState=0}if(n.startState.facet(as)!=n.state.facet(as)&&(this.viewState.mustMeasureContent=!0),(e||i||f||this.viewState.mustEnforceCursorAssoc||this.viewState.mustMeasureContent)&&this.requestMeasure(),e&&this.docViewUpdate(),!n.empty)for(let u of this.state.facet(In))try{u(n)}catch(d){St(this.state,d,"update listener")}(a||c)&&Promise.resolve().then(()=>{a&&this.state==a.startState&&this.dispatch(a),c&&!Ma(this,c)&&h.force&&$e(this.contentDOM,h.key,h.keyCode)})}setState(t){if(this.updateState!=0)throw new Error("Calls to EditorView.setState are not allowed while an update is in progress");if(this.destroyed){this.viewState.state=t;return}this.updateState=2;let e=this.hasFocus;try{for(let i of this.plugins)i.destroy(this);this.viewState=new Os(this,t),this.plugins=t.facet(Ve).map(i=>new yi(i)),this.pluginMap.clear();for(let i of this.plugins)i.update(this);this.docView.destroy(),this.docView=new ws(this),this.inputState.ensureHandlers(this.plugins),this.mountStyles(),this.updateAttrs(),this.bidiCache=[]}finally{this.updateState=0}e&&this.focus(),this.requestMeasure()}updatePlugins(t){let e=t.startState.facet(Ve),i=t.state.facet(Ve);if(e!=i){let n=[];for(let r of i){let o=e.indexOf(r);if(o<0)n.push(new yi(r));else{let l=this.plugins[o];l.mustUpdate=t,n.push(l)}}for(let r of this.plugins)r.mustUpdate!=t&&r.destroy(this);this.plugins=n,this.pluginMap.clear()}else for(let n of this.plugins)n.mustUpdate=t;for(let n=0;n<this.plugins.length;n++)this.plugins[n].update(this);e!=i&&this.inputState.ensureHandlers(this.plugins)}docViewUpdate(){for(let t of this.plugins){let e=t.value;if(e&&e.docViewUpdate)try{e.docViewUpdate(this)}catch(i){St(this.state,i,"doc view update listener")}}}measure(t=!0){if(this.destroyed)return;if(this.measureScheduled>-1&&this.win.cancelAnimationFrame(this.measureScheduled),this.observer.delayedAndroidKey){this.measureScheduled=-1,this.requestMeasure();return}this.measureScheduled=0,t&&this.observer.forceFlush();let e=null,i=this.viewState.scrollParent,n=this.viewState.getScrollOffset(),{scrollAnchorPos:r,scrollAnchorHeight:o}=this.viewState;Math.abs(n-this.viewState.scrollOffset)>1&&(o=-1),this.viewState.scrollAnchorHeight=-1;try{for(let l=0;;l++){if(o<0)if(sa(i||this.win))r=-1,o=this.viewState.heightMap.height;else{let d=this.viewState.scrollAnchorAt(n);r=d.from,o=d.top}this.updateState=1;let a=this.viewState.measure();if(!a&&!this.measureRequests.length&&this.viewState.scrollTarget==null)break;if(l>5){console.warn(this.measureRequests.length?"Measure loop restarted more than 5 times":"Viewport failed to stabilize");break}let h=[];a&4||([this.measureRequests,h]=[h,this.measureRequests]);let c=h.map(d=>{try{return d.read(this)}catch(p){return St(this.state,p),ql}}),f=ys.create(this,this.state,[]),u=!1;f.flags|=a,e?e.flags|=a:e=f,this.updateState=2,f.empty||(this.updatePlugins(f),this.inputState.update(f),this.updateAttrs(),u=this.docView.update(f),u&&this.docViewUpdate());for(let d=0;d<h.length;d++)if(c[d]!=ql)try{let p=h[d];p.write&&p.write(c[d],this)}catch(p){St(this.state,p)}if(u&&this.docView.updateSelection(!0),!f.viewportChanged&&this.measureRequests.length==0){if(this.viewState.editorHeight)if(this.viewState.scrollTarget){this.docView.scrollIntoView(this.viewState.scrollTarget),this.viewState.scrollTarget=null,o=-1;continue}else{let p=((r<0?this.viewState.heightMap.height:this.viewState.lineBlockAt(r).top)-o)/this.scaleY;if((p>1||p<-1)&&(i==this.scrollDOM||this.hasFocus||Math.max(this.inputState.lastWheelEvent,this.inputState.lastTouchTime)>Date.now()-100)){n=n+p,i?i.scrollTop+=p:this.win.scrollBy(0,p),o=-1;continue}}break}}}finally{this.updateState=0,this.measureScheduled=-1}if(e&&!e.empty)for(let l of this.state.facet(In))l(e)}get themeClasses(){return rr+" "+(this.state.facet(nr)?Fa:Ha)+" "+this.state.facet(as)}updateAttrs(){let t=Ul(this,ka,{class:"cm-editor"+(this.hasFocus?" cm-focused ":" ")+this.themeClasses}),e={spellcheck:"false",autocorrect:"off",autocapitalize:"off",writingsuggestions:"false",translate:"no",contenteditable:this.state.facet(zt)?"true":"false",class:"cm-content",style:`${x.tabSize}: ${this.state.tabSize}`,role:"textbox","aria-multiline":"true"};this.state.readOnly&&(e["aria-readonly"]="true"),Ul(this,br,e);let i=this.observer.ignore(()=>{let n=Tl(this.contentDOM,this.contentAttrs,e),r=Tl(this.dom,this.editorAttrs,t);return n||r});return this.editorAttrs=t,this.contentAttrs=e,i}showAnnouncements(t){let e=!0;for(let i of t)for(let n of i.effects)if(n.is(s.announce)){e&&(this.announceDOM.textContent=""),e=!1;let r=this.announceDOM.appendChild(document.createElement("div"));r.textContent=n.value}}mountStyles(){this.styleModules=this.state.facet(ui);let t=this.state.facet(s.cspNonce);pt.mount(this.root,this.styleModules.concat(au).reverse(),t?{nonce:t}:void 0)}readMeasured(){if(this.updateState==2)throw new Error("Reading the editor layout isn't allowed during an update");this.updateState==0&&this.measureScheduled>-1&&this.measure(!1)}requestMeasure(t){if(this.measureScheduled<0&&(this.measureScheduled=this.win.requestAnimationFrame(()=>this.measure())),t){if(this.measureRequests.indexOf(t)>-1)return;if(t.key!=null){for(let e=0;e<this.measureRequests.length;e++)if(this.measureRequests[e].key===t.key){this.measureRequests[e]=t;return}}this.measureRequests.push(t)}}plugin(t){let e=this.pluginMap.get(t);return(e===void 0||e&&e.plugin!=t)&&this.pluginMap.set(t,e=this.plugins.find(i=>i.plugin==t)||null),e&&e.update(this).value}get documentTop(){return this.contentDOM.getBoundingClientRect().top+this.viewState.paddingTop}get documentPadding(){return{top:this.viewState.paddingTop,bottom:this.viewState.paddingBottom}}get scaleX(){return this.viewState.scaleX}get scaleY(){return this.viewState.scaleY}elementAtHeight(t){return this.readMeasured(),this.viewState.elementAtHeight(t)}lineBlockAtHeight(t){return this.readMeasured(),this.viewState.lineBlockAtHeight(t)}get viewportLineBlocks(){return this.viewState.viewportLines}lineBlockAt(t){return this.viewState.lineBlockAt(t)}get contentHeight(){return this.viewState.contentHeight}moveByChar(t,e,i){return vn(this,t,El(this,t,e,i))}moveByGroup(t,e){return vn(this,t,El(this,t,e,i=>Bf(this,t.head,i)))}visualLineSide(t,e){let i=this.bidiSpans(t),n=this.textDirectionAt(t.from),r=i[e?i.length-1:0];return w.cursor(r.side(e,n)+t.from,r.forward(!e,n)?1:-1)}moveToLineBoundary(t,e,i=!0){return Ef(this,t,e,i)}moveVertically(t,e,i){return vn(this,t,Lf(this,t,e,i))}domAtPos(t,e=1){return this.docView.domAtPos(t,e)}posAtDOM(t,e=0){return this.docView.posFromDOM(t,e)}posAtCoords(t,e=!0){this.readMeasured();let i=Qn(this,t,e);return i&&i.pos}posAndSideAtCoords(t,e=!0){return this.readMeasured(),Qn(this,t,e)}coordsAtPos(t,e=1){this.readMeasured();let i=this.docView.coordsAt(t,e);if(!i||i.left==i.right)return i;let n=this.state.doc.lineAt(t),r=this.bidiSpans(n),o=r[yt.find(r,t-n.from,-1,e)];return ms(i,o.dir==X.LTR==e>0)}coordsForChar(t){return this.readMeasured(),this.docView.coordsForChar(t)}get defaultCharacterWidth(){return this.viewState.heightOracle.charWidth}get defaultLineHeight(){return this.viewState.heightOracle.lineHeight}get textDirection(){return this.viewState.defaultTextDirection}textDirectionAt(t){return!this.state.facet(ya)||t<this.viewport.from||t>this.viewport.to?this.textDirection:(this.readMeasured(),this.docView.textDirectionAt(t))}get lineWrapping(){return this.viewState.heightOracle.lineWrapping}bidiSpans(t){if(t.length>fu)return ha(t.length);let e=this.textDirectionAt(t.from),i;for(let r of this.bidiCache)if(r.from==t.from&&r.dir==e&&(r.fresh||aa(r.isolates,i=Ml(this,t))))return r.order;i||(i=Ml(this,t));let n=cf(t.text,e,i);return this.bidiCache.push(new Cs(t.from,t.to,e,i,!0,n)),n}get hasFocus(){var t;return(this.dom.ownerDocument.hasFocus()||x.safari&&((t=this.inputState)===null||t===void 0?void 0:t.lastContextMenu)>Date.now()-3e4)&&this.root.activeElement==this.contentDOM}focus(){this.observer.ignore(()=>{ia(this.contentDOM),this.docView.updateSelection()})}setRoot(t){this._root!=t&&(this._root=t,this.observer.setWindow((t.nodeType==9?t:t.ownerDocument).defaultView||window),this.mountStyles())}destroy(){this.root.activeElement==this.contentDOM&&this.contentDOM.blur();for(let t of this.plugins)t.destroy(this);this.plugins=[],this.inputState.destroy(),this.docView.destroy(),this.dom.remove(),this.observer.destroy(),this.measureScheduled>-1&&this.win.cancelAnimationFrame(this.measureScheduled),this.destroyed=!0}static scrollIntoView(t,e={}){return ss.of(new mi(typeof t=="number"?w.cursor(t):t,e.y,e.x,e.yMargin,e.xMargin))}scrollSnapshot(){let{scrollTop:t,scrollLeft:e}=this.scrollDOM,i=this.viewState.scrollAnchorAt(t);return ss.of(new mi(w.cursor(i.from),"start","start",i.top-t,e,!0))}setTabFocusMode(t){t==null?this.inputState.tabFocusMode=this.inputState.tabFocusMode<0?0:-1:typeof t=="boolean"?this.inputState.tabFocusMode=t?0:-1:this.inputState.tabFocusMode!=0&&(this.inputState.tabFocusMode=Date.now()+t)}static domEventHandlers(t){return Xt.define(()=>({}),{eventHandlers:t})}static domEventObservers(t){return Xt.define(()=>({}),{eventObservers:t})}static theme(t,e){let i=pt.newName(),n=[as.of(i),ui.of(or(`.${i}`,t))];return e&&e.dark&&n.push(nr.of(!0)),n}static baseTheme(t){return ci.lowest(ui.of(or("."+rr,t,Wa)))}static findFromDOM(t){var e;let i=t.querySelector(".cm-content"),n=i&&z.get(i)||z.get(t);return((e=n?.root)===null||e===void 0?void 0:e.view)||null}};A.styleModule=ui;A.inputHandler=ga;A.clipboardInputFilter=mr;A.clipboardOutputFilter=yr;A.scrollHandler=ba;A.focusChangeEffect=ma;A.perLineTextDirection=ya;A.exceptionSink=pa;A.updateListener=In;A.editable=zt;A.mouseSelectionStyle=da;A.dragMovesSelection=ua;A.clickAddsSelectionRange=fa;A.decorations=Ms;A.blockWrappers=xa;A.outerDecorations=wr;A.atomicRanges=Ti;A.bidiIsolatedRanges=va;A.scrollMargins=Sa;A.darkTheme=nr;A.cspNonce=S.define({combine:s=>s.length?s[0]:""});A.contentAttributes=br;A.editorAttributes=ka;A.lineWrapping=A.contentAttributes.of({class:"cm-lineWrapping"});A.announce=F.define();var fu=4096,ql={},Cs=class s{constructor(t,e,i,n,r,o){this.from=t,this.to=e,this.dir=i,this.isolates=n,this.fresh=r,this.order=o}static update(t,e){if(e.empty&&!t.some(r=>r.fresh))return t;let i=[],n=t.length?t[t.length-1].dir:X.LTR;for(let r=Math.max(0,t.length-10);r<t.length;r++){let o=t[r];o.dir==n&&!e.touchesRange(o.from,o.to)&&i.push(new s(e.mapPos(o.from,1),e.mapPos(o.to,-1),o.dir,o.isolates,!1,o.order))}return i}};function Ul(s,t,e){for(let i=s.state.facet(t),n=i.length-1;n>=0;n--){let r=i[n],o=typeof r=="function"?r(s):r;o&&dr(o,e)}return e}var uu=x.mac?"mac":x.windows?"win":x.linux?"linux":"key";function du(s,t){let e=s.split(/-(?!$)/),i=e[e.length-1];i=="Space"&&(i=" ");let n,r,o,l;for(let a=0;a<e.length-1;++a){let h=e[a];if(/^(cmd|meta|m)$/i.test(h))l=!0;else if(/^a(lt)?$/i.test(h))n=!0;else if(/^(c|ctrl|control)$/i.test(h))r=!0;else if(/^s(hift)?$/i.test(h))o=!0;else if(/^mod$/i.test(h))t=="mac"?l=!0:r=!0;else throw new Error("Unrecognized modifier name: "+h)}return n&&(i="Alt-"+i),r&&(i="Ctrl-"+i),l&&(i="Meta-"+i),o&&(i="Shift-"+i),i}function hs(s,t,e){return t.altKey&&(s="Alt-"+s),t.ctrlKey&&(s="Ctrl-"+s),t.metaKey&&(s="Meta-"+s),e!==!1&&t.shiftKey&&(s="Shift-"+s),s}var pu=ci.default(A.domEventHandlers({keydown(s,t){return bu(gu(t.state),s,t,"editor")}})),xr=S.define({enables:pu}),Gl=new WeakMap;function gu(s){let t=s.facet(xr),e=Gl.get(t);return e||Gl.set(t,e=yu(t.reduce((i,n)=>i.concat(n),[]))),e}var Jt=null,mu=4e3;function yu(s,t=uu){let e=Object.create(null),i=Object.create(null),n=(o,l)=>{let a=i[o];if(a==null)i[o]=l;else if(a!=l)throw new Error("Key binding "+o+" is used both as a regular binding and as a multi-stroke prefix")},r=(o,l,a,h,c)=>{var f,u;let d=e[o]||(e[o]=Object.create(null)),p=l.split(/ (?!$)/).map(b=>du(b,t));for(let b=1;b<p.length;b++){let v=p.slice(0,b).join(" ");n(v,!0),d[v]||(d[v]={preventDefault:!0,stopPropagation:!1,run:[C=>{let B=Jt={view:C,prefix:v,scope:o};return setTimeout(()=>{Jt==B&&(Jt=null)},mu),!0}]})}let g=p.join(" ");n(g,!1);let y=d[g]||(d[g]={preventDefault:!1,stopPropagation:!1,run:((u=(f=d._any)===null||f===void 0?void 0:f.run)===null||u===void 0?void 0:u.slice())||[]});a&&y.run.push(a),h&&(y.preventDefault=!0),c&&(y.stopPropagation=!0)};for(let o of s){let l=o.scope?o.scope.split(" "):["editor"];if(o.any)for(let h of l){let c=e[h]||(e[h]=Object.create(null));c._any||(c._any={preventDefault:!1,stopPropagation:!1,run:[]});let{any:f}=o;for(let u in c)c[u].run.push(d=>f(d,hr))}let a=o[t]||o.key;if(a)for(let h of l)r(h,a,o.run,o.preventDefault,o.stopPropagation),o.shift&&r(h,"Shift-"+a,o.shift,o.preventDefault,o.stopPropagation)}return e}var hr=null;function bu(s,t,e,i){hr=t;let n=vl(t),r=ol(n,0),o=ll(r)==n.length&&n!=" ",l="",a=!1,h=!1,c=!1;Jt&&Jt.view==e&&Jt.scope==i&&(l=Jt.prefix+" ",Ea.indexOf(t.keyCode)<0&&(h=!0,Jt=null));let f=new Set,u=y=>{if(y){for(let b of y.run)if(!f.has(b)&&(f.add(b),b(e)))return y.stopPropagation&&(c=!0),!0;y.preventDefault&&(y.stopPropagation&&(c=!0),h=!0)}return!1},d=s[i],p,g;return d&&(u(d[l+hs(n,t,!o)])?a=!0:o&&(t.altKey||t.metaKey||t.ctrlKey)&&!(x.windows&&t.ctrlKey&&t.altKey)&&!(x.mac&&t.altKey&&!(t.ctrlKey||t.metaKey))&&(p=$t[t.keyCode])&&p!=n?(u(d[l+hs(p,t,!0)])||t.shiftKey&&(g=Ne[t.keyCode])!=n&&g!=p&&u(d[l+hs(g,t,!1)]))&&(a=!0):o&&t.shiftKey&&u(d[l+hs(n,t,!0)])&&(a=!0),!a&&u(d._any)&&(a=!0)),h&&(a=!0),a&&c&&t.stopPropagation(),hr=null,a}var Cg=/x/.unicode!=null?"gu":"g";function $a(){return ku}var wu=$.line({class:"cm-activeLine"}),ku=Xt.fromClass(class{constructor(s){this.decorations=this.getDeco(s)}update(s){(s.docChanged||s.selectionSet)&&(this.decorations=this.getDeco(s.view))}getDeco(s){let t=-1,e=[];for(let i of s.state.selection.ranges){let n=s.lineBlockAt(i.head);n.from>t&&(e.push(wu.range(n.from)),t=n.from)}return $.set(e)}},{decorations:s=>s.decorations});var It=class extends Et{compare(t){return this==t||this.constructor==t.constructor&&this.eq(t)}eq(t){return!1}destroy(t){}};It.prototype.elementClass="";It.prototype.toDOM=void 0;It.prototype.mapMode=nt.TrackBefore;It.prototype.startSide=It.prototype.endSide=-1;It.prototype.point=!0;var On=S.define(),xu=S.define();var us=S.define();var cr=S.define({combine:s=>s.some(t=>t)});function vu(s){let t=[Su];return s&&s.fixed===!1&&t.push(cr.of(!0)),t}var Su=Xt.fromClass(class{constructor(s){this.view=s,this.domAfter=null,this.prevViewport=s.viewport,this.dom=document.createElement("div"),this.dom.className="cm-gutters cm-gutters-before",this.dom.setAttribute("aria-hidden","true"),this.dom.style.minHeight=this.view.contentHeight/this.view.scaleY+"px",this.gutters=s.state.facet(us).map(t=>new Ts(s,t)),this.fixed=!s.state.facet(cr);for(let t of this.gutters)t.config.side=="after"?this.getDOMAfter().appendChild(t.dom):this.dom.appendChild(t.dom);this.fixed&&(this.dom.style.position="sticky"),this.syncGutters(!1),s.scrollDOM.insertBefore(this.dom,s.contentDOM)}getDOMAfter(){return this.domAfter||(this.domAfter=document.createElement("div"),this.domAfter.className="cm-gutters cm-gutters-after",this.domAfter.setAttribute("aria-hidden","true"),this.domAfter.style.minHeight=this.view.contentHeight/this.view.scaleY+"px",this.domAfter.style.position=this.fixed?"sticky":"",this.view.scrollDOM.appendChild(this.domAfter)),this.domAfter}update(s){if(this.updateGutters(s)){let t=this.prevViewport,e=s.view.viewport,i=Math.min(t.to,e.to)-Math.max(t.from,e.from);this.syncGutters(i<(e.to-e.from)*.8)}if(s.geometryChanged){let t=this.view.contentHeight/this.view.scaleY+"px";this.dom.style.minHeight=t,this.domAfter&&(this.domAfter.style.minHeight=t)}this.view.state.facet(cr)!=!this.fixed&&(this.fixed=!this.fixed,this.dom.style.position=this.fixed?"sticky":"",this.domAfter&&(this.domAfter.style.position=this.fixed?"sticky":"")),this.prevViewport=s.view.viewport}syncGutters(s){let t=this.dom.nextSibling;s&&(this.dom.remove(),this.domAfter&&this.domAfter.remove());let e=L.iter(this.view.state.facet(On),this.view.viewport.from),i=[],n=this.gutters.map(r=>new ur(r,this.view.viewport,-this.view.documentPadding.top));for(let r of this.view.viewportLineBlocks)if(i.length&&(i=[]),Array.isArray(r.type)){let o=!0;for(let l of r.type)if(l.type==ot.Text&&o){fr(e,i,l.from);for(let a of n)a.line(this.view,l,i);o=!1}else if(l.widget)for(let a of n)a.widget(this.view,l)}else if(r.type==ot.Text){fr(e,i,r.from);for(let o of n)o.line(this.view,r,i)}else if(r.widget)for(let o of n)o.widget(this.view,r);for(let r of n)r.finish();s&&(this.view.scrollDOM.insertBefore(this.dom,t),this.domAfter&&this.view.scrollDOM.appendChild(this.domAfter))}updateGutters(s){let t=s.startState.facet(us),e=s.state.facet(us),i=s.docChanged||s.heightChanged||s.viewportChanged||!L.eq(s.startState.facet(On),s.state.facet(On),s.view.viewport.from,s.view.viewport.to);if(t==e)for(let n of this.gutters)n.update(s)&&(i=!0);else{i=!0;let n=[];for(let r of e){let o=t.indexOf(r);o<0?n.push(new Ts(this.view,r)):(this.gutters[o].update(s),n.push(this.gutters[o]))}for(let r of this.gutters)r.dom.remove(),n.indexOf(r)<0&&r.destroy();for(let r of n)r.config.side=="after"?this.getDOMAfter().appendChild(r.dom):this.dom.appendChild(r.dom);this.gutters=n}return i}destroy(){for(let s of this.gutters)s.destroy();this.dom.remove(),this.domAfter&&this.domAfter.remove()}},{provide:s=>A.scrollMargins.of(t=>{let e=t.plugin(s);if(!e||e.gutters.length==0||!e.fixed)return null;let i=e.dom.offsetWidth*t.scaleX,n=e.domAfter?e.domAfter.offsetWidth*t.scaleX:0;return t.textDirection==X.LTR?{left:i,right:n}:{right:i,left:n}})});function Kl(s){return Array.isArray(s)?s:[s]}function fr(s,t,e){for(;s.value&&s.from<=e;)s.from==e&&t.push(s.value),s.next()}var ur=class{constructor(t,e,i){this.gutter=t,this.height=i,this.i=0,this.cursor=L.iter(t.markers,e.from)}addElement(t,e,i){let{gutter:n}=this,r=(e.top-this.height)/t.scaleY,o=e.height/t.scaleY;if(this.i==n.elements.length){let l=new As(t,o,r,i);n.elements.push(l),n.dom.appendChild(l.dom)}else n.elements[this.i].update(t,o,r,i);this.height=e.bottom,this.i++}line(t,e,i){let n=[];fr(this.cursor,n,e.from),i.length&&(n=n.concat(i));let r=this.gutter.config.lineMarker(t,e,n);r&&n.unshift(r);let o=this.gutter;n.length==0&&!o.config.renderEmptyElements||this.addElement(t,e,n)}widget(t,e){let i=this.gutter.config.widgetMarker(t,e.widget,e),n=i?[i]:null;for(let r of t.state.facet(xu)){let o=r(t,e.widget,e);o&&(n||(n=[])).push(o)}n&&this.addElement(t,e,n)}finish(){let t=this.gutter;for(;t.elements.length>this.i;){let e=t.elements.pop();t.dom.removeChild(e.dom),e.destroy()}}},Ts=class{constructor(t,e){this.view=t,this.config=e,this.elements=[],this.spacer=null,this.dom=document.createElement("div"),this.dom.className="cm-gutter"+(this.config.class?" "+this.config.class:"");for(let i in e.domEventHandlers)this.dom.addEventListener(i,n=>{let r=n.target,o;if(r!=this.dom&&this.dom.contains(r)){for(;r.parentNode!=this.dom;)r=r.parentNode;let a=r.getBoundingClientRect();o=(a.top+a.bottom)/2}else o=n.clientY;let l=t.lineBlockAtHeight(o-t.documentTop);e.domEventHandlers[i](t,l,n)&&n.preventDefault()});this.markers=Kl(e.markers(t)),e.initialSpacer&&(this.spacer=new As(t,0,0,[e.initialSpacer(t)]),this.dom.appendChild(this.spacer.dom),this.spacer.dom.style.cssText+="visibility: hidden; pointer-events: none")}update(t){let e=this.markers;if(this.markers=Kl(this.config.markers(t.view)),this.spacer&&this.config.updateSpacer){let n=this.config.updateSpacer(this.spacer.markers[0],t);n!=this.spacer.markers[0]&&this.spacer.update(t.view,0,0,[n])}let i=t.view.viewport;return!L.eq(this.markers,e,i.from,i.to)||(this.config.lineMarkerChange?this.config.lineMarkerChange(t):!1)}destroy(){for(let t of this.elements)t.destroy()}},As=class{constructor(t,e,i,n){this.height=-1,this.above=0,this.markers=[],this.dom=document.createElement("div"),this.dom.className="cm-gutterElement",this.update(t,e,i,n)}update(t,e,i,n){this.height!=e&&(this.height=e,this.dom.style.height=e+"px"),this.above!=i&&(this.dom.style.marginTop=(this.above=i)?i+"px":""),Ou(this.markers,n)||this.setMarkers(t,n)}setMarkers(t,e){let i="cm-gutterElement",n=this.dom.firstChild;for(let r=0,o=0;;){let l=o,a=r<e.length?e[r++]:null,h=!1;if(a){let c=a.elementClass;c&&(i+=" "+c);for(let f=o;f<this.markers.length;f++)if(this.markers[f].compare(a)){l=f,h=!0;break}}else l=this.markers.length;for(;o<l;){let c=this.markers[o++];if(c.toDOM){c.destroy(n);let f=n.nextSibling;n.remove(),n=f}}if(!a)break;a.toDOM&&(h?n=n.nextSibling:this.dom.insertBefore(a.toDOM(t),n)),h&&o++}this.dom.className=i,this.markers=e}destroy(){this.setMarkers(null,[])}};function Ou(s,t){if(s.length!=t.length)return!1;for(let e=0;e<s.length;e++)if(!s[e].compare(t[e]))return!1;return!0}var Cu=S.define(),Tu=S.define(),Fe=S.define({combine(s){return fi(s,{formatNumber:String,domEventHandlers:{}},{domEventHandlers(t,e){let i=Object.assign({},t);for(let n in e){let r=i[n],o=e[n];i[n]=r?(l,a,h)=>r(l,a,h)||o(l,a,h):o}return i}})}}),xi=class extends It{constructor(t){super(),this.number=t}eq(t){return this.number==t.number}toDOM(){return document.createTextNode(this.number)}};function Cn(s,t){return s.state.facet(Fe).formatNumber(t,s.state)}var Au=us.compute([Fe],s=>({class:"cm-lineNumbers",renderEmptyElements:!1,markers(t){return t.state.facet(Cu)},lineMarker(t,e,i){return i.some(n=>n.toDOM)?null:new xi(Cn(t,t.state.doc.lineAt(e.from).number))},widgetMarker:(t,e,i)=>{for(let n of t.state.facet(Tu)){let r=n(t,e,i);if(r)return r}return null},lineMarkerChange:t=>t.startState.facet(Fe)!=t.state.facet(Fe),initialSpacer(t){return new xi(Cn(t,Yl(t.state.doc.lines)))},updateSpacer(t,e){let i=Cn(e.view,Yl(e.view.state.doc.lines));return i==t.number?t:new xi(i)},domEventHandlers:s.facet(Fe).domEventHandlers,side:"before"}));function za(s={}){return[Fe.of(s),vu(),Au]}function Yl(s){let t=9;for(;t<s;)t=t*10+9;return t}var Pu=0,Ai=class{constructor(t,e){this.from=t,this.to=e}},D=class{constructor(t={}){this.id=Pu++,this.perNode=!!t.perNode,this.deserialize=t.deserialize||(()=>{throw new Error("This node type doesn't define a deserialize function")}),this.combine=t.combine||null}add(t){if(this.perNode)throw new RangeError("Can't add per-node props to node types");return typeof t!="function"&&(t=tt.match(t)),e=>{let i=t(e);return i===void 0?null:[this,i]}}};D.closedBy=new D({deserialize:s=>s.split(" ")});D.openedBy=new D({deserialize:s=>s.split(" ")});D.group=new D({deserialize:s=>s.split(" ")});D.isolate=new D({deserialize:s=>{if(s&&s!="rtl"&&s!="ltr"&&s!="auto")throw new RangeError("Invalid value for isolate: "+s);return s||"auto"}});D.contextHash=new D({perNode:!0});D.lookAhead=new D({perNode:!0});D.mounted=new D({perNode:!0});var xe=class{constructor(t,e,i,n=!1){this.tree=t,this.overlay=e,this.parser=i,this.bracketed=n}static get(t){return t&&t.props&&t.props[D.mounted.id]}},Mu=Object.create(null),tt=class s{constructor(t,e,i,n=0){this.name=t,this.props=e,this.id=i,this.flags=n}static define(t){let e=t.props&&t.props.length?Object.create(null):Mu,i=(t.top?1:0)|(t.skipped?2:0)|(t.error?4:0)|(t.name==null?8:0),n=new s(t.name||"",e,t.id,i);if(t.props){for(let r of t.props)if(Array.isArray(r)||(r=r(n)),r){if(r[0].perNode)throw new RangeError("Can't store a per-node prop on a node type");e[r[0].id]=r[1]}}return n}prop(t){return this.props[t.id]}get isTop(){return(this.flags&1)>0}get isSkipped(){return(this.flags&2)>0}get isError(){return(this.flags&4)>0}get isAnonymous(){return(this.flags&8)>0}is(t){if(typeof t=="string"){if(this.name==t)return!0;let e=this.prop(D.group);return e?e.indexOf(t)>-1:!1}return this.id==t}static match(t){let e=Object.create(null);for(let i in t)for(let n of i.split(" "))e[n]=t[i];return i=>{for(let n=i.prop(D.group),r=-1;r<(n?n.length:0);r++){let o=e[r<0?i.name:n[r]];if(o)return o}}}};tt.none=new tt("",Object.create(null),0,8);var Pi=class s{constructor(t){this.types=t;for(let e=0;e<t.length;e++)if(t[e].id!=e)throw new RangeError("Node type ids should correspond to array positions when creating a node set")}extend(...t){let e=[];for(let i of this.types){let n=null;for(let r of t){let o=r(i);if(o){n||(n=Object.assign({},i.props));let l=o[1],a=o[0];a.combine&&a.id in n&&(l=a.combine(n[a.id],l)),n[a.id]=l}}e.push(n?new tt(i.name,n,i.id,i.flags):i)}return new s(e)}},Rs=new WeakMap,Qa=new WeakMap,Q;(function(s){s[s.ExcludeBuffers=1]="ExcludeBuffers",s[s.IncludeAnonymous=2]="IncludeAnonymous",s[s.IgnoreMounts=4]="IgnoreMounts",s[s.IgnoreOverlays=8]="IgnoreOverlays",s[s.EnterBracketed=16]="EnterBracketed"})(Q||(Q={}));var U=class s{constructor(t,e,i,n,r){if(this.type=t,this.children=e,this.positions=i,this.length=n,this.props=null,r&&r.length){this.props=Object.create(null);for(let[o,l]of r)this.props[typeof o=="number"?o:o.id]=l}}toString(){let t=xe.get(this);if(t&&!t.overlay)return t.tree.toString();let e="";for(let i of this.children){let n=i.toString();n&&(e&&(e+=","),e+=n)}return this.type.name?(/\W/.test(this.type.name)&&!this.type.isError?JSON.stringify(this.type.name):this.type.name)+(e.length?"("+e+")":""):e}cursor(t=0){return new Ri(this.topNode,t)}cursorAt(t,e=0,i=0){let n=Rs.get(this)||this.topNode,r=new Ri(n);return r.moveTo(t,e),Rs.set(this,r._tree),r}get topNode(){return new Nt(this,0,0,null)}resolve(t,e=0){let i=Mi(Rs.get(this)||this.topNode,t,e,!1);return Rs.set(this,i),i}resolveInner(t,e=0){let i=Mi(Qa.get(this)||this.topNode,t,e,!0);return Qa.set(this,i),i}resolveStack(t,e=0){return Du(this,t,e)}iterate(t){let{enter:e,leave:i,from:n=0,to:r=this.length}=t,o=t.mode||0,l=(o&Q.IncludeAnonymous)>0;for(let a=this.cursor(o|Q.IncludeAnonymous);;){let h=!1;if(a.from<=r&&a.to>=n&&(!l&&a.type.isAnonymous||e(a)!==!1)){if(a.firstChild())continue;h=!0}for(;h&&i&&(l||!a.type.isAnonymous)&&i(a),!a.nextSibling();){if(!a.parent())return;h=!0}}}prop(t){return t.perNode?this.props?this.props[t.id]:void 0:this.type.prop(t)}get propValues(){let t=[];if(this.props)for(let e in this.props)t.push([+e,this.props[e]]);return t}balance(t={}){return this.children.length<=8?this:Pr(tt.none,this.children,this.positions,0,this.children.length,0,this.length,(e,i,n)=>new s(this.type,e,i,n,this.propValues),t.makeTree||((e,i,n)=>new s(tt.none,e,i,n)))}static build(t){return Ru(t)}};U.empty=new U(tt.none,[],[],0);var vr=class s{constructor(t,e){this.buffer=t,this.index=e}get id(){return this.buffer[this.index-4]}get start(){return this.buffer[this.index-3]}get end(){return this.buffer[this.index-2]}get size(){return this.buffer[this.index-1]}get pos(){return this.index}next(){this.index-=4}fork(){return new s(this.buffer,this.index)}},ie=class s{constructor(t,e,i){this.buffer=t,this.length=e,this.set=i}get type(){return tt.none}toString(){let t=[];for(let e=0;e<this.buffer.length;)t.push(this.childString(e)),e=this.buffer[e+3];return t.join(",")}childString(t){let e=this.buffer[t],i=this.buffer[t+3],n=this.set.types[e],r=n.name;if(/\W/.test(r)&&!n.isError&&(r=JSON.stringify(r)),t+=4,i==t)return r;let o=[];for(;t<i;)o.push(this.childString(t)),t=this.buffer[t+3];return r+"("+o.join(",")+")"}findChild(t,e,i,n,r){let{buffer:o}=this,l=-1;for(let a=t;a!=e&&!(qa(r,n,o[a+1],o[a+2])&&(l=a,i>0));a=o[a+3]);return l}slice(t,e,i){let n=this.buffer,r=new Uint16Array(e-t),o=0;for(let l=t,a=0;l<e;){r[a++]=n[l++],r[a++]=n[l++]-i;let h=r[a++]=n[l++]-i;r[a++]=n[l++]-t,o=Math.max(o,h)}return new s(r,o,this.set)}};function qa(s,t,e,i){switch(s){case-2:return e<t;case-1:return i>=t&&e<t;case 0:return e<t&&i>t;case 1:return e<=t&&i>t;case 2:return i>t;case 4:return!0}}function Mi(s,t,e,i){for(var n;s.from==s.to||(e<1?s.from>=t:s.from>t)||(e>-1?s.to<=t:s.to<t);){let o=!i&&s instanceof Nt&&s.index<0?null:s.parent;if(!o)return s;s=o}let r=i?0:Q.IgnoreOverlays;if(i)for(let o=s,l=o.parent;l;o=l,l=o.parent)o instanceof Nt&&o.index<0&&((n=l.enter(t,e,r))===null||n===void 0?void 0:n.from)!=o.from&&(s=l);for(;;){let o=s.enter(t,e,r);if(!o)return s;s=o}}var Bs=class{cursor(t=0){return new Ri(this,t)}getChild(t,e=null,i=null){let n=ja(this,t,e,i);return n.length?n[0]:null}getChildren(t,e=null,i=null){return ja(this,t,e,i)}resolve(t,e=0){return Mi(this,t,e,!1)}resolveInner(t,e=0){return Mi(this,t,e,!0)}matchContext(t){return Sr(this.parent,t)}enterUnfinishedNodesBefore(t){let e=this.childBefore(t),i=this;for(;e;){let n=e.lastChild;if(!n||n.to!=e.to)break;n.type.isError&&n.from==n.to?(i=e,e=n.prevSibling):e=n}return i}get node(){return this}get next(){return this.parent}},Nt=class s extends Bs{constructor(t,e,i,n){super(),this._tree=t,this.from=e,this.index=i,this._parent=n}get type(){return this._tree.type}get name(){return this._tree.type.name}get to(){return this.from+this._tree.length}nextChild(t,e,i,n,r=0){for(let o=this;;){for(let{children:l,positions:a}=o._tree,h=e>0?l.length:-1;t!=h;t+=e){let c=l[t],f=a[t]+o.from,u;if(!(!(r&Q.EnterBracketed&&c instanceof U&&(u=xe.get(c))&&!u.overlay&&u.bracketed&&i>=f&&i<=f+c.length)&&!qa(n,i,f,f+c.length))){if(c instanceof ie){if(r&Q.ExcludeBuffers)continue;let d=c.findChild(0,c.buffer.length,e,i-f,n);if(d>-1)return new Di(new Or(o,c,t,f),null,d)}else if(r&Q.IncludeAnonymous||!c.type.isAnonymous||Ar(c)){let d;if(!(r&Q.IgnoreMounts)&&(d=xe.get(c))&&!d.overlay)return new s(d.tree,f,t,o);let p=new s(c,f,t,o);return r&Q.IncludeAnonymous||!p.type.isAnonymous?p:p.nextChild(e<0?c.children.length-1:0,e,i,n,r)}}}if(r&Q.IncludeAnonymous||!o.type.isAnonymous||(o.index>=0?t=o.index+e:t=e<0?-1:o._parent._tree.children.length,o=o._parent,!o))return null}}get firstChild(){return this.nextChild(0,1,0,4)}get lastChild(){return this.nextChild(this._tree.children.length-1,-1,0,4)}childAfter(t){return this.nextChild(0,1,t,2)}childBefore(t){return this.nextChild(this._tree.children.length-1,-1,t,-2)}prop(t){return this._tree.prop(t)}enter(t,e,i=0){let n;if(!(i&Q.IgnoreOverlays)&&(n=xe.get(this._tree))&&n.overlay){let r=t-this.from,o=i&Q.EnterBracketed&&n.bracketed;for(let{from:l,to:a}of n.overlay)if((e>0||o?l<=r:l<r)&&(e<0||o?a>=r:a>r))return new s(n.tree,n.overlay[0].from+this.from,-1,this)}return this.nextChild(0,1,t,e,i)}nextSignificantParent(){let t=this;for(;t.type.isAnonymous&&t._parent;)t=t._parent;return t}get parent(){return this._parent?this._parent.nextSignificantParent():null}get nextSibling(){return this._parent&&this.index>=0?this._parent.nextChild(this.index+1,1,0,4):null}get prevSibling(){return this._parent&&this.index>=0?this._parent.nextChild(this.index-1,-1,0,4):null}get tree(){return this._tree}toTree(){return this._tree}toString(){return this._tree.toString()}};function ja(s,t,e,i){let n=s.cursor(),r=[];if(!n.firstChild())return r;if(e!=null){for(let o=!1;!o;)if(o=n.type.is(e),!n.nextSibling())return r}for(;;){if(i!=null&&n.type.is(i))return r;if(n.type.is(t)&&r.push(n.node),!n.nextSibling())return i==null?r:[]}}function Sr(s,t,e=t.length-1){for(let i=s;e>=0;i=i.parent){if(!i)return!1;if(!i.type.isAnonymous){if(t[e]&&t[e]!=i.name)return!1;e--}}return!0}var Or=class{constructor(t,e,i,n){this.parent=t,this.buffer=e,this.index=i,this.start=n}},Di=class s extends Bs{get name(){return this.type.name}get from(){return this.context.start+this.context.buffer.buffer[this.index+1]}get to(){return this.context.start+this.context.buffer.buffer[this.index+2]}constructor(t,e,i){super(),this.context=t,this._parent=e,this.index=i,this.type=t.buffer.set.types[t.buffer.buffer[i]]}child(t,e,i){let{buffer:n}=this.context,r=n.findChild(this.index+4,n.buffer[this.index+3],t,e-this.context.start,i);return r<0?null:new s(this.context,this,r)}get firstChild(){return this.child(1,0,4)}get lastChild(){return this.child(-1,0,4)}childAfter(t){return this.child(1,t,2)}childBefore(t){return this.child(-1,t,-2)}prop(t){return this.type.prop(t)}enter(t,e,i=0){if(i&Q.ExcludeBuffers)return null;let{buffer:n}=this.context,r=n.findChild(this.index+4,n.buffer[this.index+3],e>0?1:-1,t-this.context.start,e);return r<0?null:new s(this.context,this,r)}get parent(){return this._parent||this.context.parent.nextSignificantParent()}externalSibling(t){return this._parent?null:this.context.parent.nextChild(this.context.index+t,t,0,4)}get nextSibling(){let{buffer:t}=this.context,e=t.buffer[this.index+3];return e<(this._parent?t.buffer[this._parent.index+3]:t.buffer.length)?new s(this.context,this._parent,e):this.externalSibling(1)}get prevSibling(){let{buffer:t}=this.context,e=this._parent?this._parent.index+4:0;return this.index==e?this.externalSibling(-1):new s(this.context,this._parent,t.findChild(e,this.index,-1,0,4))}get tree(){return null}toTree(){let t=[],e=[],{buffer:i}=this.context,n=this.index+4,r=i.buffer[this.index+3];if(r>n){let o=i.buffer[this.index+1];t.push(i.slice(n,r,o)),e.push(0)}return new U(this.type,t,e,this.to-this.from)}toString(){return this.context.buffer.childString(this.index)}};function Ua(s){if(!s.length)return null;let t=0,e=s[0];for(let r=1;r<s.length;r++){let o=s[r];(o.from>e.from||o.to<e.to)&&(e=o,t=r)}let i=e instanceof Nt&&e.index<0?null:e.parent,n=s.slice();return i?n[t]=i:n.splice(t,1),new Cr(n,e)}var Cr=class{constructor(t,e){this.heads=t,this.node=e}get next(){return Ua(this.heads)}};function Du(s,t,e){let i=s.resolveInner(t,e),n=null;for(let r=i instanceof Nt?i:i.context.parent;r;r=r.parent)if(r.index<0){let o=r.parent;(n||(n=[i])).push(o.resolve(t,e)),r=o}else{let o=xe.get(r.tree);if(o&&o.overlay&&o.overlay[0].from<=t&&o.overlay[o.overlay.length-1].to>=t){let l=new Nt(o.tree,o.overlay[0].from+r.from,-1,r);(n||(n=[i])).push(Mi(l,t,e,!1))}}return n?Ua(n):i}var Ri=class{get name(){return this.type.name}constructor(t,e=0){if(this.buffer=null,this.stack=[],this.index=0,this.bufferNode=null,this.mode=e&~Q.EnterBracketed,t instanceof Nt)this.yieldNode(t);else{this._tree=t.context.parent,this.buffer=t.context;for(let i=t._parent;i;i=i._parent)this.stack.unshift(i.index);this.bufferNode=t,this.yieldBuf(t.index)}}yieldNode(t){return t?(this._tree=t,this.type=t.type,this.from=t.from,this.to=t.to,!0):!1}yieldBuf(t,e){this.index=t;let{start:i,buffer:n}=this.buffer;return this.type=e||n.set.types[n.buffer[t]],this.from=i+n.buffer[t+1],this.to=i+n.buffer[t+2],!0}yield(t){return t?t instanceof Nt?(this.buffer=null,this.yieldNode(t)):(this.buffer=t.context,this.yieldBuf(t.index,t.type)):!1}toString(){return this.buffer?this.buffer.buffer.childString(this.index):this._tree.toString()}enterChild(t,e,i){if(!this.buffer)return this.yield(this._tree.nextChild(t<0?this._tree._tree.children.length-1:0,t,e,i,this.mode));let{buffer:n}=this.buffer,r=n.findChild(this.index+4,n.buffer[this.index+3],t,e-this.buffer.start,i);return r<0?!1:(this.stack.push(this.index),this.yieldBuf(r))}firstChild(){return this.enterChild(1,0,4)}lastChild(){return this.enterChild(-1,0,4)}childAfter(t){return this.enterChild(1,t,2)}childBefore(t){return this.enterChild(-1,t,-2)}enter(t,e,i=this.mode){return this.buffer?i&Q.ExcludeBuffers?!1:this.enterChild(1,t,e):this.yield(this._tree.enter(t,e,i))}parent(){if(!this.buffer)return this.yieldNode(this.mode&Q.IncludeAnonymous?this._tree._parent:this._tree.parent);if(this.stack.length)return this.yieldBuf(this.stack.pop());let t=this.mode&Q.IncludeAnonymous?this.buffer.parent:this.buffer.parent.nextSignificantParent();return this.buffer=null,this.yieldNode(t)}sibling(t){if(!this.buffer)return this._tree._parent?this.yield(this._tree.index<0?null:this._tree._parent.nextChild(this._tree.index+t,t,0,4,this.mode)):!1;let{buffer:e}=this.buffer,i=this.stack.length-1;if(t<0){let n=i<0?0:this.stack[i]+4;if(this.index!=n)return this.yieldBuf(e.findChild(n,this.index,-1,0,4))}else{let n=e.buffer[this.index+3];if(n<(i<0?e.buffer.length:e.buffer[this.stack[i]+3]))return this.yieldBuf(n)}return i<0?this.yield(this.buffer.parent.nextChild(this.buffer.index+t,t,0,4,this.mode)):!1}nextSibling(){return this.sibling(1)}prevSibling(){return this.sibling(-1)}atLastNode(t){let e,i,{buffer:n}=this;if(n){if(t>0){if(this.index<n.buffer.buffer.length)return!1}else for(let r=0;r<this.index;r++)if(n.buffer.buffer[r+3]<this.index)return!1;({index:e,parent:i}=n)}else({index:e,_parent:i}=this._tree);for(;i;{index:e,_parent:i}=i)if(e>-1)for(let r=e+t,o=t<0?-1:i._tree.children.length;r!=o;r+=t){let l=i._tree.children[r];if(this.mode&Q.IncludeAnonymous||l instanceof ie||!l.type.isAnonymous||Ar(l))return!1}return!0}move(t,e){if(e&&this.enterChild(t,0,4))return!0;for(;;){if(this.sibling(t))return!0;if(this.atLastNode(t)||!this.parent())return!1}}next(t=!0){return this.move(1,t)}prev(t=!0){return this.move(-1,t)}moveTo(t,e=0){for(;(this.from==this.to||(e<1?this.from>=t:this.from>t)||(e>-1?this.to<=t:this.to<t))&&this.parent(););for(;this.enterChild(1,t,e););return this}get node(){if(!this.buffer)return this._tree;let t=this.bufferNode,e=null,i=0;if(t&&t.context==this.buffer)t:for(let n=this.index,r=this.stack.length;r>=0;){for(let o=t;o;o=o._parent)if(o.index==n){if(n==this.index)return o;e=o,i=r+1;break t}n=this.stack[--r]}for(let n=i;n<this.stack.length;n++)e=new Di(this.buffer,e,this.stack[n]);return this.bufferNode=new Di(this.buffer,e,this.index)}get tree(){return this.buffer?null:this._tree._tree}iterate(t,e){for(let i=0;;){let n=!1;if(this.type.isAnonymous||t(this)!==!1){if(this.firstChild()){i++;continue}this.type.isAnonymous||(n=!0)}for(;;){if(n&&e&&e(this),n=this.type.isAnonymous,!i)return;if(this.nextSibling())break;this.parent(),i--,n=!0}}}matchContext(t){if(!this.buffer)return Sr(this.node.parent,t);let{buffer:e}=this.buffer,{types:i}=e.set;for(let n=t.length-1,r=this.stack.length-1;n>=0;r--){if(r<0)return Sr(this._tree,t,n);let o=i[e.buffer[this.stack[r]]];if(!o.isAnonymous){if(t[n]&&t[n]!=o.name)return!1;n--}}return!0}};function Ar(s){return s.children.some(t=>t instanceof ie||!t.type.isAnonymous||Ar(t))}function Ru(s){var t;let{buffer:e,nodeSet:i,maxBufferLength:n=1024,reused:r=[],minRepeatType:o=i.types.length}=s,l=Array.isArray(e)?new vr(e,e.length):e,a=i.types,h=0,c=0;function f(P,T,O,V,N,q){let{id:E,start:M,end:H,size:j}=l,J=c,Kt=h;if(j<0)if(l.next(),j==-1){let Ft=r[E];O.push(Ft),V.push(M-P);return}else if(j==-3){h=E;return}else if(j==-4){c=E;return}else throw new RangeError(`Unrecognized record size: ${j}`);let ni=a[E],zi,le,vo=M-P;if(H-M<=n&&(le=y(l.pos-T,N))){let Ft=new Uint16Array(le.size-le.skip),ut=l.pos-le.size,Rt=Ft.length;for(;l.pos>ut;)Rt=b(le.start,Ft,Rt);zi=new ie(Ft,H-le.start,i),vo=le.start-P}else{let Ft=l.pos-j;l.next();let ut=[],Rt=[],ae=E>=o?E:-1,De=0,Qi=H;for(;l.pos>Ft;)ae>=0&&l.id==ae&&l.size>=0?(l.end<=Qi-n&&(p(ut,Rt,M,De,l.end,Qi,ae,J,Kt),De=ut.length,Qi=l.end),l.next()):q>2500?u(M,Ft,ut,Rt):f(M,Ft,ut,Rt,ae,q+1);if(ae>=0&&De>0&&De<ut.length&&p(ut,Rt,M,De,M,Qi,ae,J,Kt),ut.reverse(),Rt.reverse(),ae>-1&&De>0){let So=d(ni,Kt);zi=Pr(ni,ut,Rt,0,ut.length,0,H-M,So,So)}else zi=g(ni,ut,Rt,H-M,J-H,Kt)}O.push(zi),V.push(vo)}function u(P,T,O,V){let N=[],q=0,E=-1;for(;l.pos>T;){let{id:M,start:H,end:j,size:J}=l;if(J>4)l.next();else{if(E>-1&&H<E)break;E<0&&(E=j-n),N.push(M,H,j),q++,l.next()}}if(q){let M=new Uint16Array(q*4),H=N[N.length-2];for(let j=N.length-3,J=0;j>=0;j-=3)M[J++]=N[j],M[J++]=N[j+1]-H,M[J++]=N[j+2]-H,M[J++]=J;O.push(new ie(M,N[2]-H,i)),V.push(H-P)}}function d(P,T){return(O,V,N)=>{let q=0,E=O.length-1,M,H;if(E>=0&&(M=O[E])instanceof U){if(!E&&M.type==P&&M.length==N)return M;(H=M.prop(D.lookAhead))&&(q=V[E]+M.length+H)}return g(P,O,V,N,q,T)}}function p(P,T,O,V,N,q,E,M,H){let j=[],J=[];for(;P.length>V;)j.push(P.pop()),J.push(T.pop()+O-N);P.push(g(i.types[E],j,J,q-N,M-q,H)),T.push(N-O)}function g(P,T,O,V,N,q,E){if(q){let M=[D.contextHash,q];E=E?[M].concat(E):[M]}if(N>25){let M=[D.lookAhead,N];E=E?[M].concat(E):[M]}return new U(P,T,O,V,E)}function y(P,T){let O=l.fork(),V=0,N=0,q=0,E=O.end-n,M={size:0,start:0,skip:0};t:for(let H=O.pos-P;O.pos>H;){let j=O.size;if(O.id==T&&j>=0){M.size=V,M.start=N,M.skip=q,q+=4,V+=4,O.next();continue}let J=O.pos-j;if(j<0||J<H||O.start<E)break;let Kt=O.id>=o?4:0,ni=O.start;for(O.next();O.pos>J;){if(O.size<0)if(O.size==-3||O.size==-4)Kt+=4;else break t;else O.id>=o&&(Kt+=4);O.next()}N=ni,V+=j,q+=Kt}return(T<0||V==P)&&(M.size=V,M.start=N,M.skip=q),M.size>4?M:void 0}function b(P,T,O){let{id:V,start:N,end:q,size:E}=l;if(l.next(),E>=0&&V<o){let M=O;if(E>4){let H=l.pos-(E-4);for(;l.pos>H;)O=b(P,T,O)}T[--O]=M,T[--O]=q-P,T[--O]=N-P,T[--O]=V}else E==-3?h=V:E==-4&&(c=V);return O}let v=[],C=[];for(;l.pos>0;)f(s.start||0,s.bufferStart||0,v,C,-1,0);let B=(t=s.length)!==null&&t!==void 0?t:v.length?C[0]+v[0].length:0;return new U(a[s.topID],v.reverse(),C.reverse(),B)}var Xa=new WeakMap;function Es(s,t){if(!s.isAnonymous||t instanceof ie||t.type!=s)return 1;let e=Xa.get(t);if(e==null){e=1;for(let i of t.children){if(i.type!=s||!(i instanceof U)){e=1;break}e+=Es(s,i)}Xa.set(t,e)}return e}function Pr(s,t,e,i,n,r,o,l,a){let h=0;for(let p=i;p<n;p++)h+=Es(s,t[p]);let c=Math.ceil(h*1.5/8),f=[],u=[];function d(p,g,y,b,v){for(let C=y;C<b;){let B=C,P=g[C],T=Es(s,p[C]);for(C++;C<b;C++){let O=Es(s,p[C]);if(T+O>=c)break;T+=O}if(C==B+1){if(T>c){let O=p[B];d(O.children,O.positions,0,O.children.length,g[B]+v);continue}f.push(p[B])}else{let O=g[C-1]+p[C-1].length-P;f.push(Pr(s,p,g,B,C,P,O,null,a))}u.push(P+v-r)}}return d(t,e,i,n,0),(l||a)(f,u,o)}var ve=class s{constructor(t,e,i,n,r=!1,o=!1){this.from=t,this.to=e,this.tree=i,this.offset=n,this.open=(r?1:0)|(o?2:0)}get openStart(){return(this.open&1)>0}get openEnd(){return(this.open&2)>0}static addTree(t,e=[],i=!1){let n=[new s(0,t.length,t,0,!1,i)];for(let r of e)r.to>t.length&&n.push(r);return n}static applyChanges(t,e,i=128){if(!e.length)return t;let n=[],r=1,o=t.length?t[0]:null;for(let l=0,a=0,h=0;;l++){let c=l<e.length?e[l]:null,f=c?c.fromA:1e9;if(f-a>=i)for(;o&&o.from<f;){let u=o;if(a>=u.from||f<=u.to||h){let d=Math.max(u.from,a)-h,p=Math.min(u.to,f)-h;u=d>=p?null:new s(d,p,u.tree,u.offset+h,l>0,!!c)}if(u&&n.push(u),o.to>f)break;o=r<t.length?t[r++]:null}if(!c)break;a=c.toA,h=c.toA-c.toB}return n}},Ue=class{startParse(t,e,i){return typeof t=="string"&&(t=new Tr(t)),i=i?i.length?i.map(n=>new Ai(n.from,n.to)):[new Ai(0,0)]:[new Ai(0,t.length)],this.createParse(t,e||[],i)}parse(t,e,i){let n=this.startParse(t,e,i);for(;;){let r=n.advance();if(r)return r}}},Tr=class{constructor(t){this.string=t}get length(){return this.string.length}chunk(t){return this.string.slice(t)}get lineChunks(){return!1}read(t,e){return this.string.slice(t,e)}};var Dg=new D({perNode:!0});var Eu=0,Tt=class s{constructor(t,e,i,n){this.name=t,this.set=e,this.base=i,this.modified=n,this.id=Eu++}toString(){let{name:t}=this;for(let e of this.modified)e.name&&(t=`${e.name}(${t})`);return t}static define(t,e){let i=typeof t=="string"?t:"?";if(t instanceof s&&(e=t),e?.base)throw new Error("Can not derive from a modified tag");let n=new s(i,[],null,[]);if(n.set.push(n),e)for(let r of e.set)n.set.push(r);return n}static defineModifier(t){let e=new _s(t);return i=>i.modified.indexOf(e)>-1?i:_s.get(i.base||i,i.modified.concat(e).sort((n,r)=>n.id-r.id))}},Bu=0,_s=class s{constructor(t){this.name=t,this.instances=[],this.id=Bu++}static get(t,e){if(!e.length)return t;let i=e[0].instances.find(l=>l.base==t&&Lu(e,l.modified));if(i)return i;let n=[],r=new Tt(t.name,n,t,e);for(let l of e)l.instances.push(r);let o=Iu(e);for(let l of t.set)if(!l.modified.length)for(let a of o)n.push(s.get(l,a));return r}};function Lu(s,t){return s.length==t.length&&s.every((e,i)=>e==t[i])}function Iu(s){let t=[[]];for(let e=0;e<s.length;e++)for(let i=0,n=t.length;i<n;i++)t.push(t[i].concat(s[e]));return t.sort((e,i)=>i.length-e.length)}function Ge(s){let t=Object.create(null);for(let e in s){let i=s[e];Array.isArray(i)||(i=[i]);for(let n of e.split(" "))if(n){let r=[],o=2,l=n;for(let f=0;;){if(l=="..."&&f>0&&f+3==n.length){o=1;break}let u=/^"(?:[^"\\]|\\.)*?"|[^\/!]+/.exec(l);if(!u)throw new RangeError("Invalid path: "+n);if(r.push(u[0]=="*"?"":u[0][0]=='"'?JSON.parse(u[0]):u[0]),f+=u[0].length,f==n.length)break;let d=n[f++];if(f==n.length&&d=="!"){o=0;break}if(d!="/")throw new RangeError("Invalid path: "+n);l=n.slice(f)}let a=r.length-1,h=r[a];if(!h)throw new RangeError("Invalid path: "+n);let c=new Oe(i,o,a>0?r.slice(0,a):null);t[h]=c.sort(t[h])}}return Ya.add(t)}var Ya=new D({combine(s,t){let e,i,n;for(;s||t;){if(!s||t&&s.depth>=t.depth?(n=t,t=t.next):(n=s,s=s.next),e&&e.mode==n.mode&&!n.context&&!e.context)continue;let r=new Oe(n.tags,n.mode,n.context);e?e.next=r:i=r,e=r}return i}}),Oe=class{constructor(t,e,i,n){this.tags=t,this.mode=e,this.context=i,this.next=n}get opaque(){return this.mode==0}get inherit(){return this.mode==1}sort(t){return!t||t.depth<this.depth?(this.next=t,this):(t.next=this.sort(t.next),t)}get depth(){return this.context?this.context.length:0}};Oe.empty=new Oe([],2,null);function Er(s,t){let e=Object.create(null);for(let r of s)if(!Array.isArray(r.tag))e[r.tag.id]=r.class;else for(let o of r.tag)e[o.id]=r.class;let{scope:i,all:n=null}=t||{};return{style:r=>{let o=n;for(let l of r)for(let a of l.set){let h=e[a.id];if(h){o=o?o+" "+h:h;break}}return o},scope:i}}function Nu(s,t){let e=null;for(let i of s){let n=i.style(t);n&&(e=e?e+" "+n:n)}return e}function Ja(s,t,e,i=0,n=s.length){let r=new Dr(i,Array.isArray(t)?t:[t],e);r.highlightRange(s.cursor(),i,n,"",r.highlighters),r.flush(n)}var Dr=class{constructor(t,e,i){this.at=t,this.highlighters=e,this.span=i,this.class=""}startSpan(t,e){e!=this.class&&(this.flush(t),t>this.at&&(this.at=t),this.class=e)}flush(t){t>this.at&&this.class&&this.span(this.at,t,this.class)}highlightRange(t,e,i,n,r){let{type:o,from:l,to:a}=t;if(l>=i||a<=e)return;o.isTop&&(r=this.highlighters.filter(d=>!d.scope||d.scope(o)));let h=n,c=_u(t)||Oe.empty,f=Nu(r,c.tags);if(f&&(h&&(h+=" "),h+=f,c.mode==1&&(n+=(n?" ":"")+f)),this.startSpan(Math.max(e,l),h),c.opaque)return;let u=t.tree&&t.tree.prop(D.mounted);if(u&&u.overlay){let d=t.node.enter(u.overlay[0].from+l,1),p=this.highlighters.filter(y=>!y.scope||y.scope(u.tree.type)),g=t.firstChild();for(let y=0,b=l;;y++){let v=y<u.overlay.length?u.overlay[y]:null,C=v?v.from+l:a,B=Math.max(e,b),P=Math.min(i,C);if(B<P&&g)for(;t.from<P&&(this.highlightRange(t,B,P,n,r),this.startSpan(Math.min(P,t.to),h),!(t.to>=C||!t.nextSibling())););if(!v||C>i)break;b=v.to+l,b>e&&(this.highlightRange(d.cursor(),Math.max(e,v.from+l),Math.min(i,b),"",p),this.startSpan(Math.min(i,b),h))}g&&t.parent()}else if(t.firstChild()){u&&(n="");do if(!(t.to<=e)){if(t.from>=i)break;this.highlightRange(t,e,i,n,r),this.startSpan(Math.min(i,t.to),h)}while(t.nextSibling());t.parent()}}};function _u(s){let t=s.type.prop(Ya);for(;t&&t.context&&!s.matchContext(t.context);)t=t.next;return t||null}var k=Tt.define,Ls=k(),se=k(),Ga=k(se),Ka=k(se),ne=k(),Is=k(ne),Mr=k(ne),Ht=k(),Se=k(Ht),_t=k(),Vt=k(),Rr=k(),Ei=k(Rr),Ns=k(),m={comment:Ls,lineComment:k(Ls),blockComment:k(Ls),docComment:k(Ls),name:se,variableName:k(se),typeName:Ga,tagName:k(Ga),propertyName:Ka,attributeName:k(Ka),className:k(se),labelName:k(se),namespace:k(se),macroName:k(se),literal:ne,string:Is,docString:k(Is),character:k(Is),attributeValue:k(Is),number:Mr,integer:k(Mr),float:k(Mr),bool:k(ne),regexp:k(ne),escape:k(ne),color:k(ne),url:k(ne),keyword:_t,self:k(_t),null:k(_t),atom:k(_t),unit:k(_t),modifier:k(_t),operatorKeyword:k(_t),controlKeyword:k(_t),definitionKeyword:k(_t),moduleKeyword:k(_t),operator:Vt,derefOperator:k(Vt),arithmeticOperator:k(Vt),logicOperator:k(Vt),bitwiseOperator:k(Vt),compareOperator:k(Vt),updateOperator:k(Vt),definitionOperator:k(Vt),typeOperator:k(Vt),controlOperator:k(Vt),punctuation:Rr,separator:k(Rr),bracket:Ei,angleBracket:k(Ei),squareBracket:k(Ei),paren:k(Ei),brace:k(Ei),content:Ht,heading:Se,heading1:k(Se),heading2:k(Se),heading3:k(Se),heading4:k(Se),heading5:k(Se),heading6:k(Se),contentSeparator:k(Ht),list:k(Ht),quote:k(Ht),emphasis:k(Ht),strong:k(Ht),link:k(Ht),monospace:k(Ht),strikethrough:k(Ht),inserted:k(),deleted:k(),changed:k(),invalid:k(),meta:Ns,documentMeta:k(Ns),annotation:k(Ns),processingInstruction:k(Ns),definition:Tt.defineModifier("definition"),constant:Tt.defineModifier("constant"),function:Tt.defineModifier("function"),standard:Tt.defineModifier("standard"),local:Tt.defineModifier("local"),special:Tt.defineModifier("special")};for(let s in m){let t=m[s];t instanceof Tt&&(t.name=s)}var Bg=Er([{tag:m.link,class:"tok-link"},{tag:m.heading,class:"tok-heading"},{tag:m.emphasis,class:"tok-emphasis"},{tag:m.strong,class:"tok-strong"},{tag:m.keyword,class:"tok-keyword"},{tag:m.atom,class:"tok-atom"},{tag:m.bool,class:"tok-bool"},{tag:m.url,class:"tok-url"},{tag:m.labelName,class:"tok-labelName"},{tag:m.inserted,class:"tok-inserted"},{tag:m.deleted,class:"tok-deleted"},{tag:m.literal,class:"tok-literal"},{tag:m.string,class:"tok-string"},{tag:m.number,class:"tok-number"},{tag:[m.regexp,m.escape,m.special(m.string)],class:"tok-string2"},{tag:m.variableName,class:"tok-variableName"},{tag:m.local(m.variableName),class:"tok-variableName tok-local"},{tag:m.definition(m.variableName),class:"tok-variableName tok-definition"},{tag:m.special(m.variableName),class:"tok-variableName2"},{tag:m.definition(m.propertyName),class:"tok-propertyName tok-definition"},{tag:m.typeName,class:"tok-typeName"},{tag:m.namespace,class:"tok-namespace"},{tag:m.className,class:"tok-className"},{tag:m.macroName,class:"tok-macroName"},{tag:m.propertyName,class:"tok-propertyName"},{tag:m.operator,class:"tok-operator"},{tag:m.comment,class:"tok-comment"},{tag:m.meta,class:"tok-meta"},{tag:m.invalid,class:"tok-invalid"},{tag:m.punctuation,class:"tok-punctuation"}]);var Br,Ke=new D;function Vu(s){return S.define({combine:s?t=>t.concat(s):void 0})}var Hu=new D,ct=class{constructor(t,e,i=[],n=""){this.data=t,this.name=n,W.prototype.hasOwnProperty("tree")||Object.defineProperty(W.prototype,"tree",{get(){return At(this)}}),this.parser=e,this.extension=[Ye.of(this),W.languageData.of((r,o,l)=>{let a=Za(r,o,l),h=a.type.prop(Ke);if(!h)return[];let c=r.facet(h),f=a.type.prop(Hu);if(f){let u=a.resolve(o-a.from,l);for(let d of f)if(d.test(u,r)){let p=r.facet(d.facet);return d.type=="replace"?p:p.concat(c)}}return c})].concat(i)}isActiveAt(t,e,i=-1){return Za(t,e,i).type.prop(Ke)==this.data}findRegions(t){let e=t.facet(Ye);if(e?.data==this.data)return[{from:0,to:t.doc.length}];if(!e||!e.allowsNesting)return[];let i=[],n=(r,o)=>{if(r.prop(Ke)==this.data){i.push({from:o,to:o+r.length});return}let l=r.prop(D.mounted);if(l){if(l.tree.prop(Ke)==this.data){if(l.overlay)for(let a of l.overlay)i.push({from:a.from+o,to:a.to+o});else i.push({from:o,to:o+r.length});return}else if(l.overlay){let a=i.length;if(n(l.tree,l.overlay[0].from+o),i.length>a)return}}for(let a=0;a<r.children.length;a++){let h=r.children[a];h instanceof U&&n(h,r.positions[a]+o)}};return n(At(t),0),i}get allowsNesting(){return!0}};ct.setState=F.define();function Za(s,t,e){let i=s.facet(Ye),n=At(s).topNode;if(!i||i.allowsNesting)for(let r=n;r;r=r.enter(t,e,Q.ExcludeBuffers|Q.EnterBracketed))r.type.isTop&&(n=r);return n}var Li=class s extends ct{constructor(t,e,i){super(t,e,[],i),this.parser=e}static define(t){let e=Vu(t.languageData);return new s(e,t.parser.configure({props:[Ke.add(i=>i.isTop?e:void 0)]}),t.name)}configure(t,e){return new s(this.data,this.parser.configure(t),e||this.name)}get allowsNesting(){return this.parser.hasWrappers()}};function At(s){let t=s.field(ct.state,!1);return t?t.tree:U.empty}var _r=class{constructor(t){this.doc=t,this.cursorPos=0,this.string="",this.cursor=t.iter()}get length(){return this.doc.length}syncTo(t){return this.string=this.cursor.next(t-this.cursorPos).value,this.cursorPos=t+this.string.length,this.cursorPos-this.string.length}chunk(t){return this.syncTo(t),this.string}get lineChunks(){return!0}read(t,e){let i=this.cursorPos-this.string.length;return t<i||e>=this.cursorPos?this.doc.sliceString(t,e):this.string.slice(t-i,e-i)}},Bi=null,Vr=class s{constructor(t,e,i=[],n,r,o,l,a){this.parser=t,this.state=e,this.fragments=i,this.tree=n,this.treeLen=r,this.viewport=o,this.skipped=l,this.scheduleOn=a,this.parse=null,this.tempSkipped=[]}static create(t,e,i){return new s(t,e,[],U.empty,0,i,[],null)}startParse(){return this.parser.startParse(new _r(this.state.doc),this.fragments)}work(t,e){return e!=null&&e>=this.state.doc.length&&(e=void 0),this.tree!=U.empty&&this.isDone(e??this.state.doc.length)?(this.takeTree(),!0):this.withContext(()=>{var i;if(typeof t=="number"){let n=Date.now()+t;t=()=>Date.now()>n}for(this.parse||(this.parse=this.startParse()),e!=null&&(this.parse.stoppedAt==null||this.parse.stoppedAt>e)&&e<this.state.doc.length&&this.parse.stopAt(e);;){let n=this.parse.advance();if(n)if(this.fragments=this.withoutTempSkipped(ve.addTree(n,this.fragments,this.parse.stoppedAt!=null)),this.treeLen=(i=this.parse.stoppedAt)!==null&&i!==void 0?i:this.state.doc.length,this.tree=n,this.parse=null,this.treeLen<(e??this.state.doc.length))this.parse=this.startParse();else return!0;if(t())return!1}})}takeTree(){let t,e;this.parse&&(t=this.parse.parsedPos)>=this.treeLen&&((this.parse.stoppedAt==null||this.parse.stoppedAt>t)&&this.parse.stopAt(t),this.withContext(()=>{for(;!(e=this.parse.advance()););}),this.treeLen=t,this.tree=e,this.fragments=this.withoutTempSkipped(ve.addTree(this.tree,this.fragments,!0)),this.parse=null)}withContext(t){let e=Bi;Bi=this;try{return t()}finally{Bi=e}}withoutTempSkipped(t){for(let e;e=this.tempSkipped.pop();)t=th(t,e.from,e.to);return t}changes(t,e){let{fragments:i,tree:n,treeLen:r,viewport:o,skipped:l}=this;if(this.takeTree(),!t.empty){let a=[];if(t.iterChangedRanges((h,c,f,u)=>a.push({fromA:h,toA:c,fromB:f,toB:u})),i=ve.applyChanges(i,a),n=U.empty,r=0,o={from:t.mapPos(o.from,-1),to:t.mapPos(o.to,1)},this.skipped.length){l=[];for(let h of this.skipped){let c=t.mapPos(h.from,1),f=t.mapPos(h.to,-1);c<f&&l.push({from:c,to:f})}}}return new s(this.parser,e,i,n,r,o,l,this.scheduleOn)}updateViewport(t){if(this.viewport.from==t.from&&this.viewport.to==t.to)return!1;this.viewport=t;let e=this.skipped.length;for(let i=0;i<this.skipped.length;i++){let{from:n,to:r}=this.skipped[i];n<t.to&&r>t.from&&(this.fragments=th(this.fragments,n,r),this.skipped.splice(i--,1))}return this.skipped.length>=e?!1:(this.reset(),!0)}reset(){this.parse&&(this.takeTree(),this.parse=null)}skipUntilInView(t,e){this.skipped.push({from:t,to:e})}static getSkippingParser(t){return new class extends Ue{createParse(e,i,n){let r=n[0].from,o=n[n.length-1].to;return{parsedPos:r,advance(){let a=Bi;if(a){for(let h of n)a.tempSkipped.push(h);t&&(a.scheduleOn=a.scheduleOn?Promise.all([a.scheduleOn,t]):t)}return this.parsedPos=o,new U(tt.none,[],[],o-r)},stoppedAt:null,stopAt(){}}}}}isDone(t){t=Math.min(t,this.state.doc.length);let e=this.fragments;return this.treeLen>=t&&e.length&&e[0].from==0&&e[0].to>=t}static get(){return Bi}};function th(s,t,e){return ve.applyChanges(s,[{fromA:t,toA:e,fromB:t,toB:e}])}var Ii=class s{constructor(t){this.context=t,this.tree=t.tree}apply(t){if(!t.docChanged&&this.tree==this.context.tree)return this;let e=this.context.changes(t.changes,t.state),i=this.context.treeLen==t.startState.doc.length?void 0:Math.max(t.changes.mapPos(this.context.treeLen),e.viewport.to);return e.work(20,i)||e.takeTree(),new s(e)}static init(t){let e=Math.min(3e3,t.doc.length),i=Vr.create(t.facet(Ye).parser,t,{from:0,to:e});return i.work(20,e)||i.takeTree(),new s(i)}};ct.state=xt.define({create:Ii.init,update(s,t){for(let e of t.effects)if(e.is(ct.setState))return e.value;return t.startState.facet(Ye)!=t.state.facet(Ye)?Ii.init(t.state):s.apply(t)}});var nh=s=>{let t=setTimeout(()=>s(),500);return()=>clearTimeout(t)};typeof requestIdleCallback<"u"&&(nh=s=>{let t=-1,e=setTimeout(()=>{t=requestIdleCallback(s,{timeout:400})},100);return()=>t<0?clearTimeout(e):cancelIdleCallback(t)});var Lr=typeof navigator<"u"&&(!((Br=navigator.scheduling)===null||Br===void 0)&&Br.isInputPending)?()=>navigator.scheduling.isInputPending():null,Fu=Xt.fromClass(class{constructor(t){this.view=t,this.working=null,this.workScheduled=0,this.chunkEnd=-1,this.chunkBudget=-1,this.work=this.work.bind(this),this.scheduleWork()}update(t){let e=this.view.state.field(ct.state).context;(e.updateViewport(t.view.viewport)||this.view.viewport.to>e.treeLen)&&this.scheduleWork(),(t.docChanged||t.selectionSet)&&(this.view.hasFocus&&(this.chunkBudget+=50),this.scheduleWork()),this.checkAsyncSchedule(e)}scheduleWork(){if(this.working)return;let{state:t}=this.view,e=t.field(ct.state);(e.tree!=e.context.tree||!e.context.isDone(t.doc.length))&&(this.working=nh(this.work))}work(t){this.working=null;let e=Date.now();if(this.chunkEnd<e&&(this.chunkEnd<0||this.view.hasFocus)&&(this.chunkEnd=e+3e4,this.chunkBudget=3e3),this.chunkBudget<=0)return;let{state:i,viewport:{to:n}}=this.view,r=i.field(ct.state);if(r.tree==r.context.tree&&r.context.isDone(n+1e5))return;let o=Date.now()+Math.min(this.chunkBudget,100,t&&!Lr?Math.max(25,t.timeRemaining()-5):1e9),l=r.context.treeLen<n&&i.doc.length>n+1e3,a=r.context.work(()=>Lr&&Lr()||Date.now()>o,n+(l?0:1e5));this.chunkBudget-=Date.now()-e,(a||this.chunkBudget<=0)&&(r.context.takeTree(),this.view.dispatch({effects:ct.setState.of(new Ii(r.context))})),this.chunkBudget>0&&!(a&&!l)&&this.scheduleWork(),this.checkAsyncSchedule(r.context)}checkAsyncSchedule(t){t.scheduleOn&&(this.workScheduled++,t.scheduleOn.then(()=>this.scheduleWork()).catch(e=>St(this.view.state,e)).then(()=>this.workScheduled--),t.scheduleOn=null)}destroy(){this.working&&this.working()}isWorking(){return!!(this.working||this.workScheduled>0)}},{eventHandlers:{focus(){this.scheduleWork()}}}),Ye=S.define({combine(s){return s.length?s[0]:null},enables:s=>[ct.state,Fu,A.contentAttributes.compute([s],t=>{let e=t.facet(s);return e&&e.name?{"data-language":e.name}:{}})]}),Vs=class{constructor(t,e=[]){this.language=t,this.support=e,this.extension=[t,e]}};var Wu=S.define(),Hs=S.define({combine:s=>{if(!s.length)return"  ";let t=s[0];if(!t||/\S/.test(t)||Array.from(t).some(e=>e!=t[0]))throw new Error("Invalid indent unit: "+JSON.stringify(s[0]));return t}});function Ni(s){let t=s.facet(Hs);return t.charCodeAt(0)==9?s.tabSize*t.length:t.length}function _i(s,t){let e="",i=s.tabSize,n=s.facet(Hs)[0];if(n=="	"){for(;t>=i;)e+="	",t-=i;n=" "}for(let r=0;r<t;r++)e+=n;return e}function Qr(s,t){s instanceof W&&(s=new Ce(s));for(let i of s.state.facet(Wu)){let n=i(s,t);if(n!==void 0)return n}let e=At(s.state);return e.length>=t?$u(s,e,t):null}var Ce=class{constructor(t,e={}){this.state=t,this.options=e,this.unit=Ni(t)}lineAt(t,e=1){let i=this.state.doc.lineAt(t),{simulateBreak:n,simulateDoubleBreak:r}=this.options;return n!=null&&n>=i.from&&n<=i.to?r&&n==t?{text:"",from:t}:(e<0?n<t:n<=t)?{text:i.text.slice(n-i.from),from:n}:{text:i.text.slice(0,n-i.from),from:i.from}:i}textAfterPos(t,e=1){if(this.options.simulateDoubleBreak&&t==this.options.simulateBreak)return"";let{text:i,from:n}=this.lineAt(t,e);return i.slice(t-n,Math.min(i.length,t+100-n))}column(t,e=1){let{text:i,from:n}=this.lineAt(t,e),r=this.countColumn(i,t-n),o=this.options.overrideIndentation?this.options.overrideIndentation(n):-1;return o>-1&&(r+=o-this.countColumn(i,i.search(/\S|$/))),r}countColumn(t,e=t.length){return pe(t,this.state.tabSize,e)}lineIndent(t,e=1){let{text:i,from:n}=this.lineAt(t,e),r=this.options.overrideIndentation;if(r){let o=r(n);if(o>-1)return o}return this.countColumn(i,i.search(/\S|$/))}get simulatedBreak(){return this.options.simulateBreak||null}},jr=new D;function $u(s,t,e){let i=t.resolveStack(e),n=t.resolveInner(e,-1).resolve(e,0).enterUnfinishedNodesBefore(e);if(n!=i.node){let r=[];for(let o=n;o&&!(o.from<i.node.from||o.to>i.node.to||o.from==i.node.from&&o.type==i.node.type);o=o.parent)r.push(o);for(let o=r.length-1;o>=0;o--)i={node:r[o],next:i}}return rh(i,s,e)}function rh(s,t,e){for(let i=s;i;i=i.next){let n=Qu(i.node);if(n)return n(Hr.create(t,e,i))}return 0}function zu(s){return s.pos==s.options.simulateBreak&&s.options.simulateDoubleBreak}function Qu(s){let t=s.type.prop(jr);if(t)return t;let e=s.firstChild,i;if(e&&(i=e.type.prop(D.closedBy))){let n=s.lastChild,r=n&&i.indexOf(n.name)>-1;return o=>oh(o,!0,1,void 0,r&&!zu(o)?n.from:void 0)}return s.parent==null?ju:null}function ju(){return 0}var Hr=class s extends Ce{constructor(t,e,i){super(t.state,t.options),this.base=t,this.pos=e,this.context=i}get node(){return this.context.node}static create(t,e,i){return new s(t,e,i)}get textAfter(){return this.textAfterPos(this.pos)}get baseIndent(){return this.baseIndentFor(this.node)}baseIndentFor(t){let e=this.state.doc.lineAt(t.from);for(;;){let i=t.resolve(e.from);for(;i.parent&&i.parent.from==i.from;)i=i.parent;if(Xu(i,t))break;e=this.state.doc.lineAt(i.from)}return this.lineIndent(e.from)}continue(){return rh(this.context.next,this.base,this.pos)}};function Xu(s,t){for(let e=t;e;e=e.parent)if(s==e)return!0;return!1}function qu(s){let t=s.node,e=t.childAfter(t.from),i=t.lastChild;if(!e)return null;let n=s.options.simulateBreak,r=s.state.doc.lineAt(e.from),o=n==null||n<=r.from?r.to:Math.min(r.to,n);for(let l=e.to;;){let a=t.childAfter(l);if(!a||a==i)return null;if(!a.type.isSkipped){if(a.from>=o)return null;let h=/^ */.exec(r.text.slice(e.to-r.from))[0].length;return{from:e.from,to:e.to+h}}l=a.to}}function Xr({closing:s,align:t=!0,units:e=1}){return i=>oh(i,t,e,s)}function oh(s,t,e,i,n){let r=s.textAfter,o=r.match(/^\s*/)[0].length,l=i&&r.slice(o,o+i.length)==i||n==s.pos+o,a=t?qu(s):null;return a?l?s.column(a.from):s.column(a.to):s.baseIndent+(l?0:s.unit*e)}var lh=new D;function ah(s){let t=s.firstChild,e=s.lastChild;return t&&t.to<e.from?{from:t.to,to:e.type.isError?s.to:e.from}:null}var Je=class s{constructor(t,e){this.specs=t;let i;function n(l){let a=pt.newName();return(i||(i=Object.create(null)))["."+a]=l,a}let r=typeof e.all=="string"?e.all:e.all?n(e.all):void 0,o=e.scope;this.scope=o instanceof ct?l=>l.prop(Ke)==o.data:o?l=>l==o:void 0,this.style=Er(t.map(l=>({tag:l.tag,class:l.class||n(Object.assign({},l,{tag:null}))})),{all:r}).style,this.module=i?new pt(i):null,this.themeType=e.themeType}static define(t,e){return new s(t,e||{})}},Fr=S.define(),hh=S.define({combine(s){return s.length?[s[0]]:null}});function Ir(s){let t=s.facet(Fr);return t.length?t:s.facet(hh)}function ch(s,t){let e=[Uu],i;return s instanceof Je&&(s.module&&e.push(A.styleModule.of(s.module)),i=s.themeType),t?.fallback?e.push(hh.of(s)):i?e.push(Fr.computeN([A.darkTheme],n=>n.facet(A.darkTheme)==(i=="dark")?[s]:[])):e.push(Fr.of(s)),e}var Wr=class{constructor(t){this.markCache=Object.create(null),this.tree=At(t.state),this.decorations=this.buildDeco(t,Ir(t.state)),this.decoratedTo=t.viewport.to}update(t){let e=At(t.state),i=Ir(t.state),n=i!=Ir(t.startState),{viewport:r}=t.view,o=t.changes.mapPos(this.decoratedTo,1);e.length<r.to&&!n&&e.type==this.tree.type&&o>=r.to?(this.decorations=this.decorations.map(t.changes),this.decoratedTo=o):(e!=this.tree||t.viewportChanged||n)&&(this.tree=e,this.decorations=this.buildDeco(t.view,i),this.decoratedTo=r.to)}buildDeco(t,e){if(!e||!this.tree.length)return $.none;let i=new de;for(let{from:n,to:r}of t.visibleRanges)Ja(this.tree,e,(o,l,a)=>{i.add(o,l,this.markCache[a]||(this.markCache[a]=$.mark({class:a})))},n,r);return i.finish()}},Uu=ci.high(Xt.fromClass(Wr,{decorations:s=>s.decorations})),$g=Je.define([{tag:m.meta,color:"#404740"},{tag:m.link,textDecoration:"underline"},{tag:m.heading,textDecoration:"underline",fontWeight:"bold"},{tag:m.emphasis,fontStyle:"italic"},{tag:m.strong,fontWeight:"bold"},{tag:m.strikethrough,textDecoration:"line-through"},{tag:m.keyword,color:"#708"},{tag:[m.atom,m.bool,m.url,m.contentSeparator,m.labelName],color:"#219"},{tag:[m.literal,m.inserted],color:"#164"},{tag:[m.string,m.deleted],color:"#a11"},{tag:[m.regexp,m.escape,m.special(m.string)],color:"#e40"},{tag:m.definition(m.variableName),color:"#00f"},{tag:m.local(m.variableName),color:"#30a"},{tag:[m.typeName,m.namespace],color:"#085"},{tag:m.className,color:"#167"},{tag:[m.special(m.variableName),m.macroName],color:"#256"},{tag:m.definition(m.propertyName),color:"#00c"},{tag:m.comment,color:"#940"},{tag:m.invalid,color:"#f00"}]);var Gu=1e4,Ku="()[]{}";var Yu=new D;function $r(s,t,e){let i=s.prop(t<0?D.openedBy:D.closedBy);if(i)return i;if(s.name.length==1){let n=e.indexOf(s.name);if(n>-1&&n%2==(t<0?1:0))return[e[n+t]]}return null}function zr(s){let t=s.type.prop(Yu);return t?t(s.node):s}function Te(s,t,e,i={}){let n=i.maxScanDistance||Gu,r=i.brackets||Ku,o=At(s),l=o.resolveInner(t,e);for(let a=l;a;a=a.parent){let h=$r(a.type,e,r);if(h&&a.from<a.to){let c=zr(a);if(c&&(e>0?t>=c.from&&t<c.to:t>c.from&&t<=c.to))return Ju(s,t,e,a,c,h,r)}}return Zu(s,t,e,o,l.type,n,r)}function Ju(s,t,e,i,n,r,o){let l=i.parent,a={from:n.from,to:n.to},h=0,c=l?.cursor();if(c&&(e<0?c.childBefore(i.from):c.childAfter(i.to)))do if(e<0?c.to<=i.from:c.from>=i.to){if(h==0&&r.indexOf(c.type.name)>-1&&c.from<c.to){let f=zr(c);return{start:a,end:f?{from:f.from,to:f.to}:void 0,matched:!0}}else if($r(c.type,e,o))h++;else if($r(c.type,-e,o)){if(h==0){let f=zr(c);return{start:a,end:f&&f.from<f.to?{from:f.from,to:f.to}:void 0,matched:!1}}h--}}while(e<0?c.prevSibling():c.nextSibling());return{start:a,matched:!1}}function Zu(s,t,e,i,n,r,o){let l=e<0?s.sliceDoc(t-1,t):s.sliceDoc(t,t+1),a=o.indexOf(l);if(a<0||a%2==0!=e>0)return null;let h={from:e<0?t-1:t,to:e>0?t+1:t},c=s.doc.iterRange(t,e>0?s.doc.length:0),f=0;for(let u=0;!c.next().done&&u<=r;){let d=c.value;e<0&&(u+=d.length);let p=t+u*e;for(let g=e>0?0:d.length-1,y=e>0?d.length:-1;g!=y;g+=e){let b=o.indexOf(d[g]);if(!(b<0||i.resolveInner(p+g,1).type!=n))if(b%2==0==e>0)f++;else{if(f==1)return{start:h,end:{from:p+g,to:p+g+1},matched:b>>1==a>>1};f--}}e>0&&(u+=d.length)}return c.done?{start:h,matched:!1}:null}var td=Object.create(null),eh=[tt.none];var ih=[],sh=Object.create(null),ed=Object.create(null);for(let[s,t]of[["variable","variableName"],["variable-2","variableName.special"],["string-2","string.special"],["def","variableName.definition"],["tag","tagName"],["attribute","attributeName"],["type","typeName"],["builtin","variableName.standard"],["qualifier","modifier"],["error","invalid"],["header","heading"],["property","propertyName"]])ed[s]=id(td,t);function Nr(s,t){ih.indexOf(s)>-1||(ih.push(s),console.warn(t))}function id(s,t){let e=[];for(let l of t.split(" ")){let a=[];for(let h of l.split(".")){let c=s[h]||m[h];c?typeof c=="function"?a.length?a=a.map(c):Nr(h,`Modifier ${h} used at start of tag`):a.length?Nr(h,`Tag ${h} used as modifier`):a=Array.isArray(c)?c:[c]:Nr(h,`Unknown highlighting tag ${h}`)}for(let h of a)e.push(h)}if(!e.length)return 0;let i=t.replace(/ /g,"_"),n=i+" "+e.map(l=>l.id),r=sh[n];if(r)return r.id;let o=sh[n]=tt.define({id:eh.length,name:i,props:[Ge({[i]:e})]});return eh.push(o),o.id}var zg={rtl:$.mark({class:"cm-iso",inclusive:!0,attributes:{dir:"rtl"},bidiIsolate:X.RTL}),ltr:$.mark({class:"cm-iso",inclusive:!0,attributes:{dir:"ltr"},bidiIsolate:X.LTR}),auto:$.mark({class:"cm-iso",inclusive:!0,attributes:{dir:"auto"},bidiIsolate:null})};var sd=s=>{let{state:t}=s,e=t.doc.lineAt(t.selection.main.from),i=Zr(s.state,e.from);return i.line?nd(s):i.block?od(s):!1};function Jr(s,t){return({state:e,dispatch:i})=>{if(e.readOnly)return!1;let n=s(t,e);return n?(i(e.update(n)),!0):!1}}var nd=Jr(hd,0);var rd=Jr(wh,0);var od=Jr((s,t)=>wh(s,t,ad(t)),0);function Zr(s,t){let e=s.languageDataAt("commentTokens",t,1);return e.length?e[0]:{}}var Vi=50;function ld(s,{open:t,close:e},i,n){let r=s.sliceDoc(i-Vi,i),o=s.sliceDoc(n,n+Vi),l=/\s*$/.exec(r)[0].length,a=/^\s*/.exec(o)[0].length,h=r.length-l;if(r.slice(h-t.length,h)==t&&o.slice(a,a+e.length)==e)return{open:{pos:i-l,margin:l&&1},close:{pos:n+a,margin:a&&1}};let c,f;n-i<=2*Vi?c=f=s.sliceDoc(i,n):(c=s.sliceDoc(i,i+Vi),f=s.sliceDoc(n-Vi,n));let u=/^\s*/.exec(c)[0].length,d=/\s*$/.exec(f)[0].length,p=f.length-d-e.length;return c.slice(u,u+t.length)==t&&f.slice(p,p+e.length)==e?{open:{pos:i+u+t.length,margin:/\s/.test(c.charAt(u+t.length))?1:0},close:{pos:n-d-e.length,margin:/\s/.test(f.charAt(p-1))?1:0}}:null}function ad(s){let t=[];for(let e of s.selection.ranges){let i=s.doc.lineAt(e.from),n=e.to<=i.to?i:s.doc.lineAt(e.to);n.from>i.from&&n.from==e.to&&(n=e.to==i.to+1?i:s.doc.lineAt(e.to-1));let r=t.length-1;r>=0&&t[r].to>i.from?t[r].to=n.to:t.push({from:i.from+/^\s*/.exec(i.text)[0].length,to:n.to})}return t}function wh(s,t,e=t.selection.ranges){let i=e.map(r=>Zr(t,r.from).block);if(!i.every(r=>r))return null;let n=e.map((r,o)=>ld(t,i[o],r.from,r.to));if(s!=2&&!n.every(r=>r))return{changes:t.changes(e.map((r,o)=>n[o]?[]:[{from:r.from,insert:i[o].open+" "},{from:r.to,insert:" "+i[o].close}]))};if(s!=1&&n.some(r=>r)){let r=[];for(let o=0,l;o<n.length;o++)if(l=n[o]){let a=i[o],{open:h,close:c}=l;r.push({from:h.pos-a.open.length,to:h.pos+h.margin},{from:c.pos-c.margin,to:c.pos+a.close.length})}return{changes:r}}return null}function hd(s,t,e=t.selection.ranges){let i=[],n=-1;t:for(let{from:r,to:o}of e){let l=i.length,a=1e9,h;for(let c=r;c<=o;){let f=t.doc.lineAt(c);if(h==null&&(h=Zr(t,f.from).line,!h))continue t;if(f.from>n&&(r==o||o>f.from)){n=f.from;let u=/^\s*/.exec(f.text)[0].length,d=u==f.length,p=f.text.slice(u,u+h.length)==h?u:-1;u<f.text.length&&u<a&&(a=u),i.push({line:f,comment:p,token:h,indent:u,empty:d,single:!1})}c=f.to+1}if(a<1e9)for(let c=l;c<i.length;c++)i[c].indent<i[c].line.text.length&&(i[c].indent=a);i.length==l+1&&(i[l].single=!0)}if(s!=2&&i.some(r=>r.comment<0&&(!r.empty||r.single))){let r=[];for(let{line:l,token:a,indent:h,empty:c,single:f}of i)(f||!c)&&r.push({from:l.from+h,insert:a+" "});let o=t.changes(r);return{changes:o,selection:t.selection.map(o,1)}}else if(s!=1&&i.some(r=>r.comment>=0)){let r=[];for(let{line:o,comment:l,token:a}of i)if(l>=0){let h=o.from+l,c=h+a.length;o.text[c-o.from]==" "&&c++,r.push({from:h,to:c})}return{changes:r}}return null}var Ur=dt.define(),cd=dt.define(),fd=S.define(),kh=S.define({combine(s){return fi(s,{minDepth:100,newGroupDelay:500,joinToEvent:(t,e)=>e},{minDepth:Math.max,newGroupDelay:Math.min,joinToEvent:(t,e)=>(i,n)=>t(i,n)||e(i,n)})}}),xh=xt.define({create(){return Ae.empty},update(s,t){let e=t.state.facet(kh),i=t.annotation(Ur);if(i){let a=Pt.fromTransaction(t,i.selection),h=i.side,c=h==0?s.undone:s.done;return a?c=Ws(c,c.length,e.minDepth,a):c=Ch(c,t.startState.selection),new Ae(h==0?i.rest:c,h==0?c:i.rest)}let n=t.annotation(cd);if((n=="full"||n=="before")&&(s=s.isolate()),t.annotation(K.addToHistory)===!1)return t.changes.empty?s:s.addMapping(t.changes.desc);let r=Pt.fromTransaction(t),o=t.annotation(K.time),l=t.annotation(K.userEvent);return r?s=s.addChanges(r,o,l,e,t):t.selection&&(s=s.addSelection(t.startState.selection,o,l,e.newGroupDelay)),(n=="full"||n=="after")&&(s=s.isolate()),s},toJSON(s){return{done:s.done.map(t=>t.toJSON()),undone:s.undone.map(t=>t.toJSON())}},fromJSON(s){return new Ae(s.done.map(Pt.fromJSON),s.undone.map(Pt.fromJSON))}});function vh(s={}){return[xh,kh.of(s),A.domEventHandlers({beforeinput(t,e){let i=t.inputType=="historyUndo"?Sh:t.inputType=="historyRedo"?Gr:null;return i?(t.preventDefault(),i(e)):!1}})]}function $s(s,t){return function({state:e,dispatch:i}){if(!t&&e.readOnly)return!1;let n=e.field(xh,!1);if(!n)return!1;let r=n.pop(s,e,t);return r?(i(r),!0):!1}}var Sh=$s(0,!1),Gr=$s(1,!1),ud=$s(0,!0),dd=$s(1,!0);var Pt=class s{constructor(t,e,i,n,r){this.changes=t,this.effects=e,this.mapped=i,this.startSelection=n,this.selectionsAfter=r}setSelAfter(t){return new s(this.changes,this.effects,this.mapped,this.startSelection,t)}toJSON(){var t,e,i;return{changes:(t=this.changes)===null||t===void 0?void 0:t.toJSON(),mapped:(e=this.mapped)===null||e===void 0?void 0:e.toJSON(),startSelection:(i=this.startSelection)===null||i===void 0?void 0:i.toJSON(),selectionsAfter:this.selectionsAfter.map(n=>n.toJSON())}}static fromJSON(t){return new s(t.changes&&it.fromJSON(t.changes),[],t.mapped&&Wt.fromJSON(t.mapped),t.startSelection&&w.fromJSON(t.startSelection),t.selectionsAfter.map(w.fromJSON))}static fromTransaction(t,e){let i=bt;for(let n of t.startState.facet(fd)){let r=n(t);r.length&&(i=i.concat(r))}return!i.length&&t.changes.empty?null:new s(t.changes.invert(t.startState.doc),i,void 0,e||t.startState.selection,bt)}static selection(t){return new s(void 0,bt,void 0,void 0,t)}};function Ws(s,t,e,i){let n=t+1>e+20?t-e-1:0,r=s.slice(n,t);return r.push(i),r}function pd(s,t){let e=[],i=!1;return s.iterChangedRanges((n,r)=>e.push(n,r)),t.iterChangedRanges((n,r,o,l)=>{for(let a=0;a<e.length;){let h=e[a++],c=e[a++];l>=h&&o<=c&&(i=!0)}}),i}function gd(s,t){return s.ranges.length==t.ranges.length&&s.ranges.filter((e,i)=>e.empty!=t.ranges[i].empty).length===0}function Oh(s,t){return s.length?t.length?s.concat(t):s:t}var bt=[],md=200;function Ch(s,t){if(s.length){let e=s[s.length-1],i=e.selectionsAfter.slice(Math.max(0,e.selectionsAfter.length-md));return i.length&&i[i.length-1].eq(t)?s:(i.push(t),Ws(s,s.length-1,1e9,e.setSelAfter(i)))}else return[Pt.selection([t])]}function yd(s){let t=s[s.length-1],e=s.slice();return e[s.length-1]=t.setSelAfter(t.selectionsAfter.slice(0,t.selectionsAfter.length-1)),e}function qr(s,t){if(!s.length)return s;let e=s.length,i=bt;for(;e;){let n=bd(s[e-1],t,i);if(n.changes&&!n.changes.empty||n.effects.length){let r=s.slice(0,e);return r[e-1]=n,r}else t=n.mapped,e--,i=n.selectionsAfter}return i.length?[Pt.selection(i)]:bt}function bd(s,t,e){let i=Oh(s.selectionsAfter.length?s.selectionsAfter.map(l=>l.map(t)):bt,e);if(!s.changes)return Pt.selection(i);let n=s.changes.map(t),r=t.mapDesc(s.changes,!0),o=s.mapped?s.mapped.composeDesc(r):r;return new Pt(n,F.mapEffects(s.effects,t),o,s.startSelection.map(r),i)}var wd=/^(input\.type|delete)($|\.)/,Ae=class s{constructor(t,e,i=0,n=void 0){this.done=t,this.undone=e,this.prevTime=i,this.prevUserEvent=n}isolate(){return this.prevTime?new s(this.done,this.undone):this}addChanges(t,e,i,n,r){let o=this.done,l=o[o.length-1];return l&&l.changes&&!l.changes.empty&&t.changes&&(!i||wd.test(i))&&(!l.selectionsAfter.length&&e-this.prevTime<n.newGroupDelay&&n.joinToEvent(r,pd(l.changes,t.changes))||i=="input.type.compose")?o=Ws(o,o.length-1,n.minDepth,new Pt(t.changes.compose(l.changes),Oh(F.mapEffects(t.effects,l.changes),l.effects),l.mapped,l.startSelection,bt)):o=Ws(o,o.length,n.minDepth,t),new s(o,bt,e,i)}addSelection(t,e,i,n){let r=this.done.length?this.done[this.done.length-1].selectionsAfter:bt;return r.length>0&&e-this.prevTime<n&&i==this.prevUserEvent&&i&&/^select($|\.)/.test(i)&&gd(r[r.length-1],t)?this:new s(Ch(this.done,t),this.undone,e,i)}addMapping(t){return new s(qr(this.done,t),qr(this.undone,t),this.prevTime,this.prevUserEvent)}pop(t,e,i){let n=t==0?this.done:this.undone;if(n.length==0)return null;let r=n[n.length-1],o=r.selectionsAfter[0]||(r.startSelection?r.startSelection.map(r.changes.invertedDesc,1):e.selection);if(i&&r.selectionsAfter.length)return e.update({selection:r.selectionsAfter[r.selectionsAfter.length-1],annotations:Ur.of({side:t,rest:yd(n),selection:o}),userEvent:t==0?"select.undo":"select.redo",scrollIntoView:!0});if(r.changes){let l=n.length==1?bt:n.slice(0,n.length-1);return r.mapped&&(l=qr(l,r.mapped)),e.update({changes:r.changes,selection:r.startSelection,effects:r.effects,annotations:Ur.of({side:t,rest:l,selection:o}),filter:!1,userEvent:t==0?"undo":"redo",scrollIntoView:!0})}else return null}};Ae.empty=new Ae(bt,bt);var Th=[{key:"Mod-z",run:Sh,preventDefault:!0},{key:"Mod-y",mac:"Mod-Shift-z",run:Gr,preventDefault:!0},{linux:"Ctrl-Shift-z",run:Gr,preventDefault:!0},{key:"Mod-u",run:ud,preventDefault:!0},{key:"Alt-u",mac:"Mod-Shift-u",run:dd,preventDefault:!0}];function Ze(s,t){return w.create(s.ranges.map(t),s.mainIndex)}function Mt(s,t){return s.update({selection:t,scrollIntoView:!0,userEvent:"select"})}function Dt({state:s,dispatch:t},e){let i=Ze(s.selection,e);return i.eq(s.selection,!0)?!1:(t(Mt(s,i)),!0)}function zs(s,t){return w.cursor(t?s.to:s.from)}function Ah(s,t){return Dt(s,e=>e.empty?s.moveByChar(e,t):zs(e,t))}function et(s){return s.textDirectionAt(s.state.selection.main.head)==X.LTR}var Ph=s=>Ah(s,!et(s)),Mh=s=>Ah(s,et(s));function Dh(s,t){return Dt(s,e=>e.empty?s.moveByGroup(e,t):zs(e,t))}var kd=s=>Dh(s,!et(s)),xd=s=>Dh(s,et(s));var Kg=typeof Intl<"u"&&Intl.Segmenter?new Intl.Segmenter(void 0,{granularity:"word"}):null;function vd(s,t,e){if(t.type.prop(e))return!0;let i=t.to-t.from;return i&&(i>2||/[^\s,.;:]/.test(s.sliceDoc(t.from,t.to)))||t.firstChild}function Qs(s,t,e){let i=At(s).resolveInner(t.head),n=e?D.closedBy:D.openedBy;for(let a=t.head;;){let h=e?i.childAfter(a):i.childBefore(a);if(!h)break;vd(s,h,n)?i=h:a=e?h.to:h.from}let r=i.type.prop(n),o,l;return r&&(o=e?Te(s,i.from,1):Te(s,i.to,-1))&&o.matched?l=e?o.end.to:o.end.from:l=e?i.to:i.from,w.cursor(l,e?-1:1)}var Sd=s=>Dt(s,t=>Qs(s.state,t,!et(s))),Od=s=>Dt(s,t=>Qs(s.state,t,et(s)));function Rh(s,t){return Dt(s,e=>{if(!e.empty)return zs(e,t);let i=s.moveVertically(e,t);return i.head!=e.head?i:s.moveToLineBoundary(e,t)})}var Eh=s=>Rh(s,!1),Bh=s=>Rh(s,!0);function Lh(s){let t=s.scrollDOM.clientHeight<s.scrollDOM.scrollHeight-2,e=0,i=0,n;if(t){for(let r of s.state.facet(A.scrollMargins)){let o=r(s);o?.top&&(e=Math.max(o?.top,e)),o?.bottom&&(i=Math.max(o?.bottom,i))}n=s.scrollDOM.clientHeight-e-i}else n=(s.dom.ownerDocument.defaultView||window).innerHeight;return{marginTop:e,marginBottom:i,selfScroll:t,height:Math.max(s.defaultLineHeight,n-5)}}function Ih(s,t){let e=Lh(s),{state:i}=s,n=Ze(i.selection,o=>o.empty?s.moveVertically(o,t,e.height):zs(o,t));if(n.eq(i.selection))return!1;let r;if(e.selfScroll){let o=s.coordsAtPos(i.selection.main.head),l=s.scrollDOM.getBoundingClientRect(),a=l.top+e.marginTop,h=l.bottom-e.marginBottom;o&&o.top>a&&o.bottom<h&&(r=A.scrollIntoView(n.main.head,{y:"start",yMargin:o.top-a}))}return s.dispatch(Mt(i,n),{effects:r}),!0}var fh=s=>Ih(s,!1),Kr=s=>Ih(s,!0);function re(s,t,e){let i=s.lineBlockAt(t.head),n=s.moveToLineBoundary(t,e);if(n.head==t.head&&n.head!=(e?i.to:i.from)&&(n=s.moveToLineBoundary(t,e,!1)),!e&&n.head==i.from&&i.length){let r=/^\s*/.exec(s.state.sliceDoc(i.from,Math.min(i.from+100,i.to)))[0].length;r&&t.head!=i.from+r&&(n=w.cursor(i.from+r))}return n}var Cd=s=>Dt(s,t=>re(s,t,!0)),Td=s=>Dt(s,t=>re(s,t,!1)),Ad=s=>Dt(s,t=>re(s,t,!et(s))),Pd=s=>Dt(s,t=>re(s,t,et(s))),Md=s=>Dt(s,t=>w.cursor(s.lineBlockAt(t.head).from,1)),Dd=s=>Dt(s,t=>w.cursor(s.lineBlockAt(t.head).to,-1));function Rd(s,t,e){let i=!1,n=Ze(s.selection,r=>{let o=Te(s,r.head,-1)||Te(s,r.head,1)||r.head>0&&Te(s,r.head-1,1)||r.head<s.doc.length&&Te(s,r.head+1,-1);if(!o||!o.end)return r;i=!0;let l=o.start.from==r.head?o.end.to:o.end.from;return e?w.range(r.anchor,l):w.cursor(l)});return i?(t(Mt(s,n)),!0):!1}var Ed=({state:s,dispatch:t})=>Rd(s,t,!1);function wt(s,t){let e=Ze(s.state.selection,i=>{let n=t(i);return w.range(i.anchor,n.head,n.goalColumn,n.bidiLevel||void 0)});return e.eq(s.state.selection)?!1:(s.dispatch(Mt(s.state,e)),!0)}function Nh(s,t){return wt(s,e=>s.moveByChar(e,t))}var _h=s=>Nh(s,!et(s)),Vh=s=>Nh(s,et(s));function Hh(s,t){return wt(s,e=>s.moveByGroup(e,t))}var Bd=s=>Hh(s,!et(s)),Ld=s=>Hh(s,et(s));var Id=s=>wt(s,t=>Qs(s.state,t,!et(s))),Nd=s=>wt(s,t=>Qs(s.state,t,et(s)));function Fh(s,t){return wt(s,e=>s.moveVertically(e,t))}var Wh=s=>Fh(s,!1),$h=s=>Fh(s,!0);function zh(s,t){return wt(s,e=>s.moveVertically(e,t,Lh(s).height))}var uh=s=>zh(s,!1),dh=s=>zh(s,!0),_d=s=>wt(s,t=>re(s,t,!0)),Vd=s=>wt(s,t=>re(s,t,!1)),Hd=s=>wt(s,t=>re(s,t,!et(s))),Fd=s=>wt(s,t=>re(s,t,et(s))),Wd=s=>wt(s,t=>w.cursor(s.lineBlockAt(t.head).from)),$d=s=>wt(s,t=>w.cursor(s.lineBlockAt(t.head).to)),ph=({state:s,dispatch:t})=>(t(Mt(s,{anchor:0})),!0),gh=({state:s,dispatch:t})=>(t(Mt(s,{anchor:s.doc.length})),!0),mh=({state:s,dispatch:t})=>(t(Mt(s,{anchor:s.selection.main.anchor,head:0})),!0),yh=({state:s,dispatch:t})=>(t(Mt(s,{anchor:s.selection.main.anchor,head:s.doc.length})),!0),zd=({state:s,dispatch:t})=>(t(s.update({selection:{anchor:0,head:s.doc.length},userEvent:"select"})),!0),Qd=({state:s,dispatch:t})=>{let e=js(s).map(({from:i,to:n})=>w.range(i,Math.min(n+1,s.doc.length)));return t(s.update({selection:w.create(e),userEvent:"select"})),!0},jd=({state:s,dispatch:t})=>{let e=Ze(s.selection,i=>{let n=At(s),r=n.resolveStack(i.from,1);if(i.empty){let o=n.resolveStack(i.from,-1);o.node.from>=r.node.from&&o.node.to<=r.node.to&&(r=o)}for(let o=r;o;o=o.next){let{node:l}=o;if((l.from<i.from&&l.to>=i.to||l.to>i.to&&l.from<=i.from)&&o.next)return w.range(l.to,l.from)}return i});return e.eq(s.selection)?!1:(t(Mt(s,e)),!0)};function Qh(s,t){let{state:e}=s,i=e.selection,n=e.selection.ranges.slice();for(let r of e.selection.ranges){let o=e.doc.lineAt(r.head);if(t?o.to<s.state.doc.length:o.from>0)for(let l=r;;){let a=s.moveVertically(l,t);if(a.head<o.from||a.head>o.to){n.some(h=>h.head==a.head)||n.push(a);break}else{if(a.head==l.head)break;l=a}}}return n.length==i.ranges.length?!1:(s.dispatch(Mt(e,w.create(n,n.length-1))),!0)}var Xd=s=>Qh(s,!1),qd=s=>Qh(s,!0),Ud=({state:s,dispatch:t})=>{let e=s.selection,i=null;return e.ranges.length>1?i=w.create([e.main]):e.main.empty||(i=w.create([w.cursor(e.main.head)])),i?(t(Mt(s,i)),!0):!1};function Hi(s,t){if(s.state.readOnly)return!1;let e="delete.selection",{state:i}=s,n=i.changeByRange(r=>{let{from:o,to:l}=r;if(o==l){let a=t(r);a<o?(e="delete.backward",a=Fs(s,a,!1)):a>o&&(e="delete.forward",a=Fs(s,a,!0)),o=Math.min(o,a),l=Math.max(l,a)}else o=Fs(s,o,!1),l=Fs(s,l,!0);return o==l?{range:r}:{changes:{from:o,to:l},range:w.cursor(o,o<r.head?-1:1)}});return n.changes.empty?!1:(s.dispatch(i.update(n,{scrollIntoView:!0,userEvent:e,effects:e=="delete.selection"?A.announce.of(i.phrase("Selection deleted")):void 0})),!0)}function Fs(s,t,e){if(s instanceof A)for(let i of s.state.facet(A.atomicRanges).map(n=>n(s)))i.between(t,t,(n,r)=>{n<t&&r>t&&(t=e?r:n)});return t}var jh=(s,t,e)=>Hi(s,i=>{let n=i.from,{state:r}=s,o=r.doc.lineAt(n),l,a;if(e&&!t&&n>o.from&&n<o.from+200&&!/[^ \t]/.test(l=o.text.slice(0,n-o.from))){if(l[l.length-1]=="	")return n-1;let h=pe(l,r.tabSize),c=h%Ni(r)||Ni(r);for(let f=0;f<c&&l[l.length-1-f]==" ";f++)n--;a=n}else a=Y(o.text,n-o.from,t,t)+o.from,a==n&&o.number!=(t?r.doc.lines:1)?a+=t?1:-1:!t&&/[\ufe00-\ufe0f]/.test(o.text.slice(a-o.from,n-o.from))&&(a=Y(o.text,a-o.from,!1,!1)+o.from);return a}),Yr=s=>jh(s,!1,!0);var Xh=s=>jh(s,!0,!1),qh=(s,t)=>Hi(s,e=>{let i=e.head,{state:n}=s,r=n.doc.lineAt(i),o=n.charCategorizer(i);for(let l=null;;){if(i==(t?r.to:r.from)){i==e.head&&r.number!=(t?n.doc.lines:1)&&(i+=t?1:-1);break}let a=Y(r.text,i-r.from,t)+r.from,h=r.text.slice(Math.min(i,a)-r.from,Math.max(i,a)-r.from),c=o(h);if(l!=null&&c!=l)break;(h!=" "||i!=e.head)&&(l=c),i=a}return i}),Uh=s=>qh(s,!1),Gd=s=>qh(s,!0);var Kd=s=>Hi(s,t=>{let e=s.lineBlockAt(t.head).to;return t.head<e?e:Math.min(s.state.doc.length,t.head+1)});var Yd=s=>Hi(s,t=>{let e=s.moveToLineBoundary(t,!1).head;return t.head>e?e:Math.max(0,t.head-1)}),Jd=s=>Hi(s,t=>{let e=s.moveToLineBoundary(t,!0).head;return t.head<e?e:Math.min(s.state.doc.length,t.head+1)});var Zd=({state:s,dispatch:t})=>{if(s.readOnly)return!1;let e=s.changeByRange(i=>({changes:{from:i.from,to:i.to,insert:R.of(["",""])},range:w.cursor(i.from)}));return t(s.update(e,{scrollIntoView:!0,userEvent:"input"})),!0},tp=({state:s,dispatch:t})=>{if(s.readOnly)return!1;let e=s.changeByRange(i=>{if(!i.empty||i.from==0||i.from==s.doc.length)return{range:i};let n=i.from,r=s.doc.lineAt(n),o=n==r.from?n-1:Y(r.text,n-r.from,!1)+r.from,l=n==r.to?n+1:Y(r.text,n-r.from,!0)+r.from;return{changes:{from:o,to:l,insert:s.doc.slice(n,l).append(s.doc.slice(o,n))},range:w.cursor(l)}});return e.changes.empty?!1:(t(s.update(e,{scrollIntoView:!0,userEvent:"move.character"})),!0)};function js(s){let t=[],e=-1;for(let i of s.selection.ranges){let n=s.doc.lineAt(i.from),r=s.doc.lineAt(i.to);if(!i.empty&&i.to==r.from&&(r=s.doc.lineAt(i.to-1)),e>=n.number){let o=t[t.length-1];o.to=r.to,o.ranges.push(i)}else t.push({from:n.from,to:r.to,ranges:[i]});e=r.number+1}return t}function Gh(s,t,e){if(s.readOnly)return!1;let i=[],n=[];for(let r of js(s)){if(e?r.to==s.doc.length:r.from==0)continue;let o=s.doc.lineAt(e?r.to+1:r.from-1),l=o.length+1;if(e){i.push({from:r.to,to:o.to},{from:r.from,insert:o.text+s.lineBreak});for(let a of r.ranges)n.push(w.range(Math.min(s.doc.length,a.anchor+l),Math.min(s.doc.length,a.head+l)))}else{i.push({from:o.from,to:r.from},{from:r.to,insert:s.lineBreak+o.text});for(let a of r.ranges)n.push(w.range(a.anchor-l,a.head-l))}}return i.length?(t(s.update({changes:i,scrollIntoView:!0,selection:w.create(n,s.selection.mainIndex),userEvent:"move.line"})),!0):!1}var ep=({state:s,dispatch:t})=>Gh(s,t,!1),ip=({state:s,dispatch:t})=>Gh(s,t,!0);function Kh(s,t,e){if(s.readOnly)return!1;let i=[];for(let r of js(s))e?i.push({from:r.from,insert:s.doc.slice(r.from,r.to)+s.lineBreak}):i.push({from:r.to,insert:s.lineBreak+s.doc.slice(r.from,r.to)});let n=s.changes(i);return t(s.update({changes:n,selection:s.selection.map(n,e?1:-1),scrollIntoView:!0,userEvent:"input.copyline"})),!0}var sp=({state:s,dispatch:t})=>Kh(s,t,!1),np=({state:s,dispatch:t})=>Kh(s,t,!0),rp=s=>{if(s.state.readOnly)return!1;let{state:t}=s,e=t.changes(js(t).map(({from:n,to:r})=>(n>0?n--:r<t.doc.length&&r++,{from:n,to:r}))),i=Ze(t.selection,n=>{let r;if(s.lineWrapping){let o=s.lineBlockAt(n.head),l=s.coordsAtPos(n.head,n.assoc||1);l&&(r=o.bottom+s.documentTop-l.bottom+s.defaultLineHeight/2)}return s.moveVertically(n,!0,r)}).map(e);return s.dispatch({changes:e,selection:i,scrollIntoView:!0,userEvent:"delete.line"}),!0};function op(s,t){if(/\(\)|\[\]|\{\}/.test(s.sliceDoc(t-1,t+1)))return{from:t,to:t};let e=At(s).resolveInner(t),i=e.childBefore(t),n=e.childAfter(t),r;return i&&n&&i.to<=t&&n.from>=t&&(r=i.type.prop(D.closedBy))&&r.indexOf(n.name)>-1&&s.doc.lineAt(i.to).from==s.doc.lineAt(n.from).from&&!/\S/.test(s.sliceDoc(i.to,n.from))?{from:i.to,to:n.from}:null}var bh=Yh(!1),lp=Yh(!0);function Yh(s){return({state:t,dispatch:e})=>{if(t.readOnly)return!1;let i=t.changeByRange(n=>{let{from:r,to:o}=n,l=t.doc.lineAt(r),a=!s&&r==o&&op(t,r);s&&(r=o=(o<=l.to?l:t.doc.lineAt(o)).to);let h=new Ce(t,{simulateBreak:r,simulateDoubleBreak:!!a}),c=Qr(h,r);for(c==null&&(c=pe(/^\s*/.exec(t.doc.lineAt(r).text)[0],t.tabSize));o<l.to&&/\s/.test(l.text[o-l.from]);)o++;a?{from:r,to:o}=a:r>l.from&&r<l.from+100&&!/\S/.test(l.text.slice(0,r))&&(r=l.from);let f=["",_i(t,c)];return a&&f.push(_i(t,h.lineIndent(l.from,-1))),{changes:{from:r,to:o,insert:R.of(f)},range:w.cursor(r+1+f[1].length)}});return e(t.update(i,{scrollIntoView:!0,userEvent:"input"})),!0}}function to(s,t){let e=-1;return s.changeByRange(i=>{let n=[];for(let o=i.from;o<=i.to;){let l=s.doc.lineAt(o);l.number>e&&(i.empty||i.to>l.from)&&(t(l,n,i),e=l.number),o=l.to+1}let r=s.changes(n);return{changes:n,range:w.range(r.mapPos(i.anchor,1),r.mapPos(i.head,1))}})}var ap=({state:s,dispatch:t})=>{if(s.readOnly)return!1;let e=Object.create(null),i=new Ce(s,{overrideIndentation:r=>{let o=e[r];return o??-1}}),n=to(s,(r,o,l)=>{let a=Qr(i,r.from);if(a==null)return;/\S/.test(r.text)||(a=0);let h=/^\s*/.exec(r.text)[0],c=_i(s,a);(h!=c||l.from<r.from+h.length)&&(e[r.from]=a,o.push({from:r.from,to:r.from+h.length,insert:c}))});return n.changes.empty||t(s.update(n,{userEvent:"indent"})),!0},hp=({state:s,dispatch:t})=>s.readOnly?!1:(t(s.update(to(s,(e,i)=>{i.push({from:e.from,insert:s.facet(Hs)})}),{userEvent:"input.indent"})),!0),cp=({state:s,dispatch:t})=>s.readOnly?!1:(t(s.update(to(s,(e,i)=>{let n=/^\s*/.exec(e.text)[0];if(!n)return;let r=pe(n,s.tabSize),o=0,l=_i(s,Math.max(0,r-Ni(s)));for(;o<n.length&&o<l.length&&n.charCodeAt(o)==l.charCodeAt(o);)o++;i.push({from:e.from+o,to:e.from+n.length,insert:l.slice(o)})}),{userEvent:"delete.dedent"})),!0),fp=s=>(s.setTabFocusMode(),!0);var up=[{key:"Ctrl-b",run:Ph,shift:_h,preventDefault:!0},{key:"Ctrl-f",run:Mh,shift:Vh},{key:"Ctrl-p",run:Eh,shift:Wh},{key:"Ctrl-n",run:Bh,shift:$h},{key:"Ctrl-a",run:Md,shift:Wd},{key:"Ctrl-e",run:Dd,shift:$d},{key:"Ctrl-d",run:Xh},{key:"Ctrl-h",run:Yr},{key:"Ctrl-k",run:Kd},{key:"Ctrl-Alt-h",run:Uh},{key:"Ctrl-o",run:Zd},{key:"Ctrl-t",run:tp},{key:"Ctrl-v",run:Kr}],dp=[{key:"ArrowLeft",run:Ph,shift:_h,preventDefault:!0},{key:"Mod-ArrowLeft",mac:"Alt-ArrowLeft",run:kd,shift:Bd,preventDefault:!0},{mac:"Cmd-ArrowLeft",run:Ad,shift:Hd,preventDefault:!0},{key:"ArrowRight",run:Mh,shift:Vh,preventDefault:!0},{key:"Mod-ArrowRight",mac:"Alt-ArrowRight",run:xd,shift:Ld,preventDefault:!0},{mac:"Cmd-ArrowRight",run:Pd,shift:Fd,preventDefault:!0},{key:"ArrowUp",run:Eh,shift:Wh,preventDefault:!0},{mac:"Cmd-ArrowUp",run:ph,shift:mh},{mac:"Ctrl-ArrowUp",run:fh,shift:uh},{key:"ArrowDown",run:Bh,shift:$h,preventDefault:!0},{mac:"Cmd-ArrowDown",run:gh,shift:yh},{mac:"Ctrl-ArrowDown",run:Kr,shift:dh},{key:"PageUp",run:fh,shift:uh},{key:"PageDown",run:Kr,shift:dh},{key:"Home",run:Td,shift:Vd,preventDefault:!0},{key:"Mod-Home",run:ph,shift:mh},{key:"End",run:Cd,shift:_d,preventDefault:!0},{key:"Mod-End",run:gh,shift:yh},{key:"Enter",run:bh,shift:bh},{key:"Mod-a",run:zd},{key:"Backspace",run:Yr,shift:Yr,preventDefault:!0},{key:"Delete",run:Xh,preventDefault:!0},{key:"Mod-Backspace",mac:"Alt-Backspace",run:Uh,preventDefault:!0},{key:"Mod-Delete",mac:"Alt-Delete",run:Gd,preventDefault:!0},{mac:"Mod-Backspace",run:Yd,preventDefault:!0},{mac:"Mod-Delete",run:Jd,preventDefault:!0}].concat(up.map(s=>({mac:s.key,run:s.run,shift:s.shift}))),Jh=[{key:"Alt-ArrowLeft",mac:"Ctrl-ArrowLeft",run:Sd,shift:Id},{key:"Alt-ArrowRight",mac:"Ctrl-ArrowRight",run:Od,shift:Nd},{key:"Alt-ArrowUp",run:ep},{key:"Shift-Alt-ArrowUp",run:sp},{key:"Alt-ArrowDown",run:ip},{key:"Shift-Alt-ArrowDown",run:np},{key:"Mod-Alt-ArrowUp",run:Xd},{key:"Mod-Alt-ArrowDown",run:qd},{key:"Escape",run:Ud},{key:"Mod-Enter",run:lp},{key:"Alt-l",mac:"Ctrl-l",run:Qd},{key:"Mod-i",run:jd,preventDefault:!0},{key:"Mod-[",run:cp},{key:"Mod-]",run:hp},{key:"Mod-Alt-\\",run:ap},{key:"Shift-Mod-k",run:rp},{key:"Shift-Mod-\\",run:Ed},{key:"Mod-/",run:sd},{key:"Alt-A",run:rd},{key:"Ctrl-m",mac:"Shift-Alt-m",run:fp}].concat(dp);var so=class s{constructor(t,e,i,n,r,o,l,a,h,c=0,f){this.p=t,this.stack=e,this.state=i,this.reducePos=n,this.pos=r,this.score=o,this.buffer=l,this.bufferBase=a,this.curContext=h,this.lookAhead=c,this.parent=f}toString(){return`[${this.stack.filter((t,e)=>e%3==0).concat(this.state)}]@${this.pos}${this.score?"!"+this.score:""}`}static start(t,e,i=0){let n=t.parser.context;return new s(t,[],e,i,i,0,[],0,n?new Xs(n,n.start):null,0,null)}get context(){return this.curContext?this.curContext.context:null}pushState(t,e){this.stack.push(this.state,e,this.bufferBase+this.buffer.length),this.state=t}reduce(t){var e;let i=t>>19,n=t&65535,{parser:r}=this.p,o=this.reducePos<this.pos-25&&this.setLookAhead(this.pos),l=r.dynamicPrecedence(n);if(l&&(this.score+=l),i==0){this.pushState(r.getGoto(this.state,n,!0),this.reducePos),n<r.minRepeatTerm&&this.storeNode(n,this.reducePos,this.reducePos,o?8:4,!0),this.reduceContext(n,this.reducePos);return}let a=this.stack.length-(i-1)*3-(t&262144?6:0),h=a?this.stack[a-2]:this.p.ranges[0].from,c=this.reducePos-h;c>=2e3&&!(!((e=this.p.parser.nodeSet.types[n])===null||e===void 0)&&e.isAnonymous)&&(h==this.p.lastBigReductionStart?(this.p.bigReductionCount++,this.p.lastBigReductionSize=c):this.p.lastBigReductionSize<c&&(this.p.bigReductionCount=1,this.p.lastBigReductionStart=h,this.p.lastBigReductionSize=c));let f=a?this.stack[a-1]:0,u=this.bufferBase+this.buffer.length-f;if(n<r.minRepeatTerm||t&131072){let d=r.stateFlag(this.state,1)?this.pos:this.reducePos;this.storeNode(n,h,d,u+4,!0)}if(t&262144)this.state=this.stack[a];else{let d=this.stack[a-3];this.state=r.getGoto(d,n,!0)}for(;this.stack.length>a;)this.stack.pop();this.reduceContext(n,h)}storeNode(t,e,i,n=4,r=!1){if(t==0&&(!this.stack.length||this.stack[this.stack.length-1]<this.buffer.length+this.bufferBase)){let o=this,l=this.buffer.length;if(l==0&&o.parent&&(l=o.bufferBase-o.parent.bufferBase,o=o.parent),l>0&&o.buffer[l-4]==0&&o.buffer[l-1]>-1){if(e==i)return;if(o.buffer[l-2]>=e){o.buffer[l-2]=i;return}}}if(!r||this.pos==i)this.buffer.push(t,e,i,n);else{let o=this.buffer.length;if(o>0&&(this.buffer[o-4]!=0||this.buffer[o-1]<0)){let l=!1;for(let a=o;a>0&&this.buffer[a-2]>i;a-=4)if(this.buffer[a-1]>=0){l=!0;break}if(l)for(;o>0&&this.buffer[o-2]>i;)this.buffer[o]=this.buffer[o-4],this.buffer[o+1]=this.buffer[o-3],this.buffer[o+2]=this.buffer[o-2],this.buffer[o+3]=this.buffer[o-1],o-=4,n>4&&(n-=4)}this.buffer[o]=t,this.buffer[o+1]=e,this.buffer[o+2]=i,this.buffer[o+3]=n}}shift(t,e,i,n){if(t&131072)this.pushState(t&65535,this.pos);else if((t&262144)==0){let r=t,{parser:o}=this.p;this.pos=n;let l=o.stateFlag(r,1);!l&&(n>i||e<=o.maxNode)&&(this.reducePos=n),this.pushState(r,l?i:Math.min(i,this.reducePos)),this.shiftContext(e,i),e<=o.maxNode&&this.buffer.push(e,i,n,4)}else this.pos=n,this.shiftContext(e,i),e<=this.p.parser.maxNode&&this.buffer.push(e,i,n,4)}apply(t,e,i,n){t&65536?this.reduce(t):this.shift(t,e,i,n)}useNode(t,e){let i=this.p.reused.length-1;(i<0||this.p.reused[i]!=t)&&(this.p.reused.push(t),i++);let n=this.pos;this.reducePos=this.pos=n+t.length,this.pushState(e,n),this.buffer.push(i,n,this.reducePos,-1),this.curContext&&this.updateContext(this.curContext.tracker.reuse(this.curContext.context,t,this,this.p.stream.reset(this.pos-t.length)))}split(){let t=this,e=t.buffer.length;for(;e>0&&t.buffer[e-2]>t.reducePos;)e-=4;let i=t.buffer.slice(e),n=t.bufferBase+e;for(;t&&n==t.bufferBase;)t=t.parent;return new s(this.p,this.stack.slice(),this.state,this.reducePos,this.pos,this.score,i,n,this.curContext,this.lookAhead,t)}recoverByDelete(t,e){let i=t<=this.p.parser.maxNode;i&&this.storeNode(t,this.pos,e,4),this.storeNode(0,this.pos,e,i?8:4),this.pos=this.reducePos=e,this.score-=190}canShift(t){for(let e=new no(this);;){let i=this.p.parser.stateSlot(e.state,4)||this.p.parser.hasAction(e.state,t);if(i==0)return!1;if((i&65536)==0)return!0;e.reduce(i)}}recoverByInsert(t){if(this.stack.length>=300)return[];let e=this.p.parser.nextStates(this.state);if(e.length>8||this.stack.length>=120){let n=[];for(let r=0,o;r<e.length;r+=2)(o=e[r+1])!=this.state&&this.p.parser.hasAction(o,t)&&n.push(e[r],o);if(this.stack.length<120)for(let r=0;n.length<8&&r<e.length;r+=2){let o=e[r+1];n.some((l,a)=>a&1&&l==o)||n.push(e[r],o)}e=n}let i=[];for(let n=0;n<e.length&&i.length<4;n+=2){let r=e[n+1];if(r==this.state)continue;let o=this.split();o.pushState(r,this.pos),o.storeNode(0,o.pos,o.pos,4,!0),o.shiftContext(e[n],this.pos),o.reducePos=this.pos,o.score-=200,i.push(o)}return i}forceReduce(){let{parser:t}=this.p,e=t.stateSlot(this.state,5);if((e&65536)==0)return!1;if(!t.validAction(this.state,e)){let i=e>>19,n=e&65535,r=this.stack.length-i*3;if(r<0||t.getGoto(this.stack[r],n,!1)<0){let o=this.findForcedReduction();if(o==null)return!1;e=o}this.storeNode(0,this.pos,this.pos,4,!0),this.score-=100}return this.reducePos=this.pos,this.reduce(e),!0}findForcedReduction(){let{parser:t}=this.p,e=[],i=(n,r)=>{if(!e.includes(n))return e.push(n),t.allActions(n,o=>{if(!(o&393216))if(o&65536){let l=(o>>19)-r;if(l>1){let a=o&65535,h=this.stack.length-l*3;if(h>=0&&t.getGoto(this.stack[h],a,!1)>=0)return l<<19|65536|a}}else{let l=i(o,r+1);if(l!=null)return l}})};return i(this.state,0)}forceAll(){for(;!this.p.parser.stateFlag(this.state,2);)if(!this.forceReduce()){this.storeNode(0,this.pos,this.pos,4,!0);break}return this}get deadEnd(){if(this.stack.length!=3)return!1;let{parser:t}=this.p;return t.data[t.stateSlot(this.state,1)]==65535&&!t.stateSlot(this.state,4)}restart(){this.storeNode(0,this.pos,this.pos,4,!0),this.state=this.stack[0],this.stack.length=0}sameState(t){if(this.state!=t.state||this.stack.length!=t.stack.length)return!1;for(let e=0;e<this.stack.length;e+=3)if(this.stack[e]!=t.stack[e])return!1;return!0}get parser(){return this.p.parser}dialectEnabled(t){return this.p.parser.dialect.flags[t]}shiftContext(t,e){this.curContext&&this.updateContext(this.curContext.tracker.shift(this.curContext.context,t,this,this.p.stream.reset(e)))}reduceContext(t,e){this.curContext&&this.updateContext(this.curContext.tracker.reduce(this.curContext.context,t,this,this.p.stream.reset(e)))}emitContext(){let t=this.buffer.length-1;(t<0||this.buffer[t]!=-3)&&this.buffer.push(this.curContext.hash,this.pos,this.pos,-3)}emitLookAhead(){let t=this.buffer.length-1;(t<0||this.buffer[t]!=-4)&&this.buffer.push(this.lookAhead,this.pos,this.pos,-4)}updateContext(t){if(t!=this.curContext.context){let e=new Xs(this.curContext.tracker,t);e.hash!=this.curContext.hash&&this.emitContext(),this.curContext=e}}setLookAhead(t){return t<=this.lookAhead?!1:(this.emitLookAhead(),this.lookAhead=t,!0)}close(){this.curContext&&this.curContext.tracker.strict&&this.emitContext(),this.lookAhead>0&&this.emitLookAhead()}},Xs=class{constructor(t,e){this.tracker=t,this.context=e,this.hash=t.strict?t.hash(e):0}},no=class{constructor(t){this.start=t,this.state=t.state,this.stack=t.stack,this.base=this.stack.length}reduce(t){let e=t&65535,i=t>>19;i==0?(this.stack==this.start.stack&&(this.stack=this.stack.slice()),this.stack.push(this.state,0,0),this.base+=3):this.base-=(i-1)*3;let n=this.start.p.parser.getGoto(this.stack[this.base-3],e,!0);this.state=n}},ro=class s{constructor(t,e,i){this.stack=t,this.pos=e,this.index=i,this.buffer=t.buffer,this.index==0&&this.maybeNext()}static create(t,e=t.bufferBase+t.buffer.length){return new s(t,e,e-t.bufferBase)}maybeNext(){let t=this.stack.parent;t!=null&&(this.index=this.stack.bufferBase-t.bufferBase,this.stack=t,this.buffer=t.buffer)}get id(){return this.buffer[this.index-4]}get start(){return this.buffer[this.index-3]}get end(){return this.buffer[this.index-2]}get size(){return this.buffer[this.index-1]}next(){this.index-=4,this.pos-=4,this.index==0&&this.maybeNext()}fork(){return new s(this.stack,this.pos,this.index)}};function Fi(s,t=Uint16Array){if(typeof s!="string")return s;let e=null;for(let i=0,n=0;i<s.length;){let r=0;for(;;){let o=s.charCodeAt(i++),l=!1;if(o==126){r=65535;break}o>=92&&o--,o>=34&&o--;let a=o-32;if(a>=46&&(a-=46,l=!0),r+=a,l)break;r*=46}e?e[n++]=r:e=new t(r)}return e}var ti=class{constructor(){this.start=-1,this.value=-1,this.end=-1,this.extended=-1,this.lookAhead=0,this.mask=0,this.context=0}},Zh=new ti,oo=class{constructor(t,e){this.input=t,this.ranges=e,this.chunk="",this.chunkOff=0,this.chunk2="",this.chunk2Pos=0,this.next=-1,this.token=Zh,this.rangeIndex=0,this.pos=this.chunkPos=e[0].from,this.range=e[0],this.end=e[e.length-1].to,this.readNext()}resolveOffset(t,e){let i=this.range,n=this.rangeIndex,r=this.pos+t;for(;r<i.from;){if(!n)return null;let o=this.ranges[--n];r-=i.from-o.to,i=o}for(;e<0?r>i.to:r>=i.to;){if(n==this.ranges.length-1)return null;let o=this.ranges[++n];r+=o.from-i.to,i=o}return r}clipPos(t){if(t>=this.range.from&&t<this.range.to)return t;for(let e of this.ranges)if(e.to>t)return Math.max(t,e.from);return this.end}peek(t){let e=this.chunkOff+t,i,n;if(e>=0&&e<this.chunk.length)i=this.pos+t,n=this.chunk.charCodeAt(e);else{let r=this.resolveOffset(t,1);if(r==null)return-1;if(i=r,i>=this.chunk2Pos&&i<this.chunk2Pos+this.chunk2.length)n=this.chunk2.charCodeAt(i-this.chunk2Pos);else{let o=this.rangeIndex,l=this.range;for(;l.to<=i;)l=this.ranges[++o];this.chunk2=this.input.chunk(this.chunk2Pos=i),i+this.chunk2.length>l.to&&(this.chunk2=this.chunk2.slice(0,l.to-i)),n=this.chunk2.charCodeAt(0)}}return i>=this.token.lookAhead&&(this.token.lookAhead=i+1),n}acceptToken(t,e=0){let i=e?this.resolveOffset(e,-1):this.pos;if(i==null||i<this.token.start)throw new RangeError("Token end out of bounds");this.token.value=t,this.token.end=i}acceptTokenTo(t,e){this.token.value=t,this.token.end=e}getChunk(){if(this.pos>=this.chunk2Pos&&this.pos<this.chunk2Pos+this.chunk2.length){let{chunk:t,chunkPos:e}=this;this.chunk=this.chunk2,this.chunkPos=this.chunk2Pos,this.chunk2=t,this.chunk2Pos=e,this.chunkOff=this.pos-this.chunkPos}else{this.chunk2=this.chunk,this.chunk2Pos=this.chunkPos;let t=this.input.chunk(this.pos),e=this.pos+t.length;this.chunk=e>this.range.to?t.slice(0,this.range.to-this.pos):t,this.chunkPos=this.pos,this.chunkOff=0}}readNext(){return this.chunkOff>=this.chunk.length&&(this.getChunk(),this.chunkOff==this.chunk.length)?this.next=-1:this.next=this.chunk.charCodeAt(this.chunkOff)}advance(t=1){for(this.chunkOff+=t;this.pos+t>=this.range.to;){if(this.rangeIndex==this.ranges.length-1)return this.setDone();t-=this.range.to-this.pos,this.range=this.ranges[++this.rangeIndex],this.pos=this.range.from}return this.pos+=t,this.pos>=this.token.lookAhead&&(this.token.lookAhead=this.pos+1),this.readNext()}setDone(){return this.pos=this.chunkPos=this.end,this.range=this.ranges[this.rangeIndex=this.ranges.length-1],this.chunk="",this.next=-1}reset(t,e){if(e?(this.token=e,e.start=t,e.lookAhead=t+1,e.value=e.extended=-1):this.token=Zh,this.pos!=t){if(this.pos=t,t==this.end)return this.setDone(),this;for(;t<this.range.from;)this.range=this.ranges[--this.rangeIndex];for(;t>=this.range.to;)this.range=this.ranges[++this.rangeIndex];t>=this.chunkPos&&t<this.chunkPos+this.chunk.length?this.chunkOff=t-this.chunkPos:(this.chunk="",this.chunkOff=0),this.readNext()}return this}read(t,e){if(t>=this.chunkPos&&e<=this.chunkPos+this.chunk.length)return this.chunk.slice(t-this.chunkPos,e-this.chunkPos);if(t>=this.chunk2Pos&&e<=this.chunk2Pos+this.chunk2.length)return this.chunk2.slice(t-this.chunk2Pos,e-this.chunk2Pos);if(t>=this.range.from&&e<=this.range.to)return this.input.read(t,e);let i="";for(let n of this.ranges){if(n.from>=e)break;n.to>t&&(i+=this.input.read(Math.max(n.from,t),Math.min(n.to,e)))}return i}},oe=class{constructor(t,e){this.data=t,this.id=e}token(t,e){let{parser:i}=e.p;nc(this.data,t,e,this.id,i.data,i.tokenPrecTable)}};oe.prototype.contextual=oe.prototype.fallback=oe.prototype.extend=!1;var lo=class{constructor(t,e,i){this.precTable=e,this.elseToken=i,this.data=typeof t=="string"?Fi(t):t}token(t,e){let i=t.pos,n=0;for(;;){let r=t.next<0,o=t.resolveOffset(1,1);if(nc(this.data,t,e,0,this.data,this.precTable),t.token.value>-1)break;if(this.elseToken==null)return;if(r||n++,o==null)break;t.reset(o,t.token)}n&&(t.reset(i,t.token),t.acceptToken(this.elseToken,n))}};lo.prototype.contextual=oe.prototype.fallback=oe.prototype.extend=!1;var Pe=class{constructor(t,e={}){this.token=t,this.contextual=!!e.contextual,this.fallback=!!e.fallback,this.extend=!!e.extend}};function nc(s,t,e,i,n,r){let o=0,l=1<<i,{dialect:a}=e.p.parser;t:for(;(l&s[o])!=0;){let h=s[o+1];for(let d=o+3;d<h;d+=2)if((s[d+1]&l)>0){let p=s[d];if(a.allows(p)&&(t.token.value==-1||t.token.value==p||gp(p,t.token.value,n,r))){t.acceptToken(p);break}}let c=t.next,f=0,u=s[o+2];if(t.next<0&&u>f&&s[h+u*3-3]==65535){o=s[h+u*3-1];continue t}for(;f<u;){let d=f+u>>1,p=h+d+(d<<1),g=s[p],y=s[p+1]||65536;if(c<g)u=d;else if(c>=y)f=d+1;else{o=s[p+2],t.advance();continue t}}break}}function tc(s,t,e){for(let i=t,n;(n=s[i])!=65535;i++)if(n==e)return i-t;return-1}function gp(s,t,e,i){let n=tc(e,i,t);return n<0||tc(e,i,s)<n}var ft=typeof process<"u"&&process.env&&/\bparse\b/.test(process.env.LOG),eo=null;function ec(s,t,e){let i=s.cursor(Q.IncludeAnonymous);for(i.moveTo(t);;)if(!(e<0?i.childBefore(t):i.childAfter(t)))for(;;){if((e<0?i.to<t:i.from>t)&&!i.type.isError)return e<0?Math.max(0,Math.min(i.to-1,t-25)):Math.min(s.length,Math.max(i.from+1,t+25));if(e<0?i.prevSibling():i.nextSibling())break;if(!i.parent())return e<0?0:s.length}}var ao=class{constructor(t,e){this.fragments=t,this.nodeSet=e,this.i=0,this.fragment=null,this.safeFrom=-1,this.safeTo=-1,this.trees=[],this.start=[],this.index=[],this.nextFragment()}nextFragment(){let t=this.fragment=this.i==this.fragments.length?null:this.fragments[this.i++];if(t){for(this.safeFrom=t.openStart?ec(t.tree,t.from+t.offset,1)-t.offset:t.from,this.safeTo=t.openEnd?ec(t.tree,t.to+t.offset,-1)-t.offset:t.to;this.trees.length;)this.trees.pop(),this.start.pop(),this.index.pop();this.trees.push(t.tree),this.start.push(-t.offset),this.index.push(0),this.nextStart=this.safeFrom}else this.nextStart=1e9}nodeAt(t){if(t<this.nextStart)return null;for(;this.fragment&&this.safeTo<=t;)this.nextFragment();if(!this.fragment)return null;for(;;){let e=this.trees.length-1;if(e<0)return this.nextFragment(),null;let i=this.trees[e],n=this.index[e];if(n==i.children.length){this.trees.pop(),this.start.pop(),this.index.pop();continue}let r=i.children[n],o=this.start[e]+i.positions[n];if(o>t)return this.nextStart=o,null;if(r instanceof U){if(o==t){if(o<this.safeFrom)return null;let l=o+r.length;if(l<=this.safeTo){let a=r.prop(D.lookAhead);if(!a||l+a<this.fragment.to)return r}}this.index[e]++,o+r.length>=Math.max(this.safeFrom,t)&&(this.trees.push(r),this.start.push(o),this.index.push(0))}else this.index[e]++,this.nextStart=o+r.length}}},ho=class{constructor(t,e){this.stream=e,this.tokens=[],this.mainToken=null,this.actions=[],this.tokens=t.tokenizers.map(i=>new ti)}getActions(t){let e=0,i=null,{parser:n}=t.p,{tokenizers:r}=n,o=n.stateSlot(t.state,3),l=t.curContext?t.curContext.hash:0,a=0;for(let h=0;h<r.length;h++){if((1<<h&o)==0)continue;let c=r[h],f=this.tokens[h];if(!(i&&!c.fallback)&&((c.contextual||f.start!=t.pos||f.mask!=o||f.context!=l)&&(this.updateCachedToken(f,c,t),f.mask=o,f.context=l),f.lookAhead>f.end+25&&(a=Math.max(f.lookAhead,a)),f.value!=0)){let u=e;if(f.extended>-1&&(e=this.addActions(t,f.extended,f.end,e)),e=this.addActions(t,f.value,f.end,e),!c.extend&&(i=f,e>u))break}}for(;this.actions.length>e;)this.actions.pop();return a&&t.setLookAhead(a),!i&&t.pos==this.stream.end&&(i=new ti,i.value=t.p.parser.eofTerm,i.start=i.end=t.pos,e=this.addActions(t,i.value,i.end,e)),this.mainToken=i,this.actions}getMainToken(t){if(this.mainToken)return this.mainToken;let e=new ti,{pos:i,p:n}=t;return e.start=i,e.end=Math.min(i+1,n.stream.end),e.value=i==n.stream.end?n.parser.eofTerm:0,e}updateCachedToken(t,e,i){let n=this.stream.clipPos(i.pos);if(e.token(this.stream.reset(n,t),i),t.value>-1){let{parser:r}=i.p;for(let o=0;o<r.specialized.length;o++)if(r.specialized[o]==t.value){let l=r.specializers[o](this.stream.read(t.start,t.end),i);if(l>=0&&i.p.parser.dialect.allows(l>>1)){(l&1)==0?t.value=l>>1:t.extended=l>>1;break}}}else t.value=0,t.end=this.stream.clipPos(n+1)}putAction(t,e,i,n){for(let r=0;r<n;r+=3)if(this.actions[r]==t)return n;return this.actions[n++]=t,this.actions[n++]=e,this.actions[n++]=i,n}addActions(t,e,i,n){let{state:r}=t,{parser:o}=t.p,{data:l}=o;for(let a=0;a<2;a++)for(let h=o.stateSlot(r,a?2:1);;h+=3){if(l[h]==65535)if(l[h+1]==1)h=qt(l,h+2);else{n==0&&l[h+1]==2&&(n=this.putAction(qt(l,h+2),e,i,n));break}l[h]==e&&(n=this.putAction(qt(l,h+1),e,i,n))}return n}},co=class{constructor(t,e,i,n){this.parser=t,this.input=e,this.ranges=n,this.recovering=0,this.nextStackID=9812,this.minStackPos=0,this.reused=[],this.stoppedAt=null,this.lastBigReductionStart=-1,this.lastBigReductionSize=0,this.bigReductionCount=0,this.stream=new oo(e,n),this.tokens=new ho(t,this.stream),this.topTerm=t.top[1];let{from:r}=n[0];this.stacks=[so.start(this,t.top[0],r)],this.fragments=i.length&&this.stream.end-r>t.bufferLength*4?new ao(i,t.nodeSet):null}get parsedPos(){return this.minStackPos}advance(){let t=this.stacks,e=this.minStackPos,i=this.stacks=[],n,r;if(this.bigReductionCount>300&&t.length==1){let[o]=t;for(;o.forceReduce()&&o.stack.length&&o.stack[o.stack.length-2]>=this.lastBigReductionStart;);this.bigReductionCount=this.lastBigReductionSize=0}for(let o=0;o<t.length;o++){let l=t[o];for(;;){if(this.tokens.mainToken=null,l.pos>e)i.push(l);else{if(this.advanceStack(l,i,t))continue;{n||(n=[],r=[]),n.push(l);let a=this.tokens.getMainToken(l);r.push(a.value,a.end)}}break}}if(!i.length){let o=n&&mp(n);if(o)return ft&&console.log("Finish with "+this.stackID(o)),this.stackToTree(o);if(this.parser.strict)throw ft&&n&&console.log("Stuck with token "+(this.tokens.mainToken?this.parser.getName(this.tokens.mainToken.value):"none")),new SyntaxError("No parse at "+e);this.recovering||(this.recovering=5)}if(this.recovering&&n){let o=this.stoppedAt!=null&&n[0].pos>this.stoppedAt?n[0]:this.runRecovery(n,r,i);if(o)return ft&&console.log("Force-finish "+this.stackID(o)),this.stackToTree(o.forceAll())}if(this.recovering){let o=this.recovering==1?1:this.recovering*3;if(i.length>o)for(i.sort((l,a)=>a.score-l.score);i.length>o;)i.pop();i.some(l=>l.reducePos>e)&&this.recovering--}else if(i.length>1){t:for(let o=0;o<i.length-1;o++){let l=i[o];for(let a=o+1;a<i.length;a++){let h=i[a];if(l.sameState(h)||l.buffer.length>500&&h.buffer.length>500)if((l.score-h.score||l.buffer.length-h.buffer.length)>0)i.splice(a--,1);else{i.splice(o--,1);continue t}}}i.length>12&&(i.sort((o,l)=>l.score-o.score),i.splice(12,i.length-12))}this.minStackPos=i[0].pos;for(let o=1;o<i.length;o++)i[o].pos<this.minStackPos&&(this.minStackPos=i[o].pos);return null}stopAt(t){if(this.stoppedAt!=null&&this.stoppedAt<t)throw new RangeError("Can't move stoppedAt forward");this.stoppedAt=t}advanceStack(t,e,i){let n=t.pos,{parser:r}=this,o=ft?this.stackID(t)+" -> ":"";if(this.stoppedAt!=null&&n>this.stoppedAt)return t.forceReduce()?t:null;if(this.fragments){let h=t.curContext&&t.curContext.tracker.strict,c=h?t.curContext.hash:0;for(let f=this.fragments.nodeAt(n);f;){let u=this.parser.nodeSet.types[f.type.id]==f.type?r.getGoto(t.state,f.type.id):-1;if(u>-1&&f.length&&(!h||(f.prop(D.contextHash)||0)==c))return t.useNode(f,u),ft&&console.log(o+this.stackID(t)+` (via reuse of ${r.getName(f.type.id)})`),!0;if(!(f instanceof U)||f.children.length==0||f.positions[0]>0)break;let d=f.children[0];if(d instanceof U&&f.positions[0]==0)f=d;else break}}let l=r.stateSlot(t.state,4);if(l>0)return t.reduce(l),ft&&console.log(o+this.stackID(t)+` (via always-reduce ${r.getName(l&65535)})`),!0;if(t.stack.length>=8400)for(;t.stack.length>6e3&&t.forceReduce(););let a=this.tokens.getActions(t);for(let h=0;h<a.length;){let c=a[h++],f=a[h++],u=a[h++],d=h==a.length||!i,p=d?t:t.split(),g=this.tokens.mainToken;if(p.apply(c,f,g?g.start:p.pos,u),ft&&console.log(o+this.stackID(p)+` (via ${(c&65536)==0?"shift":`reduce of ${r.getName(c&65535)}`} for ${r.getName(f)} @ ${n}${p==t?"":", split"})`),d)return!0;p.pos>n?e.push(p):i.push(p)}return!1}advanceFully(t,e){let i=t.pos;for(;;){if(!this.advanceStack(t,null,null))return!1;if(t.pos>i)return ic(t,e),!0}}runRecovery(t,e,i){let n=null,r=!1;for(let o=0;o<t.length;o++){let l=t[o],a=e[o<<1],h=e[(o<<1)+1],c=ft?this.stackID(l)+" -> ":"";if(l.deadEnd&&(r||(r=!0,l.restart(),ft&&console.log(c+this.stackID(l)+" (restarted)"),this.advanceFully(l,i))))continue;let f=l.split(),u=c;for(let d=0;d<10&&f.forceReduce()&&(ft&&console.log(u+this.stackID(f)+" (via force-reduce)"),!this.advanceFully(f,i));d++)ft&&(u=this.stackID(f)+" -> ");for(let d of l.recoverByInsert(a))ft&&console.log(c+this.stackID(d)+" (via recover-insert)"),this.advanceFully(d,i);this.stream.end>l.pos?(h==l.pos&&(h++,a=0),l.recoverByDelete(a,h),ft&&console.log(c+this.stackID(l)+` (via recover-delete ${this.parser.getName(a)})`),ic(l,i)):(!n||n.score<f.score)&&(n=f)}return n}stackToTree(t){return t.close(),U.build({buffer:ro.create(t),nodeSet:this.parser.nodeSet,topID:this.topTerm,maxBufferLength:this.parser.bufferLength,reused:this.reused,start:this.ranges[0].from,length:t.pos-this.ranges[0].from,minRepeatType:this.parser.minRepeatTerm})}stackID(t){let e=(eo||(eo=new WeakMap)).get(t);return e||eo.set(t,e=String.fromCodePoint(this.nextStackID++)),e+t}};function ic(s,t){for(let e=0;e<t.length;e++){let i=t[e];if(i.pos==s.pos&&i.sameState(s)){t[e].score<s.score&&(t[e]=s);return}}t.push(s)}var fo=class{constructor(t,e,i){this.source=t,this.flags=e,this.disabled=i}allows(t){return!this.disabled||this.disabled[t]==0}},io=s=>s,qs=class{constructor(t){this.start=t.start,this.shift=t.shift||io,this.reduce=t.reduce||io,this.reuse=t.reuse||io,this.hash=t.hash||(()=>0),this.strict=t.strict!==!1}},ei=class s extends Ue{constructor(t){if(super(),this.wrappers=[],t.version!=14)throw new RangeError(`Parser version (${t.version}) doesn't match runtime version (14)`);let e=t.nodeNames.split(" ");this.minRepeatTerm=e.length;for(let l=0;l<t.repeatNodeCount;l++)e.push("");let i=Object.keys(t.topRules).map(l=>t.topRules[l][1]),n=[];for(let l=0;l<e.length;l++)n.push([]);function r(l,a,h){n[l].push([a,a.deserialize(String(h))])}if(t.nodeProps)for(let l of t.nodeProps){let a=l[0];typeof a=="string"&&(a=D[a]);for(let h=1;h<l.length;){let c=l[h++];if(c>=0)r(c,a,l[h++]);else{let f=l[h+-c];for(let u=-c;u>0;u--)r(l[h++],a,f);h++}}}this.nodeSet=new Pi(e.map((l,a)=>tt.define({name:a>=this.minRepeatTerm?void 0:l,id:a,props:n[a],top:i.indexOf(a)>-1,error:a==0,skipped:t.skippedNodes&&t.skippedNodes.indexOf(a)>-1}))),t.propSources&&(this.nodeSet=this.nodeSet.extend(...t.propSources)),this.strict=!1,this.bufferLength=1024;let o=Fi(t.tokenData);this.context=t.context,this.specializerSpecs=t.specialized||[],this.specialized=new Uint16Array(this.specializerSpecs.length);for(let l=0;l<this.specializerSpecs.length;l++)this.specialized[l]=this.specializerSpecs[l].term;this.specializers=this.specializerSpecs.map(sc),this.states=Fi(t.states,Uint32Array),this.data=Fi(t.stateData),this.goto=Fi(t.goto),this.maxTerm=t.maxTerm,this.tokenizers=t.tokenizers.map(l=>typeof l=="number"?new oe(o,l):l),this.topRules=t.topRules,this.dialects=t.dialects||{},this.dynamicPrecedences=t.dynamicPrecedences||null,this.tokenPrecTable=t.tokenPrec,this.termNames=t.termNames||null,this.maxNode=this.nodeSet.types.length-1,this.dialect=this.parseDialect(),this.top=this.topRules[Object.keys(this.topRules)[0]]}createParse(t,e,i){let n=new co(this,t,e,i);for(let r of this.wrappers)n=r(n,t,e,i);return n}getGoto(t,e,i=!1){let n=this.goto;if(e>=n[0])return-1;for(let r=n[e+1];;){let o=n[r++],l=o&1,a=n[r++];if(l&&i)return a;for(let h=r+(o>>1);r<h;r++)if(n[r]==t)return a;if(l)return-1}}hasAction(t,e){let i=this.data;for(let n=0;n<2;n++)for(let r=this.stateSlot(t,n?2:1),o;;r+=3){if((o=i[r])==65535)if(i[r+1]==1)o=i[r=qt(i,r+2)];else{if(i[r+1]==2)return qt(i,r+2);break}if(o==e||o==0)return qt(i,r+1)}return 0}stateSlot(t,e){return this.states[t*6+e]}stateFlag(t,e){return(this.stateSlot(t,0)&e)>0}validAction(t,e){return!!this.allActions(t,i=>i==e?!0:null)}allActions(t,e){let i=this.stateSlot(t,4),n=i?e(i):void 0;for(let r=this.stateSlot(t,1);n==null;r+=3){if(this.data[r]==65535)if(this.data[r+1]==1)r=qt(this.data,r+2);else break;n=e(qt(this.data,r+1))}return n}nextStates(t){let e=[];for(let i=this.stateSlot(t,1);;i+=3){if(this.data[i]==65535)if(this.data[i+1]==1)i=qt(this.data,i+2);else break;if((this.data[i+2]&1)==0){let n=this.data[i+1];e.some((r,o)=>o&1&&r==n)||e.push(this.data[i],n)}}return e}configure(t){let e=Object.assign(Object.create(s.prototype),this);if(t.props&&(e.nodeSet=this.nodeSet.extend(...t.props)),t.top){let i=this.topRules[t.top];if(!i)throw new RangeError(`Invalid top rule name ${t.top}`);e.top=i}return t.tokenizers&&(e.tokenizers=this.tokenizers.map(i=>{let n=t.tokenizers.find(r=>r.from==i);return n?n.to:i})),t.specializers&&(e.specializers=this.specializers.slice(),e.specializerSpecs=this.specializerSpecs.map((i,n)=>{let r=t.specializers.find(l=>l.from==i.external);if(!r)return i;let o=Object.assign(Object.assign({},i),{external:r.to});return e.specializers[n]=sc(o),o})),t.contextTracker&&(e.context=t.contextTracker),t.dialect&&(e.dialect=this.parseDialect(t.dialect)),t.strict!=null&&(e.strict=t.strict),t.wrap&&(e.wrappers=e.wrappers.concat(t.wrap)),t.bufferLength!=null&&(e.bufferLength=t.bufferLength),e}hasWrappers(){return this.wrappers.length>0}getName(t){return this.termNames?this.termNames[t]:String(t<=this.maxNode&&this.nodeSet.types[t].name||t)}get eofTerm(){return this.maxNode+1}get topNode(){return this.nodeSet.types[this.top[1]]}dynamicPrecedence(t){let e=this.dynamicPrecedences;return e==null?0:e[t]||0}parseDialect(t){let e=Object.keys(this.dialects),i=e.map(()=>!1);if(t)for(let r of t.split(" ")){let o=e.indexOf(r);o>=0&&(i[o]=!0)}let n=null;for(let r=0;r<e.length;r++)if(!i[r])for(let o=this.dialects[e[r]],l;(l=this.data[o++])!=65535;)(n||(n=new Uint8Array(this.maxTerm+1)))[l]=1;return new fo(t,i,n)}static deserialize(t){return new s(t)}};function qt(s,t){return s[t]|s[t+1]<<16}function mp(s){let t=null;for(let e of s){let i=e.p.stoppedAt;(e.pos==e.p.stream.end||i!=null&&e.pos>i)&&e.p.parser.stateFlag(e.state,2)&&(!t||t.score<e.score)&&(t=e)}return t}function sc(s){if(s.external){let t=s.extend?1:0;return(e,i)=>s.external(e,i)<<1|t}return s.get}var ii=63,rc=64,yp=1,bp=2,ac=3,wp=4,hc=5,kp=6,xp=7,cc=65,vp=66,Sp=8,Op=9,Cp=10,Tp=11,Ap=12,fc=13,Pp=19,Mp=20,Dp=29,Rp=33,Ep=34,Bp=47,Lp=0,bo=1,po=2,$i=3,go=4,Ut=class{constructor(t,e,i){this.parent=t,this.depth=e,this.type=i,this.hash=(t?t.hash+t.hash<<8:0)+e+(e<<4)+i}};Ut.top=new Ut(null,-1,Lp);function Wi(s,t){for(let e=0,i=t-s.pos-1;;i--,e++){let n=s.peek(i);if(Gt(n)||n==-1)return e}}function mo(s){return s==32||s==9}function Gt(s){return s==10||s==13}function uc(s){return mo(s)||Gt(s)}function Me(s){return s<0||uc(s)}var Ip=new qs({start:Ut.top,reduce(s,t){return s.type==$i&&(t==Mp||t==Ep)?s.parent:s},shift(s,t,e,i){if(t==ac)return new Ut(s,Wi(i,i.pos),bo);if(t==cc||t==hc)return new Ut(s,Wi(i,i.pos),po);if(t==ii)return s.parent;if(t==Pp||t==Rp)return new Ut(s,0,$i);if(t==fc&&s.type==go)return s.parent;if(t==Bp){let n=/[1-9]/.exec(i.read(i.pos,e.pos));if(n)return new Ut(s,s.depth+ +n[0],go)}return s},hash(s){return s.hash}});function si(s,t,e=0){return s.peek(e)==t&&s.peek(e+1)==t&&s.peek(e+2)==t&&Me(s.peek(e+3))}var Np=new Pe((s,t)=>{if(s.next==-1&&t.canShift(rc))return s.acceptToken(rc);let e=s.peek(-1);if((Gt(e)||e<0)&&t.context.type!=$i){if(si(s,45))if(t.canShift(ii))s.acceptToken(ii);else return s.acceptToken(yp,3);if(si(s,46))if(t.canShift(ii))s.acceptToken(ii);else return s.acceptToken(bp,3);let i=0;for(;s.next==32;)i++,s.advance();(i<t.context.depth||i==t.context.depth&&t.context.type==bo&&(s.next!=45||!Me(s.peek(1))))&&s.next!=-1&&!Gt(s.next)&&s.next!=35&&s.acceptToken(ii,-i)}},{contextual:!0}),_p=new Pe((s,t)=>{if(t.context.type==$i){s.next==63&&(s.advance(),Me(s.next)&&s.acceptToken(xp));return}if(s.next==45)s.advance(),Me(s.next)&&s.acceptToken(t.context.type==bo&&t.context.depth==Wi(s,s.pos-1)?wp:ac);else if(s.next==63)s.advance(),Me(s.next)&&s.acceptToken(t.context.type==po&&t.context.depth==Wi(s,s.pos-1)?kp:hc);else{let e=s.pos;for(;;)if(mo(s.next)){if(s.pos==e)return;s.advance()}else if(s.next==33)dc(s);else if(s.next==38)yo(s);else if(s.next==42){yo(s);break}else if(s.next==39||s.next==34){if(wo(s,!0))break;return}else if(s.next==91||s.next==123){if(!Hp(s))return;break}else{pc(s,!0,!1,0);break}for(;mo(s.next);)s.advance();if(s.next==58){if(s.pos==e&&t.canShift(Dp))return;let i=s.peek(1);Me(i)&&s.acceptTokenTo(t.context.type==po&&t.context.depth==Wi(s,e)?vp:cc,e)}}},{contextual:!0});function Vp(s){return s>32&&s<127&&s!=34&&s!=37&&s!=44&&s!=60&&s!=62&&s!=92&&s!=94&&s!=96&&s!=123&&s!=124&&s!=125}function oc(s){return s>=48&&s<=57||s>=97&&s<=102||s>=65&&s<=70}function lc(s,t){return s.next==37?(s.advance(),oc(s.next)&&s.advance(),oc(s.next)&&s.advance(),!0):Vp(s.next)||t&&s.next==44?(s.advance(),!0):!1}function dc(s){if(s.advance(),s.next==60){for(s.advance();;)if(!lc(s,!0)){s.next==62&&s.advance();break}}else for(;lc(s,!1););}function yo(s){for(s.advance();!Me(s.next)&&Us(s.next)!="f";)s.advance()}function wo(s,t){let e=s.next,i=!1,n=s.pos;for(s.advance();;){let r=s.next;if(r<0)break;if(s.advance(),r==e)if(r==39)if(s.next==39)s.advance();else break;else break;else if(r==92&&e==34)s.next>=0&&s.advance();else if(Gt(r)){if(t)return!1;i=!0}else if(t&&s.pos>=n+1024)return!1}return!i}function Hp(s){for(let t=[],e=s.pos+1024;;)if(s.next==91||s.next==123)t.push(s.next),s.advance();else if(s.next==39||s.next==34){if(!wo(s,!0))return!1}else if(s.next==93||s.next==125){if(t[t.length-1]!=s.next-2)return!1;if(t.pop(),s.advance(),!t.length)return!0}else{if(s.next<0||s.pos>e||Gt(s.next))return!1;s.advance()}}var Fp="iiisiiissisfissssssssssssisssiiissssssssssssssssssssssssssfsfssissssssssssssssssssssssssssfif";function Us(s){return s<33?"u":s>125?"s":Fp[s-33]}function uo(s,t){let e=Us(s);return e!="u"&&!(t&&e=="f")}function pc(s,t,e,i){if(Us(s.next)=="s"||(s.next==63||s.next==58||s.next==45)&&uo(s.peek(1),e))s.advance();else return!1;let n=s.pos;for(;;){let r=s.next,o=0,l=i+1;for(;uc(r);){if(Gt(r)){if(t)return!1;l=0}else l++;r=s.peek(++o)}if(!(r>=0&&(r==58?uo(s.peek(o+1),e):r==35?s.peek(o-1)!=32:uo(r,e)))||!e&&l<=i||l==0&&!e&&(si(s,45,o)||si(s,46,o)))break;if(t&&Us(r)=="f")return!1;for(let h=o;h>=0;h--)s.advance();if(t&&s.pos>n+1024)return!1}return!0}var Wp=new Pe((s,t)=>{if(s.next==33)dc(s),s.acceptToken(Ap);else if(s.next==38||s.next==42){let e=s.next==38?Cp:Tp;yo(s),s.acceptToken(e)}else s.next==39||s.next==34?(wo(s,!1),s.acceptToken(Op)):pc(s,!1,t.context.type==$i,t.context.depth)&&s.acceptToken(Sp)}),$p=new Pe((s,t)=>{let e=t.context.type==go?t.context.depth:-1,i=s.pos;t:for(;;){let n=0,r=s.next;for(;r==32;)r=s.peek(++n);if(!n&&(si(s,45,n)||si(s,46,n))||!Gt(r)&&(e<0&&(e=Math.max(t.context.depth+1,n)),n<e))break;for(;;){if(s.next<0)break t;let o=Gt(s.next);if(s.advance(),o)continue t;i=s.pos}}s.acceptTokenTo(fc,i)}),zp=Ge({DirectiveName:m.keyword,DirectiveContent:m.attributeValue,"DirectiveEnd DocEnd":m.meta,QuotedLiteral:m.string,BlockLiteralHeader:m.special(m.string),BlockLiteralContent:m.content,Literal:m.content,"Key/Literal Key/QuotedLiteral":m.definition(m.propertyName),"Anchor Alias":m.labelName,Tag:m.typeName,Comment:m.lineComment,": , -":m.separator,"?":m.punctuation,"[ ]":m.squareBracket,"{ }":m.brace}),gc=ei.deserialize({version:14,states:"5lQ!ZQgOOO#PQfO'#CpO#uQfO'#DOOOQR'#Dv'#DvO$qQgO'#DRO%gQdO'#DUO%nQgO'#DUO&ROaO'#D[OOQR'#Du'#DuO&{QgO'#D^O'rQgO'#D`OOQR'#Dt'#DtO(iOqO'#DbOOQP'#Dj'#DjO(zQaO'#CmO)YQgO'#CmOOQP'#Cm'#CmQ)jQaOOQ)uQgOOQ]QgOOO*PQdO'#CrO*nQdO'#CtOOQO'#Dw'#DwO+]Q`O'#CxO+hQdO'#CwO+rQ`O'#CwOOQO'#Cv'#CvO+wQdO'#CvOOQO'#Cq'#CqO,UQ`O,59[O,^QfO,59[OOQR,59[,59[OOQO'#Cx'#CxO,eQ`O'#DPO,pQdO'#DPOOQO'#Dx'#DxO,zQdO'#DxO-XQ`O,59jO-aQfO,59jOOQR,59j,59jOOQR'#DS'#DSO-hQcO,59mO-sQgO'#DVO.TQ`O'#DVO.YQcO,59pOOQR'#DX'#DXO#|QfO'#DWO.hQcO'#DWOOQR,59v,59vO.yOWO,59vO/OOaO,59vO/WOaO,59vO/cQgO'#D_OOQR,59x,59xO0VQgO'#DaOOQR,59z,59zOOQP,59|,59|O0yOaO,59|O1ROaO,59|O1aOqO,59|OOQP-E7h-E7hO1oQgO,59XOOQP,59X,59XO2PQaO'#DeO2_QgO'#DeO2oQgO'#DkOOQP'#Dk'#DkQ)jQaOOO3PQdO'#CsOOQO,59^,59^O3kQdO'#CuOOQO,59`,59`OOQO,59c,59cO4VQdO,59cO4aQdO'#CzO4kQ`O'#CzOOQO,59b,59bOOQU,5:Q,5:QOOQR1G.v1G.vO4pQ`O1G.vOOQU-E7d-E7dO4xQdO,59kOOQO,59k,59kO5SQdO'#DQO5^Q`O'#DQOOQO,5:d,5:dOOQU,5:R,5:ROOQR1G/U1G/UO5cQ`O1G/UOOQU-E7e-E7eO5kQgO'#DhO5xQcO1G/XOOQR1G/X1G/XOOQR,59q,59qO6TQgO,59qO6eQdO'#DiO6lQgO'#DiO7PQcO1G/[OOQR1G/[1G/[OOQR,59r,59rO#|QfO,59rOOQR1G/b1G/bO7_OWO1G/bO7dOaO1G/bOOQR,59y,59yOOQR,59{,59{OOQP1G/h1G/hO7lOaO1G/hO7tOaO1G/hO8POaO1G/hOOQP1G.s1G.sO8_QgO,5:POOQP,5:P,5:POOQP,5:V,5:VOOQP-E7i-E7iOOQO,59_,59_OOQO,59a,59aOOQO1G.}1G.}OOQO,59f,59fO8oQdO,59fOOQR7+$b7+$bP,XQ`O'#DfOOQO1G/V1G/VOOQO,59l,59lO8yQdO,59lOOQR7+$p7+$pP9TQ`O'#DgOOQR'#DT'#DTOOQR,5:S,5:SOOQR-E7f-E7fOOQR7+$s7+$sOOQR1G/]1G/]O9YQgO'#DYO9jQ`O'#DYOOQR,5:T,5:TO#|QfO'#DZO9oQcO'#DZOOQR-E7g-E7gOOQR7+$v7+$vOOQR1G/^1G/^OOQR7+$|7+$|O:QOWO7+$|OOQP7+%S7+%SO:VOaO7+%SO:_OaO7+%SOOQP1G/k1G/kOOQO1G/Q1G/QOOQO1G/W1G/WOOQR,59t,59tO:jQgO,59tOOQR,59u,59uO#|QfO,59uOOQR<<Hh<<HhOOQP<<Hn<<HnO:zOaO<<HnOOQR1G/`1G/`OOQR1G/a1G/aOOQPAN>YAN>Y",stateData:";S~O!fOS!gOS^OS~OP_OQbORSOTUOWROXROYYOZZO[XOcPOqQO!PVO!V[O!cTO~O`cO~P]OVkOWROXROYeOZfO[dOcPOmhOqQO~OboO~P!bOVtOWROXROYeOZfO[dOcPOmrOqQO~OpwO~P#WORSOTUOWROXROYYOZZO[XOcPOqQO!PVO!cTO~OSvP!avP!bvP~P#|OWROXROYeOZfO[dOcPOqQO~OmzO~P%OOm!OOUzP!azP!bzP!dzP~P#|O^!SO!b!QO!f!TO!g!RO~ORSOTUOWROXROcPOqQO!PVO!cTO~OY!UOP!QXQ!QX!V!QX!`!QXS!QX!a!QX!b!QXU!QXm!QX!d!QX~P&aO[!WOP!SXQ!SX!V!SX!`!SXS!SX!a!SX!b!SXU!SXm!SX!d!SX~P&aO^!ZO!W![O!b!YO!f!]O!g!YO~OP!_O!V[OQaX!`aX~OPaXQaX!VaX!`aX~P#|OP!bOQ!cO!V[O~OP_O!V[O~P#|OWROXROY!fOcPOqQObfXmfXofXpfX~OWROXRO[!hOcPOqQObhXmhXohXphX~ObeXmlXoeX~ObkXokX~P%OOm!kO~Om!lObnPonP~P%OOb!pOo!oO~Ob!pO~P!bOm!sOosXpsX~OosXpsX~P%OOm!uOotPptP~P%OOo!xOp!yO~Op!yO~P#WOS!|O!a#OO!b#OO~OUyX!ayX!byX!dyX~P#|Om#QO~OU#SO!a#UO!b#UO!d#RO~Om#WOUzX!azX!bzX!dzX~O]#XO~O!b#XO!g#YO~O^#ZO!b#XO!g#YO~OP!RXQ!RX!V!RX!`!RXS!RX!a!RX!b!RXU!RXm!RX!d!RX~P&aOP!TXQ!TX!V!TX!`!TXS!TX!a!TX!b!TXU!TXm!TX!d!TX~P&aO!b#^O!g#^O~O^#_O!b#^O!f#`O!g#^O~O^#_O!W#aO!b#^O!g#^O~OPaaQaa!Vaa!`aa~P#|OP#cO!V[OQ!XX!`!XX~OP!XXQ!XX!V!XX!`!XX~P#|OP_O!V[OQ!_X!`!_X~P#|OWROXROcPOqQObgXmgXogXpgX~OWROXROcPOqQObiXmiXoiXpiX~Obkaoka~P%OObnXonX~P%OOm#kO~Ob#lOo!oO~Oosapsa~P%OOotXptX~P%OOm#pO~Oo!xOp#qO~OSwP!awP!bwP~P#|OS!|O!a#vO!b#vO~OUya!aya!bya!dya~P#|Om#xO~P%OOm#{OU}P!a}P!b}P!d}P~P#|OU#SO!a$OO!b$OO!d#RO~O]$QO~O!b$QO!g$RO~O!b$SO!g$SO~O^$TO!b$SO!g$SO~O^$TO!b$SO!f$UO!g$SO~OP!XaQ!Xa!V!Xa!`!Xa~P#|Obnaona~P%OOotapta~P%OOo!xO~OU|X!a|X!b|X!d|X~P#|Om$ZO~Om$]OU}X!a}X!b}X!d}X~O]$^O~O!b$_O!g$_O~O^$`O!b$_O!g$_O~OU|a!a|a!b|a!d|a~P#|O!b$cO!g$cO~O",goto:",]!mPPPPPPPPPPPPPPPPP!nPP!v#v#|$`#|$c$f$j$nP%VPPP!v%Y%^%a%{&O%a&R&U&X&_&b%aP&e&{&e'O'RPP']'a'g'm's'y(XPPPPPPPP(_)e*X+c,VUaObcR#e!c!{ROPQSTUXY_bcdehknrtvz!O!U!W!_!b!c!f!h!k!l!s!u!|#Q#R#S#W#c#k#p#x#{$Z$]QmPR!qnqfPQThknrtv!k!l!s!u#R#k#pR!gdR!ieTlPnTjPnSiPnSqQvQ{TQ!mkQ!trQ!vtR#y#RR!nkTsQvR!wt!RWOSUXY_bcz!O!U!W!_!b!c!|#Q#S#W#c#x#{$Z$]RySR#t!|R|TR|UQ!PUR#|#SR#z#RR#z#SyZOSU_bcz!O!_!b!c!|#Q#S#W#c#x#{$Z$]R!VXR!XYa]O^abc!a!c!eT!da!eQnPR!rnQvQR!{vQ!}yR#u!}Q#T|R#}#TW^Obc!cS!^^!aT!aa!eQ!eaR#f!eW`Obc!cQxSS}U#SQ!`_Q#PzQ#V!OQ#b!_Q#d!bQ#s!|Q#w#QQ$P#WQ$V#cQ$Y#xQ$[#{Q$a$ZR$b$]xZOSU_bcz!O!_!b!c!|#Q#S#W#c#x#{$Z$]Q!VXQ!XYQ#[!UR#]!W!QWOSUXY_bcz!O!U!W!_!b!c!|#Q#S#W#c#x#{$Z$]pfPQThknrtv!k!l!s!u#R#k#pQ!gdQ!ieQ#g!fR#h!hSgPn^pQTkrtv#RQ!jhQ#i!kQ#j!lQ#n!sQ#o!uQ$W#kR$X#pQuQR!zv",nodeNames:"\u26A0 DirectiveEnd DocEnd - - ? ? ? Literal QuotedLiteral Anchor Alias Tag BlockLiteralContent Comment Stream BOM Document ] [ FlowSequence Item Tagged Anchored Anchored Tagged FlowMapping Pair Key : Pair , } { FlowMapping Pair Pair BlockSequence Item Item BlockMapping Pair Pair Key Pair Pair BlockLiteral BlockLiteralHeader Tagged Anchored Anchored Tagged Directive DirectiveName DirectiveContent Document",maxTerm:74,context:Ip,nodeProps:[["isolate",-3,8,9,14,""],["openedBy",18,"[",32,"{"],["closedBy",19,"]",33,"}"]],propSources:[zp],skippedNodes:[0],repeatNodeCount:6,tokenData:"-Y~RnOX#PXY$QYZ$]Z]#P]^$]^p#Ppq$Qqs#Pst$btu#Puv$yv|#P|}&e}![#P![!]'O!]!`#P!`!a'i!a!}#P!}#O*g#O#P#P#P#Q+Q#Q#o#P#o#p+k#p#q'i#q#r,U#r;'S#P;'S;=`#z<%l?HT#P?HT?HU,o?HUO#PQ#UU!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PQ#kTOY#PZs#Pt;'S#P;'S;=`#z<%lO#PQ#}P;=`<%l#P~$VQ!f~XY$Qpq$Q~$bO!g~~$gS^~OY$bZ;'S$b;'S;=`$s<%lO$b~$vP;=`<%l$bR%OX!WQOX%kXY#PZ]%k]^#P^p%kpq#hq;'S%k;'S;=`&_<%lO%kR%rX!WQ!VPOX%kXY#PZ]%k]^#P^p%kpq#hq;'S%k;'S;=`&_<%lO%kR&bP;=`<%l%kR&lUoP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR'VUmP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR'p[!PP!WQOY#PZp#Ppq#hq{#P{|(f|}#P}!O(f!O!R#P!R![)p![;'S#P;'S;=`#z<%lO#PR(mW!PP!WQOY#PZp#Ppq#hq!R#P!R![)V![;'S#P;'S;=`#z<%lO#PR)^U!PP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR)wY!PP!WQOY#PZp#Ppq#hq{#P{|)V|}#P}!O)V!O;'S#P;'S;=`#z<%lO#PR*nUcP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR+XUbP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR+rUqP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR,]UpP!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#PR,vU`P!WQOY#PZp#Ppq#hq;'S#P;'S;=`#z<%lO#P",tokenizers:[Np,_p,Wp,$p,0,1],topRules:{Stream:[0,15]},tokenPrec:0});var Qp=ei.deserialize({version:14,states:"!vOQOPOOO]OPO'#C_OhOPO'#C^OOOO'#Cc'#CcOpOPO'#CaQOOOOOO{OPOOOOOO'#Cb'#CbO!WOPO'#C`O!`OPO,58xOOOO-E6a-E6aOOOO-E6`-E6`OOOO'#C_'#C_OOOO1G.d1G.d",stateData:"!h~OXPOYROWTP~OWVXXRXYRX~OYVOXSP~OXROYROWTX~OXROYROWTP~OYVOXSX~OX[O~OXY~",goto:"vWPPX[beioRUOQQOR]XRXQTTOUQWQRZWSSOURYS",nodeNames:"\u26A0 Document Frontmatter DashLine FrontmatterContent Body",maxTerm:10,skippedNodes:[0],repeatNodeCount:2,tokenData:"$z~RXOYnYZ!^Z]n]^!^^}n}!O!i!O;'Sn;'S;=`!c<%lOn~qXOYnYZ!^Z]n]^!^^;'Sn;'S;=`!c<%l~n~On~~!^~!cOY~~!fP;=`<%ln~!lZOYnYZ!^Z]n]^!^^}n}!O#_!O;'Sn;'S;=`!c<%l~n~On~~!^~#bZOYnYZ!^Z]n]^!^^}n}!O$T!O;'Sn;'S;=`!c<%l~n~On~~!^~$WXOYnYZ$sZ]n]^$s^;'Sn;'S;=`!c<%l~n~On~~$s~$zOX~Y~",tokenizers:[0],topRules:{Document:[0,1]},tokenPrec:67}),jp=Li.define({name:"yaml",parser:gc.configure({props:[jr.add({Stream:s=>{for(let t=s.node.resolve(s.pos,-1);t&&t.to>=s.pos;t=t.parent){if(t.name=="BlockLiteralContent"&&t.from<t.to)return s.baseIndentFor(t);if(t.name=="BlockLiteral")return s.baseIndentFor(t)+s.unit;if(t.name=="BlockSequence"||t.name=="BlockMapping")return s.column(t.from,1);if(t.name=="QuotedLiteral")return null;if(t.name=="Literal"){let e=s.column(t.from,1);if(e==s.lineIndent(t.from,1))return e;if(t.to>s.pos)return null}}return null},FlowMapping:Xr({closing:"}"}),FlowSequence:Xr({closing:"]"})}),lh.add({"FlowMapping FlowSequence":ah,"Item Pair BlockLiteral":(s,t)=>({from:t.doc.lineAt(s.from).to,to:s.to})})]}),languageData:{commentTokens:{line:"#"},indentOnInput:/^\s*[\]\}]$/}});function mc(){return new Vs(jp)}var hm=Li.define({name:"yaml-frontmatter",parser:Qp.configure({props:[Ge({DashLine:m.meta})]})});var Xp="#e5c07b",yc="#e06c75",qp="#56b6c2",Up="#ffffff",Gs="#abb2bf",xo="#7d8799",Gp="#61afef",Kp="#98c379",bc="#d19a66",Yp="#c678dd",Jp="#21252b",wc="#2c313a",kc="#282c34",ko="#353a42",Zp="#3E4451",xc="#528bff";var tg=A.theme({"&":{color:Gs,backgroundColor:kc},".cm-content":{caretColor:xc},".cm-cursor, .cm-dropCursor":{borderLeftColor:xc},"&.cm-focused > .cm-scroller > .cm-selectionLayer .cm-selectionBackground, .cm-selectionBackground, .cm-content ::selection":{backgroundColor:Zp},".cm-panels":{backgroundColor:Jp,color:Gs},".cm-panels.cm-panels-top":{borderBottom:"2px solid black"},".cm-panels.cm-panels-bottom":{borderTop:"2px solid black"},".cm-searchMatch":{backgroundColor:"#72a1ff59",outline:"1px solid #457dff"},".cm-searchMatch.cm-searchMatch-selected":{backgroundColor:"#6199ff2f"},".cm-activeLine":{backgroundColor:"#6699ff0b"},".cm-selectionMatch":{backgroundColor:"#aafe661a"},"&.cm-focused .cm-matchingBracket, &.cm-focused .cm-nonmatchingBracket":{backgroundColor:"#bad0f847"},".cm-gutters":{backgroundColor:kc,color:xo,border:"none"},".cm-activeLineGutter":{backgroundColor:wc},".cm-foldPlaceholder":{backgroundColor:"transparent",border:"none",color:"#ddd"},".cm-tooltip":{border:"none",backgroundColor:ko},".cm-tooltip .cm-tooltip-arrow:before":{borderTopColor:"transparent",borderBottomColor:"transparent"},".cm-tooltip .cm-tooltip-arrow:after":{borderTopColor:ko,borderBottomColor:ko},".cm-tooltip-autocomplete":{"& > ul > li[aria-selected]":{backgroundColor:wc,color:Gs}}},{dark:!0}),eg=Je.define([{tag:m.keyword,color:Yp},{tag:[m.name,m.deleted,m.character,m.propertyName,m.macroName],color:yc},{tag:[m.function(m.variableName),m.labelName],color:Gp},{tag:[m.color,m.constant(m.name),m.standard(m.name)],color:bc},{tag:[m.definition(m.name),m.separator],color:Gs},{tag:[m.typeName,m.className,m.number,m.changed,m.annotation,m.modifier,m.self,m.namespace],color:Xp},{tag:[m.operator,m.operatorKeyword,m.url,m.escape,m.regexp,m.link,m.special(m.string)],color:qp},{tag:[m.meta,m.comment],color:xo},{tag:m.strong,fontWeight:"bold"},{tag:m.emphasis,fontStyle:"italic"},{tag:m.strikethrough,textDecoration:"line-through"},{tag:m.link,color:xo,textDecoration:"underline"},{tag:m.heading,fontWeight:"bold",color:yc},{tag:[m.atom,m.bool,m.special(m.variableName)],color:bc},{tag:[m.processingInstruction,m.string,m.inserted],color:Kp},{tag:m.invalid,color:Up}]),vc=[tg,ch(eg)];var ig=600;function Sc(){return{step:"write",text:"",labelPrefix:"",showOptions:!1,focused:!1,submitting:!1,errorMsg:"",result:{},draft_id:"",_sseSource:null,_draftTimeout:null,initiative:"",phaseCount:0,issueCount:0,batchId:"",batchIdCopied:!1,filingProgress:"",filedIssues:[],yamlValid:!0,yamlValidationMsg:"",_validateTimer:null,streamingText:"",loadingMsg:"Amplifying your intelligence\u2026",_loadingMsgs:["Amplifying your intelligence\u2026","Untangling the dependency graph\u2026","The singularity is here\u2026","Parallelising your chaos\u2026","Turning noise into signal\u2026","Your engineers will thank you\u2026","One prompt to rule them all\u2026","Infinite leverage, loading\u2026","Sequencing your work\u2026","Finding the critical path\u2026","Collapsing the wave function\u2026","Reasoning about blast radius\u2026","Negotiating with entropy\u2026","Refactoring reality\u2026","Compiling your ambitions\u2026","Aligning the planets\u2026","Decoding the chaos\u2026","Calculating minimum viable complexity\u2026","Summoning the dependency gods\u2026","Optimising for speed of thought\u2026","Separating signal from noise\u2026","Scheduling the unschedulable\u2026","Making the implicit explicit\u2026","Turning caffeine into architecture\u2026","Mapping the unknown unknowns\u2026","Prioritising ruthlessly\u2026","Preparing your engineers for glory\u2026","Thinking ten steps ahead\u2026","Converting entropy into momentum\u2026","Eliminating the impossible\u2026"],_loadingTimer:null,_editor:null,init(){this._rotateMsgs()},_rotateMsgs(){let s=[...this._loadingMsgs];for(let e=s.length-1;e>0;e--){let i=Math.floor(Math.random()*(e+1));[s[e],s[i]]=[s[i],s[e]]}let t=0;this.loadingMsg=s[0]??"",this._loadingTimer=setInterval(()=>{t=(t+1)%s.length,this.loadingMsg=s[t]??""},4e3)},autoGrow(s){s.style.height="auto",s.style.height=Math.min(s.scrollHeight,520)+"px"},async pasteClipboard(){try{let s=await navigator.clipboard.readText();this.text=(this.text?this.text+`
 `:"")+s,await this.$nextTick(),this.autoGrow(this.$refs.textarea)}catch{}},appendSeed(s){this.text=(this.text.trim()?this.text.trim()+`
 `:"")+s,this.$nextTick(()=>this.autoGrow(this.$refs.textarea))},cancel(){this._closeSse(),this.step="write",this.submitting=!1,this.errorMsg=""},_closeSse(){this._draftTimeout!==null&&(clearTimeout(this._draftTimeout),this._draftTimeout=null),this._sseSource!==null&&(this._sseSource.close(),this._sseSource=null)},async submit(){let s=this.text.trim();if(s){this.errorMsg="",this.streamingText="",this.step="generating",this.submitting=!0;try{let t=await fetch("/api/plan/draft",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({text:s})});if(!t.ok){let i=await t.json().catch(()=>({}));throw new Error(i.detail||`HTTP ${t.status}`)}let e=await t.json();this.draft_id=e.draft_id??"",await this._waitForDraftReady()}catch(t){this.errorMsg=t.message,this.step="write"}finally{this.submitting=!1}}},_waitForDraftReady(){return new Promise((s,t)=>{let e=new EventSource("/events");this._sseSource=e,this._draftTimeout=setTimeout(()=>{this._closeSse(),t(new Error("Plan generation timed out after 60 s. Please try again."))},6e4),e.onmessage=i=>{let n;try{n=JSON.parse(i.data)}catch{return}let r=n.plan_draft_events??[];for(let o of r)if(o.draft_id===this.draft_id){if(this._closeSse(),o.event==="plan_draft_ready"){let l=o.yaml_text??"";this.step="review",this.yamlValid=!0,this.yamlValidationMsg="\u2713 Plan ready \u2014 review and edit before launching",this.$nextTick(()=>this._mountEditor(l)),s(void 0)}else t(new Error("Plan generation timed out on the server. Please try again."));return}},e.onerror=()=>{this._closeSse(),t(new Error("SSE connection lost while waiting for plan. Please try again."))}})},editPlan(){this.step="write",this.errorMsg="",this.$nextTick(()=>{this.$refs.textarea&&this.autoGrow(this.$refs.textarea)})},async launch(){let s=this._getEditorValue();if(s.trim()){if(!this.yamlValid){this.errorMsg="Fix the YAML errors before launching.";return}this.errorMsg="",this.filingProgress="Setting up labels\u2026",this.filedIssues=[],this.step="launching",this.submitting=!0;try{let t=await fetch("/api/plan/file-issues",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({yaml_text:s})});if(!t.ok){let o=await t.json().catch(()=>({}));throw new Error(o.detail||`HTTP ${t.status}`)}let e=t.body.getReader(),i=new TextDecoder,n="",r=null;t:for(;;){let{done:o,value:l}=await e.read();if(o)break;n+=i.decode(l,{stream:!0});let a=n.split(`
-`);n=a.pop()??"";for(let h of a){if(!h.startsWith("data: "))continue;let c=h.slice(6).trim();if(!c)continue;let f;try{f=JSON.parse(c)}catch{continue}if(f.t==="start")this.filingProgress=`Filing 0 / ${f.total}\u2026`;else if(f.t==="label")this.filingProgress=f.text;else if(f.t==="issue")this.filingProgress=`Filing ${f.index} / ${f.total} \u2014 ${f.phase}`,this.filedIssues.push({number:f.number,url:f.url,title:f.title,phase:f.phase});else if(f.t!=="blocked"){if(f.t==="done"){r=f;break t}else if(f.t==="error")throw new Error(f.detail||"Filing issues failed.")}}}if(!r)throw new Error("Stream ended without a done event.");this.result={initiative:r.initiative,total:r.total,issues:r.issues??[]},this.batchId=r.batch_id??"",this.initiative=r.initiative??this.initiative;try{localStorage.setItem("ac_active_batch",this.batchId),localStorage.setItem("ac_active_initiative",this.initiative)}catch{}this.step="done"}catch(t){this.errorMsg=t.message,this.step="review"}finally{this.submitting=!1}}},reset(){this._closeSse(),this.step="write",this.text="",this.labelPrefix="",this.showOptions=!1,this.errorMsg="",this.streamingText="",this.draft_id="",this.initiative="",this.phaseCount=0,this.issueCount=0,this.yamlValid=!0,this.yamlValidationMsg="",this.filingProgress="",this.filedIssues=[],this.result={},this.batchId="",this.batchIdCopied=!1,this._editor&&this._setEditorValue("")},async copyBatchId(){if(this.batchId)try{await navigator.clipboard.writeText(this.batchId),this.batchIdCopied=!0,setTimeout(()=>{this.batchIdCopied=!1},1500)}catch{}},async reRun(s){try{let t=await fetch(`/api/plan/${encodeURIComponent(s)}/plan-text`);if(!t.ok){let i=await t.json().catch(()=>({}));this.errorMsg=i.detail||`Could not load run (HTTP ${t.status})`;return}let e=await t.json();this.reset(),this.text=e.plan_text??"",await this.$nextTick(),this.$refs.textarea&&this.autoGrow(this.$refs.textarea)}catch(t){this.errorMsg=t.message||"Failed to load previous run."}},_mountEditor(s){let t=this.$refs.yamlEditor;if(!t)return;if(this._editor){this._setEditorValue(s);return}let e=this,i=A.updateListener.of(n=>{n.docChanged&&(clearTimeout(e._validateTimer),e._validateTimer=setTimeout(()=>e._validateYaml(),ig))});this._editor=new A({state:W.create({doc:s,extensions:[vh(),za(),$a(),xr.of([...Jh,...Th]),mc(),vc,A.lineWrapping,i]}),parent:t})},_getEditorValue(){return this._editor?this._editor.state.doc.toString():""},_setEditorValue(s){this._editor&&this._editor.dispatch({changes:{from:0,to:this._editor.state.doc.length,insert:s},selection:{anchor:0},scrollIntoView:!0})},async _validateYaml(){if(this.step!=="review")return;let s=this._getEditorValue();if(!s.trim()){this.yamlValid=!1,this.yamlValidationMsg="\u26A0 YAML is empty.";return}try{let e=await(await fetch("/api/plan/validate",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({yaml_text:s})})).json();e.valid?(this.yamlValid=!0,this.yamlValidationMsg=`\u2713 Valid \u2014 ${e.phase_count} phases, ${e.issue_count} issues`):(this.yamlValid=!1,this.yamlValidationMsg=`\u2717 ${e.detail||"Invalid PlanSpec"}`)}catch{this.yamlValidationMsg=""}}}}function Oc(s){return{activeIssue:null,events:[],thoughts:[],streamOpen:!1,_evtSource:null,chatMessage:"",chatSending:!1,chatError:null,agentStopping:!1,dispatchOpen:!1,dispatchIssue:null,dispatchRole:"python-developer",roleGroups:s,dispatching:!1,dispatchError:null,dispatchSuccess:!1,dispatchResult:null,labelDispatchOpen:!1,labelDispatchLabel:"",scopeMode:"full_initiative",scopePhases:[],selectedPhase:"",scopeIssues:[],selectedIssueNumber:null,labelContextLoading:!1,labelContextLoaded:!1,showAdvanced:!1,advancedRole:"",labelDispatching:!1,labelDispatchError:null,labelDispatchSuccess:!1,labelDispatchResult:null,dispatcherCopied:!1,get launchPreviewText(){let t=this.labelDispatchLabel;if(this.scopeMode==="full_initiative")return`A ${this.advancedRole.trim()||"coordinator"} will survey every open ticket under "${t}" and assemble its own team.`;if(this.scopeMode==="phase")return this.selectedPhase?`A ${this.advancedRole.trim()||"coordinator"} will handle all tickets in phase "${this.selectedPhase}".`:"Choose a phase to see the preview.";if(this.scopeMode==="issue"){if(!this.selectedIssueNumber)return"Choose a ticket to see the preview.";let e=this.scopeIssues.find(n=>n.number===this.selectedIssueNumber),i=e?e.title:`#${this.selectedIssueNumber}`;return`One leaf agent will work on #${this.selectedIssueNumber}: "${i}".`}return""},get launchReady(){return this.scopeMode==="full_initiative"?!0:this.scopeMode==="phase"?!!this.selectedPhase:this.scopeMode==="issue"?!!this.selectedIssueNumber:!1},get repo(){return window._buildRepo??""},onInspect(t){this.activeIssue?.number!==t.number&&(this._closeStream(),this.activeIssue=t,this.events=[],this.thoughts=[],t.run&&this._openStream(t.run.id))},clearInspect(){this._closeStream(),this.activeIssue=null,this.events=[],this.thoughts=[],this.chatMessage="",this.chatError=null},async sendMessage(){let t=this.chatMessage.trim();if(!t||!this.activeIssue?.run)return;let e=this.activeIssue.run.id;this.chatSending=!0,this.chatError=null;try{let i=await fetch(`/api/build/agent/${encodeURIComponent(e)}/message`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({content:t})});if(i.ok)this.chatMessage="";else{let n=await i.json().catch(()=>({}));this.chatError=n.detail??`Error ${i.status}`}}catch(i){this.chatError=`Network error: ${i.message}`}finally{this.chatSending=!1}},async stopAgent(){if(!this.activeIssue?.run)return;let t=this.activeIssue.run.id;this.agentStopping=!0;try{await fetch(`/api/build/agent/${encodeURIComponent(t)}/stop`,{method:"POST"}),this._closeStream()}catch{}finally{this.agentStopping=!1}},restartAgent(){this.activeIssue&&this.openDispatch(this.activeIssue)},_openStream(t){this._closeStream();let e=new EventSource(`/build/agent/${encodeURIComponent(t)}/stream`);this._evtSource=e,this.streamOpen=!0,e.onmessage=i=>{let n;try{n=JSON.parse(i.data)}catch{return}if(n.t!=="ping"){if(n.t==="event")this.events.push({...n,id:Date.now()+Math.random()});else if(n.t==="thought"){let r=this.thoughts[this.thoughts.length-1];r&&r.role===n.role&&this.thoughts.length>0?r.content+=`
-`+n.content:this.thoughts.push(n),this._scrollCot()}}},e.onerror=()=>{this.streamOpen=!1}},_closeStream(){this._evtSource&&(this._evtSource.close(),this._evtSource=null),this.streamOpen=!1},_scrollCot(){this.$nextTick(()=>{let t=this.$refs.cotScroll;t&&(t.scrollTop=t.scrollHeight)})},openDispatch(t){this.dispatchIssue=t,this.dispatchRole="python-developer",this.dispatchError=null,this.dispatchSuccess=!1,this.dispatchResult=null,this.dispatching=!1,this.dispatchOpen=!0},async submitDispatch(){if(this.dispatchIssue){this.dispatching=!0,this.dispatchError=null;try{let t=await fetch("/api/build/dispatch",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({issue_number:this.dispatchIssue.number,issue_title:this.dispatchIssue.title,role:this.dispatchRole,repo:this.repo})}),e=await t.json();t.ok?(this.dispatchResult=e,this.dispatchSuccess=!0):this.dispatchError=e.detail??`Error ${t.status}`}catch(t){this.dispatchError=`Network error: ${t.message}`}finally{this.dispatching=!1}}},openLabelDispatch(t){this.labelDispatchLabel=t.label??"",this.scopeMode="full_initiative",this.scopePhases=[],this.scopeIssues=[],this.selectedPhase="",this.selectedIssueNumber=null,this.labelContextLoading=!1,this.labelContextLoaded=!1,this.showAdvanced=!1,this.advancedRole="",this.labelDispatchError=null,this.labelDispatchSuccess=!1,this.labelDispatchResult=null,this.labelDispatching=!1,this.dispatcherCopied=!1,this.labelDispatchOpen=!0,this._loadLabelContext()},closeLabelDispatch(){this.labelDispatchOpen=!1},async _loadLabelContext(){if(!(this.labelContextLoaded||this.labelContextLoading)){this.labelContextLoading=!0;try{let t=`/api/build/label-context?label=${encodeURIComponent(this.labelDispatchLabel)}&repo=${encodeURIComponent(this.repo)}`,e=await fetch(t);if(e.ok){let i=await e.json();this.scopePhases=i.phases??[],this.scopeIssues=i.issues??[],this.labelContextLoaded=!0}}catch{}finally{this.labelContextLoading=!1}}},async copyDispatcherPrompt(){try{let t=await fetch("/api/build/dispatcher-prompt");if(!t.ok)throw new Error(`HTTP ${t.status}`);let e=await t.json();await navigator.clipboard.writeText(e.content),this.dispatcherCopied=!0,setTimeout(()=>{this.dispatcherCopied=!1},3e3)}catch(t){alert(`Could not copy prompt: ${t.message}`)}},async submitLabelDispatch(){if(!this.launchReady)return;this.labelDispatching=!0,this.labelDispatchError=null;let t={label:this.labelDispatchLabel,scope:this.scopeMode,repo:this.repo};this.scopeMode==="phase"&&this.selectedPhase&&(t.scope_label=this.selectedPhase),this.scopeMode==="issue"&&this.selectedIssueNumber&&(t.scope_issue_number=this.selectedIssueNumber),this.advancedRole.trim()&&(t.role=this.advancedRole.trim());try{let e=await fetch("/api/build/dispatch-label",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify(t)}),i=await e.json();e.ok?(this.labelDispatchResult=i,this.labelDispatchSuccess=!0):this.labelDispatchError=i.detail??`Error ${e.status}`}catch(e){this.labelDispatchError=`Network error: ${e.message}`}finally{this.labelDispatching=!1}},fmtTime(t){if(!t)return"";try{return new Date(t).toLocaleTimeString([],{hour:"2-digit",minute:"2-digit"})}catch{return t}},eventIcon(t){return{step_start:"\u25B6",blocker:"\u{1F6A7}",decision:"\u{1F4A1}",done:"\u2705"}[t]??"\u2022"},eventDetail(t){let e=t.payload??{};switch(t.event_type){case"step_start":return e.step??"";case"blocker":return e.description??"";case"decision":return`${e.decision??""} \u2014 ${e.rationale??""}`;case"done":return e.summary||e.pr_url||"";default:return JSON.stringify(e)}}}}function Cc(s,t,e){return{q:s||"",filterRole:t||"",filterStatus:e||"",sort:"mtime",applyFilters(){let i=this.q.toLowerCase();document.querySelectorAll("[data-transcript-row]").forEach(n=>{let r=(n.dataset.preview||"").toLowerCase(),o=n.dataset.role||"",l=n.dataset.status||"",a=!i||r.includes(i),h=!this.filterRole||o===this.filterRole,c=!this.filterStatus||l===this.filterStatus;n.style.display=a&&h&&c?"":"none"}),this.applySortDom()},applySortDom(){let i=document.getElementById("transcript-tbody");if(!i)return;let n=Array.from(i.querySelectorAll("[data-transcript-row]"));n.sort((r,o)=>{switch(this.sort){case"messages":return parseInt(o.dataset.messages)-parseInt(r.dataset.messages);case"words":return parseInt(o.dataset.words)-parseInt(r.dataset.words);case"subagents":return parseInt(o.dataset.subagents)-parseInt(r.dataset.subagents);default:return parseFloat(o.dataset.mtime)-parseFloat(r.dataset.mtime)}}),n.forEach(r=>i.appendChild(r))}}}function Tc(){return{search:"",highlight(s){if(!this.search)return this._esc(s);let t=this.search.replace(/[.*+?^${}()|[\]\\]/g,"\\$&"),e=new RegExp(t,"gi");return this._esc(s).replace(e,i=>`<mark class="hl">${i}</mark>`)},_esc(s){return String(s).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;")}}}function Ac(){return{name:"",version:"",busy:!1,toast:{msg:"",cls:""},async doExport(){if(!(!this.name||!this.version)){this.busy=!0,this._toast("","");try{let s=await fetch("/api/templates/export",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({name:this.name,version:this.version})});if(!s.ok){let l=await s.json().catch(()=>({detail:s.statusText}));throw new Error(l.detail||s.statusText)}let t=await s.blob(),i=(s.headers.get("Content-Disposition")||"").match(/filename="([^"]+)"/),n=i?i[1]:"template.tar.gz",r=URL.createObjectURL(t),o=document.createElement("a");o.href=r,o.download=n,o.click(),URL.revokeObjectURL(r),this._toast("\u2705 Exported \u2014 page will refresh to show the new entry.","ok"),setTimeout(()=>location.reload(),1500)}catch(s){this._toast("Export failed: "+s.message,"err")}finally{this.busy=!1}}},_toast(s,t){this.toast={msg:s,cls:t}}}}function Pc(){return{file:null,targetRepo:"",busy:!1,toast:{msg:"",cls:""},conflicts:[],onFileChange(s){this.file=s.target.files[0]||null},async doImport(){if(!(!this.file||!this.targetRepo)){this.busy=!0,this.conflicts=[],this._toast("","");try{let s=new FormData;s.append("file",this.file);let t="/api/templates/import?target_repo="+encodeURIComponent(this.targetRepo),e=await fetch(t,{method:"POST",body:s});if(!e.ok){let n=await e.json().catch(()=>({detail:e.statusText}));throw new Error(n.detail||e.statusText)}let i=await e.json();this.conflicts=(i.conflicts||[]).filter(n=>n.exists),this._toast("\u2705 Imported "+i.extracted.length+" file(s) ("+i.name+" v"+i.version+").","ok")}catch(s){this._toast("Import failed: "+s.message,"err")}finally{this.busy=!1}}},_toast(s,t){this.toast={msg:s,cls:t}}}}function Mc(s){return{slug:s,open:!1,loaded:!1,toggle(){if(this.open=!this.open,this.open&&!this.loaded){this.loaded=!0;let t=document.getElementById("env-detail-"+this.slug);t&&htmx.trigger(t,"detail-load")}},async deleteWorktree(t){if(!confirm(`Remove sandbox "${t}"?
+`);n=a.pop()??"";for(let h of a){if(!h.startsWith("data: "))continue;let c=h.slice(6).trim();if(!c)continue;let f;try{f=JSON.parse(c)}catch{continue}if(f.t==="start")this.filingProgress=`Filing 0 / ${f.total}\u2026`;else if(f.t==="label")this.filingProgress=f.text;else if(f.t==="issue")this.filingProgress=`Filing ${f.index} / ${f.total} \u2014 ${f.phase}`,this.filedIssues.push({number:f.number,url:f.url,title:f.title,phase:f.phase});else if(f.t!=="blocked"){if(f.t==="done"){r=f;break t}else if(f.t==="error")throw new Error(f.detail||"Filing issues failed.")}}}if(!r)throw new Error("Stream ended without a done event.");this.result={initiative:r.initiative,total:r.total,issues:r.issues??[]},this.batchId=r.batch_id??"",this.initiative=r.initiative??this.initiative;try{localStorage.setItem("ac_active_batch",this.batchId),localStorage.setItem("ac_active_initiative",this.initiative)}catch{}this.step="done"}catch(t){this.errorMsg=t.message,this.step="review"}finally{this.submitting=!1}}},reset(){this._closeSse(),this.step="write",this.text="",this.labelPrefix="",this.showOptions=!1,this.errorMsg="",this.streamingText="",this.draft_id="",this.initiative="",this.phaseCount=0,this.issueCount=0,this.yamlValid=!0,this.yamlValidationMsg="",this.filingProgress="",this.filedIssues=[],this.result={},this.batchId="",this.batchIdCopied=!1,this._editor&&this._setEditorValue("")},async copyBatchId(){if(this.batchId)try{await navigator.clipboard.writeText(this.batchId),this.batchIdCopied=!0,setTimeout(()=>{this.batchIdCopied=!1},1500)}catch{}},async reRun(s){try{let t=await fetch(`/api/plan/${encodeURIComponent(s)}/plan-text`);if(!t.ok){let i=await t.json().catch(()=>({}));this.errorMsg=i.detail||`Could not load run (HTTP ${t.status})`;return}let e=await t.json();this.reset(),this.text=e.plan_text??"",await this.$nextTick(),this.$refs.textarea&&this.autoGrow(this.$refs.textarea)}catch(t){this.errorMsg=t.message||"Failed to load previous run."}},_mountEditor(s){let t=this.$refs.yamlEditor;if(!t)return;if(this._editor){this._setEditorValue(s);return}let e=this,i=A.updateListener.of(n=>{n.docChanged&&(clearTimeout(e._validateTimer),e._validateTimer=setTimeout(()=>e._validateYaml(),ig))});this._editor=new A({state:W.create({doc:s,extensions:[vh(),za(),$a(),xr.of([...Jh,...Th]),mc(),vc,A.lineWrapping,i]}),parent:t})},_getEditorValue(){return this._editor?this._editor.state.doc.toString():""},_setEditorValue(s){this._editor&&this._editor.dispatch({changes:{from:0,to:this._editor.state.doc.length,insert:s},selection:{anchor:0},scrollIntoView:!0})},async _validateYaml(){if(this.step!=="review")return;let s=this._getEditorValue();if(!s.trim()){this.yamlValid=!1,this.yamlValidationMsg="\u26A0 YAML is empty.";return}try{let e=await(await fetch("/api/plan/validate",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({yaml_text:s})})).json();e.valid?(this.yamlValid=!0,this.yamlValidationMsg=`\u2713 Valid \u2014 ${e.phase_count} phases, ${e.issue_count} issues`):(this.yamlValid=!1,this.yamlValidationMsg=`\u2717 ${e.detail||"Invalid PlanSpec"}`)}catch{this.yamlValidationMsg=""}}}}function Oc(s){return{activeIssue:null,events:[],thoughts:[],streamOpen:!1,_evtSource:null,chatMessage:"",chatSending:!1,chatError:null,agentStopping:!1,dispatchOpen:!1,dispatchIssue:null,dispatchRole:"python-developer",roleGroups:s,dispatching:!1,dispatchError:null,dispatchSuccess:!1,dispatchResult:null,labelDispatchOpen:!1,labelDispatchLabel:"",scopeMode:"full_initiative",scopePhases:[],selectedPhase:"",scopeIssues:[],selectedIssueNumber:null,labelContextLoading:!1,labelContextLoaded:!1,showAdvanced:!1,advancedRole:"",labelDispatching:!1,labelDispatchError:null,labelDispatchSuccess:!1,labelDispatchResult:null,dispatcherCopied:!1,get launchPreviewText(){let t=this.labelDispatchLabel;if(this.scopeMode==="full_initiative")return`A ${this.advancedRole.trim()||"coordinator"} will survey every open ticket under "${t}" and assemble its own team.`;if(this.scopeMode==="phase")return this.selectedPhase?`A ${this.advancedRole.trim()||"coordinator"} will handle all tickets in phase "${this.selectedPhase}".`:"Choose a phase to see the preview.";if(this.scopeMode==="issue"){if(!this.selectedIssueNumber)return"Choose a ticket to see the preview.";let e=this.scopeIssues.find(n=>n.number===this.selectedIssueNumber),i=e?e.title:`#${this.selectedIssueNumber}`;return`One leaf agent will work on #${this.selectedIssueNumber}: "${i}".`}return""},get launchReady(){return this.scopeMode==="full_initiative"?!0:this.scopeMode==="phase"?!!this.selectedPhase:this.scopeMode==="issue"?!!this.selectedIssueNumber:!1},get repo(){return window._buildRepo??""},onInspect(t){this.activeIssue?.number!==t.number&&(this._closeStream(),this.activeIssue=t,this.events=[],this.thoughts=[],t.run&&this._openStream(t.run.id))},clearInspect(){this._closeStream(),this.activeIssue=null,this.events=[],this.thoughts=[],this.chatMessage="",this.chatError=null},async sendMessage(){let t=this.chatMessage.trim();if(!t||!this.activeIssue?.run)return;let e=this.activeIssue.run.id;this.chatSending=!0,this.chatError=null;try{let i=await fetch(`/api/runs/${encodeURIComponent(e)}/message`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({content:t})});if(i.ok)this.chatMessage="";else{let n=await i.json().catch(()=>({}));this.chatError=n.detail??`Error ${i.status}`}}catch(i){this.chatError=`Network error: ${i.message}`}finally{this.chatSending=!1}},async stopAgent(){if(!this.activeIssue?.run)return;let t=this.activeIssue.run.id;this.agentStopping=!0;try{await fetch(`/api/runs/${encodeURIComponent(t)}/stop`,{method:"POST"}),this._closeStream()}catch{}finally{this.agentStopping=!1}},restartAgent(){this.activeIssue&&this.openDispatch(this.activeIssue)},_openStream(t){this._closeStream();let e=new EventSource(`/ship/runs/${encodeURIComponent(t)}/stream`);this._evtSource=e,this.streamOpen=!0,e.onmessage=i=>{let n;try{n=JSON.parse(i.data)}catch{return}if(n.t!=="ping"){if(n.t==="event")this.events.push({...n,id:Date.now()+Math.random()});else if(n.t==="thought"){let r=this.thoughts[this.thoughts.length-1];r&&r.role===n.role&&this.thoughts.length>0?r.content+=`
+`+n.content:this.thoughts.push(n),this._scrollCot()}}},e.onerror=()=>{this.streamOpen=!1}},_closeStream(){this._evtSource&&(this._evtSource.close(),this._evtSource=null),this.streamOpen=!1},_scrollCot(){this.$nextTick(()=>{let t=this.$refs.cotScroll;t&&(t.scrollTop=t.scrollHeight)})},openDispatch(t){this.dispatchIssue=t,this.dispatchRole="python-developer",this.dispatchError=null,this.dispatchSuccess=!1,this.dispatchResult=null,this.dispatching=!1,this.dispatchOpen=!0},async submitDispatch(){if(this.dispatchIssue){this.dispatching=!0,this.dispatchError=null;try{let t=await fetch("/api/dispatch/issue",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({issue_number:this.dispatchIssue.number,issue_title:this.dispatchIssue.title,role:this.dispatchRole,repo:this.repo})}),e=await t.json();t.ok?(this.dispatchResult=e,this.dispatchSuccess=!0):this.dispatchError=e.detail??`Error ${t.status}`}catch(t){this.dispatchError=`Network error: ${t.message}`}finally{this.dispatching=!1}}},openLabelDispatch(t){this.labelDispatchLabel=t.label??"",this.scopeMode="full_initiative",this.scopePhases=[],this.scopeIssues=[],this.selectedPhase="",this.selectedIssueNumber=null,this.labelContextLoading=!1,this.labelContextLoaded=!1,this.showAdvanced=!1,this.advancedRole="",this.labelDispatchError=null,this.labelDispatchSuccess=!1,this.labelDispatchResult=null,this.labelDispatching=!1,this.dispatcherCopied=!1,this.labelDispatchOpen=!0,this._loadLabelContext()},closeLabelDispatch(){this.labelDispatchOpen=!1},async _loadLabelContext(){if(!(this.labelContextLoaded||this.labelContextLoading)){this.labelContextLoading=!0;try{let t=`/api/dispatch/context?label=${encodeURIComponent(this.labelDispatchLabel)}&repo=${encodeURIComponent(this.repo)}`,e=await fetch(t);if(e.ok){let i=await e.json();this.scopePhases=i.phases??[],this.scopeIssues=i.issues??[],this.labelContextLoaded=!0}}catch{}finally{this.labelContextLoading=!1}}},async copyDispatcherPrompt(){try{let t=await fetch("/api/dispatch/prompt");if(!t.ok)throw new Error(`HTTP ${t.status}`);let e=await t.json();await navigator.clipboard.writeText(e.content),this.dispatcherCopied=!0,setTimeout(()=>{this.dispatcherCopied=!1},3e3)}catch(t){alert(`Could not copy prompt: ${t.message}`)}},async submitLabelDispatch(){if(!this.launchReady)return;this.labelDispatching=!0,this.labelDispatchError=null;let t={label:this.labelDispatchLabel,scope:this.scopeMode,repo:this.repo};this.scopeMode==="phase"&&this.selectedPhase&&(t.scope_label=this.selectedPhase),this.scopeMode==="issue"&&this.selectedIssueNumber&&(t.scope_issue_number=this.selectedIssueNumber),this.advancedRole.trim()&&(t.role=this.advancedRole.trim());try{let e=await fetch("/api/dispatch/label",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify(t)}),i=await e.json();e.ok?(this.labelDispatchResult=i,this.labelDispatchSuccess=!0):this.labelDispatchError=i.detail??`Error ${e.status}`}catch(e){this.labelDispatchError=`Network error: ${e.message}`}finally{this.labelDispatching=!1}},fmtTime(t){if(!t)return"";try{return new Date(t).toLocaleTimeString([],{hour:"2-digit",minute:"2-digit"})}catch{return t}},eventIcon(t){return{step_start:"\u25B6",blocker:"\u{1F6A7}",decision:"\u{1F4A1}",done:"\u2705"}[t]??"\u2022"},eventDetail(t){let e=t.payload??{};switch(t.event_type){case"step_start":return e.step??"";case"blocker":return e.description??"";case"decision":return`${e.decision??""} \u2014 ${e.rationale??""}`;case"done":return e.summary||e.pr_url||"";default:return JSON.stringify(e)}}}}function Cc(s,t,e){return{q:s||"",filterRole:t||"",filterStatus:e||"",sort:"mtime",applyFilters(){let i=this.q.toLowerCase();document.querySelectorAll("[data-transcript-row]").forEach(n=>{let r=(n.dataset.preview||"").toLowerCase(),o=n.dataset.role||"",l=n.dataset.status||"",a=!i||r.includes(i),h=!this.filterRole||o===this.filterRole,c=!this.filterStatus||l===this.filterStatus;n.style.display=a&&h&&c?"":"none"}),this.applySortDom()},applySortDom(){let i=document.getElementById("transcript-tbody");if(!i)return;let n=Array.from(i.querySelectorAll("[data-transcript-row]"));n.sort((r,o)=>{switch(this.sort){case"messages":return parseInt(o.dataset.messages)-parseInt(r.dataset.messages);case"words":return parseInt(o.dataset.words)-parseInt(r.dataset.words);case"subagents":return parseInt(o.dataset.subagents)-parseInt(r.dataset.subagents);default:return parseFloat(o.dataset.mtime)-parseFloat(r.dataset.mtime)}}),n.forEach(r=>i.appendChild(r))}}}function Tc(){return{search:"",highlight(s){if(!this.search)return this._esc(s);let t=this.search.replace(/[.*+?^${}()|[\]\\]/g,"\\$&"),e=new RegExp(t,"gi");return this._esc(s).replace(e,i=>`<mark class="hl">${i}</mark>`)},_esc(s){return String(s).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;")}}}function Ac(){return{name:"",version:"",busy:!1,toast:{msg:"",cls:""},async doExport(){if(!(!this.name||!this.version)){this.busy=!0,this._toast("","");try{let s=await fetch("/api/templates/export",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({name:this.name,version:this.version})});if(!s.ok){let l=await s.json().catch(()=>({detail:s.statusText}));throw new Error(l.detail||s.statusText)}let t=await s.blob(),i=(s.headers.get("Content-Disposition")||"").match(/filename="([^"]+)"/),n=i?i[1]:"template.tar.gz",r=URL.createObjectURL(t),o=document.createElement("a");o.href=r,o.download=n,o.click(),URL.revokeObjectURL(r),this._toast("\u2705 Exported \u2014 page will refresh to show the new entry.","ok"),setTimeout(()=>location.reload(),1500)}catch(s){this._toast("Export failed: "+s.message,"err")}finally{this.busy=!1}}},_toast(s,t){this.toast={msg:s,cls:t}}}}function Pc(){return{file:null,targetRepo:"",busy:!1,toast:{msg:"",cls:""},conflicts:[],onFileChange(s){this.file=s.target.files[0]||null},async doImport(){if(!(!this.file||!this.targetRepo)){this.busy=!0,this.conflicts=[],this._toast("","");try{let s=new FormData;s.append("file",this.file);let t="/api/templates/import?target_repo="+encodeURIComponent(this.targetRepo),e=await fetch(t,{method:"POST",body:s});if(!e.ok){let n=await e.json().catch(()=>({detail:e.statusText}));throw new Error(n.detail||e.statusText)}let i=await e.json();this.conflicts=(i.conflicts||[]).filter(n=>n.exists),this._toast("\u2705 Imported "+i.extracted.length+" file(s) ("+i.name+" v"+i.version+").","ok")}catch(s){this._toast("Import failed: "+s.message,"err")}finally{this.busy=!1}}},_toast(s,t){this.toast={msg:s,cls:t}}}}function Mc(s){return{slug:s,open:!1,loaded:!1,toggle(){if(this.open=!this.open,this.open&&!this.loaded){this.loaded=!0;let t=document.getElementById("env-detail-"+this.slug);t&&htmx.trigger(t,"detail-load")}},async deleteWorktree(t){if(!confirm(`Remove sandbox "${t}"?
 
 This deletes the working directory. The branch is kept.`))return;let e=await fetch(`/api/worktrees/${t}`,{method:"DELETE"});if(e.ok)this.$el.style.opacity="0",this.$el.style.transform="translateX(-6px)",setTimeout(()=>location.reload(),300);else{let i=await e.json().catch(()=>({}));alert(`Failed to remove sandbox: ${i.detail??e.status}`)}}}}function Dc(){return{open:!1,tryOpen:!1,method:"GET",pathTemplate:"/",pathParamNames:[],pathParamValues:{},queryParams:[],queryValues:{},bodyFields:[],bodyJson:"",response:null,sending:!1,init(){this.method=this.$el.dataset.method||"GET",this.pathTemplate=this.$el.dataset.path||"/";try{this.queryParams=JSON.parse(this.$el.dataset.queryParams||"[]")}catch{this.queryParams=[]}for(let t of this.queryParams)this.queryValues[t.name]="";try{this.bodyFields=JSON.parse(this.$el.dataset.bodyFields||"[]")}catch{this.bodyFields=[]}if(this.bodyFields.length){let t={};for(let e of this.bodyFields)e.default!==null&&e.default!==void 0&&e.default!==""?t[e.name]=e.default:e.type==="integer"||e.type==="number"?t[e.name]=0:e.type==="boolean"?t[e.name]=!1:e.type&&e.type.startsWith("array")?t[e.name]=[]:t[e.name]="";this.bodyJson=JSON.stringify(t,null,2)}let s=this.pathTemplate.match(/\{(\w+)\}/g)||[];this.pathParamNames=s.map(t=>t.slice(1,-1));for(let t of this.pathParamNames)this.pathParamValues[t]=""},get resolvedPath(){let s=this.pathTemplate;for(let[e,i]of Object.entries(this.pathParamValues))s=s.replace(`{${e}}`,i?encodeURIComponent(i):`{${e}}`);let t=Object.entries(this.queryValues).filter(([,e])=>e!=="").map(([e,i])=>`${encodeURIComponent(e)}=${encodeURIComponent(i)}`);return t.length&&(s+="?"+t.join("&")),s},get curlCommand(){let s=window.location.origin,t=`curl -X ${this.method} '${s}${this.resolvedPath}'`;if(!["GET","HEAD"].includes(this.method)&&this.bodyJson.trim()){let i=this.bodyJson.replace(/'/g,"'\\''");t+=` \\
   -H 'Content-Type: application/json' \\
