+++ agentception/templates/build.html
@@ -26,7 +26,7 @@
   <nav class="build-initiative-tabs" aria-label="Initiatives">
     {% for ini in initiatives %}
     <a class="build-initiative-tab{% if ini == initiative %} build-initiative-tab--active{% endif %}"
-       href="/build?initiative={{ ini }}">
+       href="/ship/{{ ini | urlencode }}">
       {{ ini }}
     </a>
     {% endfor %}
@@ -39,7 +39,7 @@
     {# ── left: phase board ─────────────────────────────────────────────── #}
     <section class="build-board"
              id="build-board"
-             hx-get="/build/board?initiative={{ initiative }}"
+             hx-get="/ship/{{ initiative | urlencode }}/board"
              hx-trigger="load, every 10s, refreshBoard from:body"
              hx-swap="innerHTML"
     >
