+++ agentception/tests/test_label_context_and_dispatch.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 
-"""Tests for GET /api/build/label-context and the scope-based POST /api/build/dispatch-label.
+"""Tests for GET /api/dispatch/context and the scope-based POST /api/dispatch/label.
 
 Covers:
   - _role_and_node_type_for_scope() derives the correct node_type and default role for each scope.
-  - GET /api/build/label-context returns empty lists gracefully when the DB is empty.
-  - POST /api/build/dispatch-label with scope=full_initiative spawns a coordinator with role cto.
-  - POST /api/build/dispatch-label with scope=phase spawns a coordinator for the sub-label.
-  - POST /api/build/dispatch-label with scope=issue spawns a leaf for the given issue number.
-  - POST /api/build/dispatch-label respects an explicit role override in the request.
+  - GET /api/dispatch/context returns empty lists gracefully when the DB is empty.
+  - POST /api/dispatch/label with scope=full_initiative spawns a coordinator with role cto.
+  - POST /api/dispatch/label with scope=phase spawns a coordinator for the sub-label.
+  - POST /api/dispatch/label with scope=issue spawns a leaf for the given issue number.
+  - POST /api/dispatch/label respects an explicit role override in the request.
   - .agent-task file contains SCOPE_TYPE=issue and SCOPE_VALUE=<number> for issue scope.
   - .agent-task file contains INITIATIVE_LABEL for phase and issue scopes.
 
@@ -25,7 +25,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from agentception.app import app
-from agentception.routes.api.build import _role_and_node_type_for_scope
+from agentception.routes.api.dispatch import _role_and_node_type_for_scope
 
 
 # ---------------------------------------------------------------------------
@@ -75,7 +75,7 @@ def test_scope_role_override_blank_ignored() -> None:
 
 
 # ---------------------------------------------------------------------------
-# GET /api/build/label-context — graceful empty response
+# GET /api/dispatch/context — graceful empty response
 # ---------------------------------------------------------------------------
 
 
@@ -84,12 +84,12 @@ def test_label_context_returns_empty_lists_when_no_db_data(
 ) -> None:
     """label-context must return {phases:[], issues:[]} even when DB is empty."""
     with patch(
-        "agentception.routes.api.build.get_label_context",
+        "agentception.routes.api.dispatch.get_label_context",
         new_callable=AsyncMock,
         return_value={"phases": [], "issues": []},
     ):
         res = client.get(
-            "/api/build/label-context",
+            "/api/dispatch/context",
             params={"label": "ac-workflow", "repo": "cgcardona/agentception"},
         )
     assert res.status_code == 200
@@ -104,12 +104,12 @@ def test_label_context_returns_phases_and_issues(client: TestClient) -> None:
         "issues": [{"number": 42, "title": "Fix the thing"}],
     }
     with patch(
-        "agentception.routes.api.build.get_label_context",
+        "agentception.routes.api.dispatch.get_label_context",
         new_callable=AsyncMock,
         return_value=mock_ctx,
     ):
         res = client.get(
-            "/api/build/label-context",
+            "/api/dispatch/context",
             params={"label": "ac-workflow", "repo": "cgcardona/agentception"},
         )
     assert res.status_code == 200
@@ -159,7 +159,7 @@ def _dispatch_label_body(**overrides: object) -> dict[str, object]:
 
 
 # ---------------------------------------------------------------------------
-# POST /api/build/dispatch-label — scope=full_initiative
+# POST /api/dispatch/label — scope=full_initiative
 # ---------------------------------------------------------------------------
 
 
@@ -168,9 +168,9 @@ def test_dispatch_label_full_initiative_creates_coordinator(
     tmp_path: Path,
 ) -> None:
     with (
-        patch("agentception.routes.api.build.settings") as mock_settings,
-        patch("agentception.routes.api.build.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
-        patch("agentception.routes.api.build.persist_agent_run_dispatch", new_callable=AsyncMock),
+        patch("agentception.routes.api.dispatch.settings") as mock_settings,
+        patch("agentception.routes.api.dispatch.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
+        patch("agentception.routes.api.dispatch.persist_agent_run_dispatch", new_callable=AsyncMock),
     ):
         mock_settings.worktrees_dir = str(tmp_path / "worktrees")
         mock_settings.host_worktrees_dir = "/host/worktrees"
@@ -178,7 +178,7 @@ def test_dispatch_label_full_initiative_creates_coordinator(
         mock_settings.ac_url = "http://localhost:10003"
 
         res = client.post(
-            "/api/build/dispatch-label",
+            "/api/dispatch/label",
             json=_dispatch_label_body(scope="full_initiative"),
         )
 
@@ -194,9 +194,9 @@ def test_dispatch_label_full_initiative_role_override(
     tmp_path: Path,
 ) -> None:
     with (
-        patch("agentception.routes.api.build.settings") as mock_settings,
-        patch("agentception.routes.api.build.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
-        patch("agentception.routes.api.build.persist_agent_run_dispatch", new_callable=AsyncMock),
+        patch("agentception.routes.api.dispatch.settings") as mock_settings,
+        patch("agentception.routes.api.dispatch.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
+        patch("agentception.routes.api.dispatch.persist_agent_run_dispatch", new_callable=AsyncMock),
     ):
         mock_settings.worktrees_dir = str(tmp_path / "worktrees2")
         mock_settings.host_worktrees_dir = "/host/worktrees"
@@ -204,7 +204,7 @@ def test_dispatch_label_full_initiative_role_override(
         mock_settings.ac_url = "http://localhost:10003"
 
         res = client.post(
-            "/api/build/dispatch-label",
+            "/api/dispatch/label",
             json=_dispatch_label_body(scope="full_initiative", role="engineering-coordinator"),
         )
 
@@ -215,7 +215,7 @@ def test_dispatch_label_full_initiative_role_override(
 
 
 # ---------------------------------------------------------------------------
-# POST /api/build/dispatch-label — scope=phase
+# POST /api/dispatch/label — scope=phase
 # ---------------------------------------------------------------------------
 
 
@@ -224,9 +224,9 @@ def test_dispatch_label_phase_scope_is_coordinator(
     tmp_path: Path,
 ) -> None:
     with (
-        patch("agentception.routes.api.build.settings") as mock_settings,
-        patch("agentception.routes.api.build.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
-        patch("agentception.routes.api.build.persist_agent_run_dispatch", new_callable=AsyncMock),
+        patch("agentception.routes.api.dispatch.settings") as mock_settings,
+        patch("agentception.routes.api.dispatch.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
+        patch("agentception.routes.api.dispatch.persist_agent_run_dispatch", new_callable=AsyncMock),
     ):
         mock_settings.worktrees_dir = str(tmp_path / "worktrees3")
         mock_settings.host_worktrees_dir = "/host/worktrees"
@@ -234,7 +234,7 @@ def test_dispatch_label_phase_scope_is_coordinator(
         mock_settings.ac_url = "http://localhost:10003"
 
         res = client.post(
-            "/api/build/dispatch-label",
+            "/api/dispatch/label",
             json=_dispatch_label_body(
                 scope="phase",
                 scope_label="ac-workflow/5-plan-step-v2",
@@ -269,9 +269,9 @@ def test_dispatch_label_phase_scope_agent_task_scope_value(
         return original_write_text(self, data, encoding=encoding, errors=errors, newline=newline)
 
     with (
-        patch("agentception.routes.api.build.settings") as mock_settings,
-        patch("agentception.routes.api.build.asyncio.create_subprocess_exec") as mock_exec,
-        patch("agentception.routes.api.build.persist_agent_run_dispatch", new_callable=AsyncMock),
+        patch("agentception.routes.api.dispatch.settings") as mock_settings,
+        patch("agentception.routes.api.dispatch.asyncio.create_subprocess_exec") as mock_exec,
+        patch("agentception.routes.api.dispatch.persist_agent_run_dispatch", new_callable=AsyncMock),
         patch.object(Path, "write_text", _capture_write),
     ):
         wt_dir = tmp_path / "worktrees4"
@@ -283,7 +283,7 @@ def test_dispatch_label_phase_scope_agent_task_scope_value(
         mock_exec.return_value = _make_fake_proc()
 
         client.post(
-            "/api/build/dispatch-label",
+            "/api/dispatch/label",
             json=_dispatch_label_body(
                 scope="phase",
                 scope_label="ac-workflow/5-plan-step-v2",
@@ -298,7 +298,7 @@ def test_dispatch_label_phase_scope_agent_task_scope_value(
 
 
 # ---------------------------------------------------------------------------
-# POST /api/build/dispatch-label — scope=issue
+# POST /api/dispatch/label — scope=issue
 # ---------------------------------------------------------------------------
 
 
@@ -307,9 +307,9 @@ def test_dispatch_label_issue_scope_is_leaf(
     tmp_path: Path,
 ) -> None:
     with (
-        patch("agentception.routes.api.build.settings") as mock_settings,
-        patch("agentception.routes.api.build.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
-        patch("agentception.routes.api.build.persist_agent_run_dispatch", new_callable=AsyncMock),
+        patch("agentception.routes.api.dispatch.settings") as mock_settings,
+        patch("agentception.routes.api.dispatch.asyncio.create_subprocess_exec", new=_make_worktree_exec()),
+        patch("agentception.routes.api.dispatch.persist_agent_run_dispatch", new_callable=AsyncMock),
     ):
         mock_settings.worktrees_dir = str(tmp_path / "worktrees5")
         mock_settings.host_worktrees_dir = "/host/worktrees"
@@ -317,7 +317,7 @@ def test_dispatch_label_issue_scope_is_leaf(
         mock_settings.ac_url = "http://localhost:10003"
 
         res = client.post(
-            "/api/build/dispatch-label",
+            "/api/dispatch/label",
             json=_dispatch_label_body(
                 scope="issue",
                 scope_issue_number=108,
@@ -350,9 +350,9 @@ def test_dispatch_label_issue_scope_agent_task_fields(
         return original_write_text(self, data, encoding=encoding, errors=errors, newline=newline)
 
     with (
-        patch("agentception.routes.api.build.settings") as mock_settings,
-        patch("agentception.routes.api.build.asyncio.create_subprocess_exec") as mock_exec,
-        patch("agentception.routes.api.build.persist_agent_run_dispatch", new_callable=AsyncMock),
+        patch("agentception.routes.api.dispatch.settings") as mock_settings,
+        patch("agentception.routes.api.dispatch.asyncio.create_subprocess_exec") as mock_exec,
+        patch("agentception.routes.api.dispatch.persist_agent_run_dispatch", new_callable=AsyncMock),
         patch.object(Path, "write_text", _capture_write),
     ):
         wt_dir = tmp_path / "worktrees6"
@@ -364,7 +364,7 @@ def test_dispatch_label_issue_scope_agent_task_fields(
         mock_exec.return_value = _make_fake_proc()
 
         client.post(
-            "/api/build/dispatch-label",
+            "/api/dispatch/label",
             json=_dispatch_label_body(scope="issue", scope_issue_number=42),
         )
